### 2019-02-14 #1250 @whedon
[PRE REVIEW]: NeuroDSP: A package for neural digital signal processing
**Submitting author:** @voytek (<a href="http://orcid.org/0000-0003-1640-2525">Bradley Voytek</a>)
**Repository:** <a href="https://github.com/neurodsp-tools/neurodsp" target ="_blank">https://github.com/neurodsp-tools/neurodsp</a>
**Version:** v1.0.0
