### 2019-02-14 86ac971
Update paper.md
fix affiliation issue
