### 2019-02-14 #1250 @voytek
Fixed affiliation quoting issue
