### 2019-02-14 a8bb6f6
Update paper.md
Fixed affiliation quotes for everyone? Unsure the exact syntax but I think this should take care of it.
