### 2019-02-14 ea870e0
Continue refactor - mostly transition band stuff
