+++ neurodsp/filt.py
@@ -3,7 +3,6 @@
 import warnings
 
 import numpy as np
-import scipy as sp
 from scipy import signal
 
 from neurodsp.plts.filt import plot_frequency_response
@@ -68,8 +67,9 @@ def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
         returned only if 'return_kernel' == True
     """
 
-    # Check inputs
+    # Check inputs &  compute the nyquist frequency
     f_lo, f_hi = check_filter_definitions(pass_type, fc)
+    f_nyq = fs / 2.
 
     # Remove any NaN on the edges of 'sig'
     first_nonan = np.where(~np.isnan(sig))[0][0]
@@ -77,12 +77,10 @@ def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
     sig_old = np.copy(sig)
     sig = sig[first_nonan:last_nonan]
 
-    # Compute nyquist frequency
-    f_nyq = fs / 2.
-
     if iir:
-        sig_filt, b_vals, a_vals = iir_filt(sig, pass_type, f_lo, f_hi, f_nyq, butterworth_order,
-            verbose, n_seconds, remove_edge_artifacts, plot_freq_response)
+        iir_checks(pass_type, n_seconds, butterworth_order, remove_edge_artifacts, verbose)
+        sig_filt, b_vals, a_vals = iir_filt(sig, pass_type, f_lo, f_hi, f_nyq,
+                                            butterworth_order, plot_freq_response)
     else:
         sig_filt, filt_len, kernel = fir_filt(sig, fs, pass_type, f_lo, f_hi, f_nyq, n_cycles,
                                               n_seconds, plot_freq_response)
@@ -90,8 +88,7 @@ def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
     # Compute transition bandwidth
     if compute_transition_band and verbose:
         if not iir:
-            b_vals = kernel
-            a_vals = 1
+            a_vals, b_vals = 1, kernel
         calc_transition_band(b_vals, a_vals, fs, f_lo, f_hi, f_nyq, pass_type)
 
     # Remove edge artifacts
@@ -116,26 +113,24 @@ def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
 
 
 def check_filter_definitions(pass_type, fc):
+    """Words, words, words."""
 
-    # Check that frequency cutoff inputs are appropriate
+    if pass_type not in ['bandpass', 'bandstop', 'lowpass', 'highpass']:
+        raise ValueError('Filter passtype not understood.')
+
+    ## Check that frequency cutoff inputs are appropriate
+    # For band filters, 2 inputs required & second entry must be > first
     if pass_type in ('bandpass', 'bandstop'):
-        # Check, if fc is a tuple and performing bandpass/stop, that
-        #   the second cutoff frequency is greater than the first
-        if isinstance(fc, tuple):
-            if fc[0] >= fc[1]:
-                raise ValueError('Second cutoff frequency must be greater than first.')
-
-        if isinstance(fc, (int, float)):
-            raise ValueError('Two cutoff frequencies required for bandpass and bandstop filters')
-        if len(fc) != 2:
+        if isinstance(fc, tuple) and fc[0] >= fc[1]:
+            raise ValueError('Second cutoff frequency must be greater than first.')
+        elif isinstance(fc, (int, float)) or len(fc) != 2:
             raise ValueError('Two cutoff frequencies required for bandpass and bandstop filters')
 
[…]
