### 2019-02-15 1a1b618
Tutorial updates & fixes for sphinx
