+++ tutorials/plot_1-Filtering.py
@@ -1,60 +1,50 @@
 """
 Filtering
 =========
+
+This tutorial covers using digital filters, and is broken down into the following sections:
+
+1. Bandpass filter: extract a single oscillator from your signal
+2. Highpass, lowpass, and bandstop filters: remove power in unwanted frequency ranges
+3. Time-frequency resolution tradeoff: Change the filter length
+4. Infinite-impulse-response (IIR) filter option.
+5. Beta bandpass filter on a neural signal
 """
 
-###############################################################################
-# This notebook shows how to use the neurodsp module to filter neural signals.
-#
-
-###############################################################################
-#
-# Filtering neural signals using neurodsp.filt
-# --------------------------------------------
-#
-# This tutorial is broken down into the following sections:
-#
-# 1. Bandpass filter: extract a single oscillator from your signal
-# 2. Highpass, lowpass, and bandstop filters: remove power in unwanted frequency ranges
-# 3. Time-frequency resolution tradeoff: Change the filter length, n_cycles and n_taps
-# 4. Infinite-impulse-response (IIR) filter option.
-# 5. Beta bandpass filter on a neural signal
-#
-
-###############################################################################
+###################################################################################################
 
 import numpy as np
 import matplotlib.pyplot as plt
 
 from neurodsp import filt
 
-###############################################################################
+###################################################################################################
+
+# Set the random seed, for consistency simulating data
+np.random.seed(0)
+
+###################################################################################################
 #
 # 1. Bandpass filter
 # ------------------
 #
-# Extract signal within a specific frequency range (e.g. theta, 4-8Hz)
+# Extract signal within a specific frequency range (e.g. theta, 4-8 Hz).
 #
 
-###############################################################################
+###################################################################################################
 
 # Generate an oscillation with noise
-np.random.seed(0)
 t = np.arange(0,4,.001)
 x = np.random.randn(len(t)) + 5*np.sin(t*2*np.pi*6)
 fs = 1000
 
-###############################################################################
+###################################################################################################
 
-# Define the frequency band of interest
-fc = (4,8)
-
-###############################################################################
-
-# Filter the data
+# Filter the data, across a frequency band of interest
+fc = (4, 8)
 x_filt = filt.filter_signal(x, fs, 'bandpass', fc)
 
-###############################################################################
+###################################################################################################
 
 # Plot filtered signal
 plt.figure(figsize=(15,3))
@@ -62,18 +52,17 @@ plt.plot(t, x, 'k', label='raw')
 plt.plot(t, x_filt, 'r', label='filtered')
 plt.legend(loc='best')
 
-###############################################################################
+###################################################################################################
 #
 # Notice that the edges of the filtered signal are clipped (no red).
 #
-# Edge artifact removal is done by default in `filter_signal()`, because
+# Edge artifact removal is done by default in :func:`filter_signal`, because
 # the signal samples at the edges only experienced part of the filter.
-# As the saying goes, 99% of a filter is not a filter.
 #
 # To bypass this feature, set `remove_edge_artifacts=False`, but at your own risk!
 #
 
-###############################################################################
+###################################################################################################
 #
 # 2. Highpass, lowpass, and bandstop filters
 # ------------------------------------------
@@ -84,49 +73,47 @@ plt.legend(loc='best')
 # Remove low frequency drift from the data
 #
 
-###############################################################################
+###################################################################################################
 
[…]

[…]
