### 2019-02-15 cb3f24b
Various doc updates, related to sphinx site & some orgs
