+++ neurodsp/spectral.py
@@ -50,7 +50,7 @@ def compute_spectrum(sig, fs, method='mean', window='hann', nperseg=None,
     freqs : ndarray
         Array of sample frequencies.
     spectrum : ndarray
-        Power spectral density of x.
+        Power spectral density.
 
     References
     ----------
@@ -74,24 +74,28 @@ def compute_spectrum(sig, fs, method='mean', window='hann', nperseg=None,
         if noverlap is not None:
             noverlap = int(noverlap)
 
-        # call signal.spectrogram function in scipy to compute STFT
+        # Calculate the short time fourier transform with signal.spectrogram
         freqs, t_axis, spg = signal.spectrogram(sig, fs, window, nperseg, noverlap)
 
-        # pad data to 2D
+        # Pad data to 2D
         if len(sig.shape) == 1:
             sig = sig[None, :]
 
         numchan = sig.shape[0]
-        # throw out outliers if indicated
+
+        # Throw out outliers if indicated
         if spg_outlier_pct > 0.:
+
             n_discard = int(np.ceil(len(t_axis) / 100. * spg_outlier_pct))
             n_keep = int(len(t_axis)-n_discard)
             spg_temp = np.zeros((numchan, len(freqs), n_keep))
             outlier_inds = np.zeros((numchan, n_discard))
+
+            # Discard time windows with high total log powers, round up so it doesn't get a zero
             for chan in range(numchan):
-                # discard time windows with high total log powers, round up so it doesn't get a zero
                 outlier_inds[chan, :] = np.argsort(np.mean(np.log10(spg[chan, :, :]), axis=0))[-n_discard:]
                 spg_[chan, :, :] = np.delete(spg[chan], outlier_inds[chan, :], axis=-1)
+
             spg = spg_temp
 
         if method == 'mean':
@@ -101,12 +105,11 @@ def compute_spectrum(sig, fs, method='mean', window='hann', nperseg=None,
 
     elif method == 'medfilt':
 
-        # median filtered FFT spectrum
-        # take the positive half of the spectrum since it's symmetrical
+        # Take the positive half of the spectrum since it's symmetrical
         FT = np.fft.fft(sig)[:int(np.ceil(len(sig) / 2.))]
         freqs = np.fft.fftfreq(len(sig), 1. / fs)[:int(np.ceil(len(sig) / 2.))]  # get freq axis
 
-        # convert median filter length from Hz to samples
+        # Convert median filter length from Hz to samples
         filt_len_samp = int(int(filt_len / (freqs[1] - freqs[0])) / 2 * 2 + 1)
         spectrum = signal.medfilt(np.abs(FT)**2. / (fs * len(sig)), filt_len_samp)
 
@@ -145,7 +148,7 @@ def compute_scv(sig, fs, window='hann', nperseg=None, noverlap=0, outlier_pct=No
 
     Returns
     -------
-    freq : ndarray
+    freqs : ndarray
         Array of sample frequencies.
     spect_cv : ndarray
         Spectral coefficient of variation.
@@ -168,17 +171,18 @@ def compute_scv(sig, fs, window='hann', nperseg=None, noverlap=0, outlier_pct=No
     if noverlap is not None:
         noverlap = int(noverlap)
 
-    freq, _, spg = signal.spectrogram(sig, fs, window, nperseg, noverlap)
+    freqs, _, spg = signal.spectrogram(sig, fs, window, nperseg, noverlap)
+
     if outlier_pct is not None:
-        # discard time windows with high powers
-        # round up so it doesn't get a zero
+
+        # Discard time windows with high powers. Round up so it doesn't get a zero.
         discard = int(np.ceil(spg.shape[1] / 100. * outlier_pct))
[…]
