+++ neurodsp/plts/spectral.py
@@ -6,7 +6,6 @@ import matplotlib.pyplot as plt
 ###################################################################################################
 ###################################################################################################
 
-
 def plot_spectral_hist(freqs, power_bins, spect_hist, spectrum_freqs=None, spectrum=None):
     """Plot the spectral histogram.
 
