### 2019-02-15 9c733bb
lint & style on tutorials
