+++ tutorials/plot_1-Filtering.py
@@ -34,22 +34,22 @@ np.random.seed(0)
 ###################################################################################################
 
 # Generate an oscillation with noise
-t = np.arange(0,4,.001)
-x = np.random.randn(len(t)) + 5*np.sin(t*2*np.pi*6)
+times = np.arange(0, 4, .001)
+sig = np.random.randn(len(times)) + 5*np.sin(times*2*np.pi*6)
 fs = 1000
 
 ###################################################################################################
 
 # Filter the data, across a frequency band of interest
 fc = (4, 8)
-x_filt = filt.filter_signal(x, fs, 'bandpass', fc)
+sig_filt = filt.filter_signal(sig, fs, 'bandpass', fc)
 
 ###################################################################################################
 
 # Plot filtered signal
-plt.figure(figsize=(15,3))
-plt.plot(t, x, 'k', label='raw')
-plt.plot(t, x_filt, 'r', label='filtered')
+plt.figure(figsize=(15, 3))
+plt.plot(times, sig, 'k', label='raw')
+plt.plot(times, sig_filt, 'r', label='filtered')
 plt.legend(loc='best')
 
 ###################################################################################################
@@ -76,21 +76,21 @@ plt.legend(loc='best')
 ###################################################################################################
 
 # Generate a signal with a low-frequency drift
-t = np.arange(0, 6, .001)
-x = np.random.randn(len(t)) + 5 * np.sin(t*2*np.pi*3) + 4 * np.sin(t*2*np.pi*.5)
+times = np.arange(0, 6, .001)
+sig = np.random.randn(len(times)) + 5 * np.sin(times*2*np.pi*3) + 4 * np.sin(times*2*np.pi*.5)
 
 ###################################################################################################
 
 # Filter the data
 fc = 2
-x_filt = filt.filter_signal(x, fs, 'highpass', (fc, None))
+sig_filt = filt.filter_signal(sig, fs, 'highpass', (fc, None))
 
 ###################################################################################################
 
 # Plot filtered signal
 plt.figure(figsize=(15, 3))
-plt.plot(t, x, 'k', label='raw')
-plt.plot(t, x_filt, 'r', label='filtered')
+plt.plot(times, sig, 'k', label='raw')
+plt.plot(times, sig_filt, 'r', label='filtered')
 plt.legend(loc='best')
 
 ###################################################################################################
@@ -104,21 +104,21 @@ plt.legend(loc='best')
 ###################################################################################################
 
 # Generate a signal with a low-frequency drift
-t = np.arange(0, 6, .001)
-x = np.random.randn(len(t)) + 5 * np.sin(t*2*np.pi*3) + 4 * np.sin(t*2*np.pi*.5)
+times = np.arange(0, 6, .001)
+sig = np.random.randn(len(times)) + 5 * np.sin(times*2*np.pi*3) + 4 * np.sin(times*2*np.pi*.5)
 
 ###################################################################################################
 
 # Filter the data
 fc = 20
-x_filt = filt.filter_signal(x, fs, 'lowpass', (None, fc))
+sig_filt = filt.filter_signal(sig, fs, 'lowpass', (None, fc))
 
 ###################################################################################################
 
 # Plot filtered signal
-plt.figure(figsize=(15,3))
-plt.plot(t, x, 'k', label='raw')
-plt.plot(t, x_filt, 'r', label='filtered')
+plt.figure(figsize=(15, 3))
+plt.plot(times, sig, 'k', label='raw')
+plt.plot(times, sig_filt, 'r', label='filtered')
 plt.legend(loc='best')
 
 ###################################################################################################
@@ -136,21 +136,21 @@ plt.legend(loc='best')
[…]
