### 2019-02-15 27e8080
Last run through of in-code doc udpates, for consistency & related to sphinx
