+++ _downloads/3c869693f9b9df313d90bae39393b1e4/plot_2-InstantaneousMeasures.py
@@ -1,105 +1,125 @@
 """
 Time-frequency analysis
 =======================
-In this tutorial, we will show how to estimate instantaneous measures of phase,
-amplitude, and frequency.
+
+In this tutorial, we will show how to estimate instantaneous measures of
+phase, amplitude, and frequency.
 """
 
-###############################################################################
+###################################################################################################
 
 import numpy as np
-from neurodsp.timefrequency import amp_by_time, freq_by_time, phase_by_time
 import matplotlib.pyplot as plt
 
-###############################################################################
+from neurodsp.timefrequency import amp_by_time, freq_by_time, phase_by_time
+
+###################################################################################################
 #
 # Load example neural signal
 # --------------------------
 
 sig = np.load('./data/sample_data_1.npy')
-Fs = 1000
-t = np.arange(0, len(sig)/Fs, 1/Fs)
-f_range = (13,30)
+fs = 1000
+times = np.arange(0, len(sig)/fs, 1/fs)
+f_range = (13, 30)
 
 # Load filtered version of signal
 sig_filt_true = np.load('./data/sample_data_1_filt.npy')
 
+###################################################################################################
+
 # Plot signal
-plt.figure(figsize=(12,3))
-plt.plot(t, sig, 'k')
-plt.xlim((4,5))
+plt.figure(figsize=(12, 3))
+plt.plot(times, sig, 'k')
+plt.xlim((4, 5))
 plt.xlabel('Time (s)')
 plt.ylabel('Voltage (uV)')
 
-###############################################################################
+###################################################################################################
 #
 # Compute and plot instantaneous phase
 # ------------------------------------
 
-pha = phase_by_time(sig, Fs, f_range)
+###################################################################################################
+
+# Compute instaneous phase from a signal
+pha = phase_by_time(sig, fs, f_range)
+
+###################################################################################################
 
 # Plot example signal
-plt.figure(figsize=(12,4))
-plt.subplot(2,1,1)
-plt.plot(t, sig, 'k')
-plt.xlim((4,5))
+plt.figure(figsize=(12, 4))
+plt.subplot(2, 1, 1)
+plt.plot(times, sig, 'k')
+plt.xlim((4, 5))
 plt.ylabel('Voltage (uV)')
-plt.subplot(2,1,2)
-plt.plot(t, pha, 'k')
-plt.xlim((4,5))
-plt.yticks([-np.pi, 0, np.pi], ['$\pi$',0,'$\pi$'])
+plt.subplot(2, 1, 2)
+plt.plot(times, pha, 'k')
+plt.xlim((4, 5))
+plt.yticks([-np.pi, 0, np.pi], ['$\pi$', 0, '$\pi$'])
 plt.xlabel('Time (s)')
 plt.ylabel('Phase (rad)')
 
-###############################################################################
+###################################################################################################
 #
 # Compute and plot instantaneous amplitude
 # ----------------------------------------
 
-amp = amp_by_time(sig, Fs, f_range)
+###################################################################################################
+
+# Compute instaneous amplitude from a signal
+amp = amp_by_time(sig, fs, f_range)
+
+###################################################################################################
 
 # Plot example signal
-plt.figure(figsize=(12,4))
-plt.subplot(2,1,1)
-plt.plot(t, sig, 'k', label='raw voltage')
[…]
