+++ _downloads/9f6f6a6d3328dd0cd4373de0258fe99f/plot_5-BurstDetection.py
@@ -1,29 +1,33 @@
 """
 Burst Detection
 ===============
+
 This notebook shows how to use `neurodsp.burst` to determine the portions
 of a signal that contain a burst of an oscillation of interest.
 """
 
-###############################################################################
+###################################################################################################
 
 import numpy as np
-from neurodsp.sim import sim_noisy_bursty_oscillator
-from neurodsp.burst import detect_bursts_dual_threshold, compute_burst_stats
 import matplotlib.pyplot as plt
 
-###############################################################################
+from neurodsp.sim import sim_noisy_bursty_oscillator
+from neurodsp.burst import detect_bursts_dual_threshold, compute_burst_stats
+
+###################################################################################################
+
+# Set the random seed, for consistency simulating data
+np.random.seed(0)
+
+###################################################################################################
 #
 # Simulate a noisy and bursty oscillator
 # --------------------------------------
 #
-# First, we'll simulate bursting activity in the alpha range on
-# top of some background noise.
+# First, we'll simulate bursting activity in the alpha range on top of an aperiodic component.
 #
 
-###############################################################################
-
-np.random.seed(1)
+###################################################################################################
 
 n_seconds = 5
 fs = 1000
@@ -33,23 +37,24 @@ f_range = (8, 12)
 noise_model = 'synaptic'
 noise_args = {'n_neurons':1000, 'firing_rate':2, 't_ker':1.0, 'tau_r':0.002, 'tau_d':0.02}
 
-###############################################################################
+###################################################################################################
 
 sig = sim_noisy_bursty_oscillator(n_seconds, fs, freq, noise_model, noise_args,
                                   ratio_osc_var=2., prob_enter_burst=.2, prob_leave_burst=.2,)
-t = np.arange(0, n_seconds, 1/fs)
+times = np.arange(0, n_seconds, 1/fs)
 
-###############################################################################
+###################################################################################################
 
-plt.figure(figsize=(16,3))
-plt.plot(t, sig, 'k', label='simulated EEG')
+plt.figure(figsize=(16, 3))
+plt.plot(times, sig, 'k', label='Simulated EEG')
 plt.xlim((0, n_seconds))
 
-###############################################################################
+###################################################################################################
 #
 # We can see that 10Hz oscillations present 1.5-2.1 seconds and after 4.2 seconds.
+#
 
-###############################################################################
+###################################################################################################
 #
 # Dual amplitude threshold algorithm
 # -----------------------------------
@@ -73,18 +78,18 @@ amp_dual_thresh = (1, 2)
 bursting = detect_bursts_dual_threshold(sig, fs, f_range, amp_dual_thresh)
 
 # Plot original signal and burst activity
-plt.figure(figsize=(12,3));
-plt.plot(t, sig, 'k', label='simulated EEG');
-plt.plot(t[bursting], sig[bursting], 'r', label='burst detected')
-plt.legend(loc='best');
+plt.figure(figsize=(12, 3))
+plt.plot(times, sig, 'k', label='simulated EEG')
+plt.plot(times[bursting], sig[bursting], 'r', label='burst detected')
+plt.legend(loc='best')
 
-###############################################################################
+###################################################################################################
 #
 # The graph above shows the bursting activity in red. The algorithm was
 # used with thresh=(1, 2), so any time point with more than 3 times the
 # median magnitude in the alpha range (8-12 Hz) was marked as bursting activity.
 
-###############################################################################
+###################################################################################################
 #
 # Burst detection applied to real recordings
[…]
