+++ _downloads/b52349e2ed4ed17a9944d2416e40393d/plot_7-SimulatingSignals.py
@@ -1,61 +1,59 @@
 """
 Simulating Signals
 ==================
+
+Simulate periodic (oscillatory) and aperiodic (1/f-like) signals with neurodsp.sim.
 """
 
-###############################################################################
-#
-# Simulating 1/f-like noise and oscillatory signals with neurodsp.sim
-#
-# In this notebook, we simulate 'brown'-like noise from several models,
-# stationary and non-stationary oscillators, and the two together using
-# the simulation library, `sim`, in `neurodsp`.
-
-###############################################################################
+###################################################################################################
 
 import numpy as np
-np.random.seed(0)
+import matplotlib.pyplot as plt
 
 from neurodsp import spectral, sim
 
-import matplotlib.pyplot as plt
+###################################################################################################
 
+# Set the random seed, for consistency simulating data
+np.random.seed(0)
 
-###############################################################################
+###################################################################################################
 # Simulate brown noise
 # --------------------
 #
-# Often, we want to simulate noise that is comparable to what we see in
-# neural recordings. Brown noise is a good proxy for that because,
-# like neural signals, power decreases roughly linearly with frequency.
+# Often, we want to simulate noise that is comparable to what we see in neural recordings.
+#
+# Brown noise is a good proxy for that because, like neural signals,
+# power decreases roughly linearly with frequency.
+#
 
-###############################################################################
+###################################################################################################
 
 # Simulate brown noise
 n_seconds = 10
 fs = 1000
 exponent = -2
-t = np.arange(0, n_seconds, 1/fs)
+times = np.arange(0, n_seconds, 1/fs)
 br_noise = sim.sim_variable_powerlaw(n_seconds, fs, exponent)
 
-###############################################################################
+###################################################################################################
 
-# Plot time series
-plt.figure(figsize=(12,3))
-plt.plot(t, br_noise, 'k')
+# Plot the simulated data, in the time domain
+plt.figure(figsize=(12, 3))
+plt.plot(times, br_noise, 'k')
 plt.xlabel('Time (s)')
 plt.ylabel('Voltage')
 
-###############################################################################
+###################################################################################################
 
-# Plot power spectrum
+# Plot the simulated data, in the frequency domain
 freqs, psd = spectral.compute_spectrum(br_noise, fs)
-plt.figure(figsize=(5,5))
+plt.figure(figsize=(5, 5))
 plt.loglog(freqs, psd, 'k')
 plt.xlabel('Frequency (Hz)')
 plt.ylabel('Power')
 
-###############################################################################
+###################################################################################################
 #
 # Simulate filtered brown noise
 # -----------------------------
@@ -68,36 +66,39 @@ plt.ylabel('Power')
 #
 # Note this might not be ideal because it creates an "oscillation" at
 # the cutoff frequency.
+#
 
-###############################################################################
+###################################################################################################
 
 # Simulate highpass-filtered brown noise with a 1Hz cutoff frequency
 f_hipass_brown = 1
 brown_filt = sim.sim_filtered_noise(n_seconds, fs, exponent, (f_hipass_brown, None))
 
-###############################################################################
+###################################################################################################
 
-# Plot time series
-plt.figure(figsize=(12,3))
-plt.plot(t, brown_filt, 'k')
+# Plot the simulated data, in the time domain
[…]

[…]
