### 2019-02-16 #1250 @voytek
> @voytek : do you have any ideas for potential reviewers?

The most obvious would be Alexandre Gramfort (alexandre.gramfort@inria.fr), who is the first author on MNE, and Denis Engemann (denis.engemann@googlemail.com) who is also an MNE dev. Both have significant electrophysiology and Python experience.
