### 2019-02-18 #1250 @arokem
Hey @agramfort: would you be willing to review this paper for [JOSS](http://joss.theoj.org)? Thanks!
