### 2019-02-18 #1250 @agramfort
I could but I would suggest to ask @TomDLT who certainly has more time to take a close look. Otherwise when would you need a review? For when? Should the review be posted here? What are the guidelines for reviewers?
