### 2019-02-18 #1250 @arokem
Thanks for the quick response. The guidelines for reviewers are [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Generally: public/open review on GitHub, focused on a checklist of items that we would like the software to comply with, before publishing the paper. We'd like to have a review in two weeks, if possible. Can you take this on?
