### 2019-02-18 #1250 @arokem
@TomDLT: @agramfort suggested you would be a good reviewer for this article in [JOSS](http://joss.theoj.org). Would you be available to review this in addition or instead of him? Thanks!
