### 2019-02-20 #1250 @agramfort
thanks @TomDLT

let me know if you need also feedback from me @arokem
