### 2019-02-22 0b14491
Update which images are used as thumbnails
