+++ README.md
@@ -10,6 +10,10 @@
 
 A package of tools to process, analyze, and simulate neural recordings as individual voltage time series, with specific focus on time and frequency domain analyses.
 
+## Documentation
+
+Documentation for the NeuroDSP module is available [here](https://neurodsp-tools.github.io/neurodsp/).
+
 ## Dependencies
 
 NeuroDSP is written in Python, and requires Python >= 3.5 to run.
@@ -40,10 +44,13 @@ To install this cloned copy of neurodsp, move into the directory you just cloned
 
 ## Modules
 
-- ```burst``` : Detect bursting oscillators in neural signals ([Tutorial](https://github.com/neurodsp-tools/neurodsp/blob/master/tutorials/5-BurstDetection.ipynb))
-- ```filt``` : Filter data with bandpass, highpass, lowpass, or notch filters ([Tutorial](https://github.com/neurodsp-tools/neurodsp/blob/master/tutorials/1-Filtering.ipynb))
-- ```laggedcoherence``` : Estimate rhythmicity using the lagged coherence measure ([Tutorial](https://github.com/neurodsp-tools/neurodsp/blob/master/tutorials/3-LaggedCoherence.ipynb))
-- ```sim``` : Simulate bursting or stationary oscillators with brown noise ([Tutorial](https://github.com/neurodsp-tools/neurodsp/blob/master/tutorials/7-SimulatingSignals.ipynb))
-- ```spectral``` : Compute spectral domain features (PSD and 1/f slope, etc) ([Tutorial](https://github.com/neurodsp-tools/neurodsp/blob/master/tutorials/4-SpectralAnalysis.ipynb))
-- ```swm``` : Identify recurrent patterns in a signal using sliding window matching ([Tutorial](https://github.com/neurodsp-tools/neurodsp/blob/master/tutorials/6-SlidingWindowMatching.ipynb))
-- ```timefrequency``` : Estimate instantaneous measures of oscillatory activity ([Tutorial](https://github.com/neurodsp-tools/neurodsp/blob/master/tutorials/2-InstantaneousMeasures.ipynb))
+NeuroDSP includes the following modules, each of which have dedicated [tutorials](https://neurodsp-tools.github.io/neurodsp/auto_tutorials/index.html).
+
+- ```filt``` : Filter data with bandpass, highpass, lowpass, or notch filters
+- ```burst``` : Detect bursting oscillations in neural signals
+- ```laggedcoherence``` : Estimate rhythmicity using the lagged coherence measure
+- ```spectral``` : Compute spectral domain features such as power spectra
+- ```swm``` : Identify recurrent patterns in signals using sliding window matching
+- ```timefrequency``` : Estimate instantaneous measures of oscillatory activity
+- ```sim``` : Simulate periodic and aperiodic signal components
+- ```plts``` : Plotting functions
