+++ _downloads/2d3a2886b69a3fb1f0487c292e5be570/plot_1-Filtering.py
@@ -309,3 +309,9 @@ plt.legend(loc='best')
 # `bycycle <https://bycycle-tools.github.io/bycycle/>`_,
 # which is a tool for time-domain analyses of waveform shape.
 #
+
+###################################################################################################
+#
+# Sphinx settings:
+# sphinx_gallery_thumbnail_number = 12
+#
