+++ _sources/auto_tutorials/plot_5-BurstDetection.rst.txt
@@ -272,7 +272,12 @@ requires some knowledge of the data you're working with.
 
 
 
-**Total running time of the script:** ( 0 minutes  0.124 seconds)
+Sphinx settings:
+sphinx_gallery_thumbnail_number = 2
+
+
+
+**Total running time of the script:** ( 0 minutes  0.131 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_5-BurstDetection.py:
