+++ _sources/auto_tutorials/plot_6-SlidingWindowMatching.rst.txt
@@ -125,13 +125,19 @@ would help the robustness of the algorithm.
 
 
 
+
 .. image:: /auto_tutorials/images/sphx_glr_plot_6-SlidingWindowMatching_002.png
     :class: sphx-glr-single-img
 
 
 
 
-**Total running time of the script:** ( 0 minutes  1.306 seconds)
+Sphinx settings:
+sphinx_gallery_thumbnail_number = 2
+
+
+
+**Total running time of the script:** ( 0 minutes  1.346 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_6-SlidingWindowMatching.py:
