+++ _sources/auto_tutorials/plot_4-SpectralAnalysis.rst.txt
@@ -315,13 +315,19 @@ For complete details, see the **Simulation** tutorial.
 
 
 
+
 .. image:: /auto_tutorials/images/sphx_glr_plot_4-SpectralAnalysis_007.png
     :class: sphx-glr-single-img
 
 
 
 
-**Total running time of the script:** ( 0 minutes  0.989 seconds)
+Sphinx settings:
+sphinx_gallery_thumbnail_number = 2
+
+
+
+**Total running time of the script:** ( 0 minutes  0.955 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_4-SpectralAnalysis.py:
