+++ _sources/auto_tutorials/plot_2-InstantaneousMeasures.rst.txt
@@ -228,13 +228,19 @@ estimate in order to make it smoother (see e.g. Samaha & Postle, 2015)
 
 
 
+
 .. image:: /auto_tutorials/images/sphx_glr_plot_2-InstantaneousMeasures_004.png
     :class: sphx-glr-single-img
 
 
 
 
-**Total running time of the script:** ( 0 minutes  0.081 seconds)
+Sphinx settings:
+sphinx_gallery_thumbnail_number = 3
+
+
+
+**Total running time of the script:** ( 0 minutes  0.138 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_2-InstantaneousMeasures.py:
