+++ _sources/auto_tutorials/plot_3-LaggedCoherence.rst.txt
@@ -247,6 +247,7 @@ Compute lagged coherence of an example neural signal
 
 
 
+
 .. rst-class:: sphx-glr-script-out
 
  Out:
@@ -256,7 +257,12 @@ Compute lagged coherence of an example neural signal
     Lagged coherence =  0.24582055225422916
 
 
-**Total running time of the script:** ( 0 minutes  0.117 seconds)
+Sphinx settings:
+sphinx_gallery_thumbnail_number = 2
+
+
+
+**Total running time of the script:** ( 0 minutes  0.104 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_3-LaggedCoherence.py:
