+++ _sources/auto_tutorials/plot_4-SpectralAnalysis.rst.txt
@@ -327,7 +327,7 @@ sphinx_gallery_thumbnail_number = 2
 
 
 
-**Total running time of the script:** ( 0 minutes  0.955 seconds)
+**Total running time of the script:** ( 0 minutes  0.978 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_4-SpectralAnalysis.py:
