+++ _sources/auto_tutorials/plot_2-InstantaneousMeasures.rst.txt
@@ -240,7 +240,7 @@ sphinx_gallery_thumbnail_number = 3
 
 
 
-**Total running time of the script:** ( 0 minutes  0.138 seconds)
+**Total running time of the script:** ( 0 minutes  0.134 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_2-InstantaneousMeasures.py:
