+++ _sources/auto_tutorials/plot_6-SlidingWindowMatching.rst.txt
@@ -137,7 +137,7 @@ sphinx_gallery_thumbnail_number = 2
 
 
 
-**Total running time of the script:** ( 0 minutes  1.346 seconds)
+**Total running time of the script:** ( 0 minutes  1.350 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_6-SlidingWindowMatching.py:
