+++ _sources/auto_tutorials/plot_5-BurstDetection.rst.txt
@@ -277,7 +277,7 @@ sphinx_gallery_thumbnail_number = 2
 
 
 
-**Total running time of the script:** ( 0 minutes  0.131 seconds)
+**Total running time of the script:** ( 0 minutes  0.121 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_5-BurstDetection.py:
