+++ _sources/auto_tutorials/plot_7-SimulatingSignals.rst.txt
@@ -695,7 +695,7 @@ sphinx_gallery_thumbnail_number = 15
 
 
 
-**Total running time of the script:** ( 0 minutes  0.901 seconds)
+**Total running time of the script:** ( 0 minutes  0.880 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_7-SimulatingSignals.py:
