+++ _sources/auto_tutorials/plot_3-LaggedCoherence.rst.txt
@@ -262,7 +262,7 @@ sphinx_gallery_thumbnail_number = 2
 
 
 
-**Total running time of the script:** ( 0 minutes  0.104 seconds)
+**Total running time of the script:** ( 0 minutes  0.107 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_3-LaggedCoherence.py:
