+++ _sources/auto_tutorials/plot_1-Filtering.rst.txt
@@ -643,7 +643,7 @@ sphinx_gallery_thumbnail_number = 12
 
 
 
-**Total running time of the script:** ( 0 minutes  0.233 seconds)
+**Total running time of the script:** ( 0 minutes  0.241 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_1-Filtering.py:
