+++ _sources/auto_tutorials/plot_5-BurstDetection.rst.txt
@@ -10,8 +10,9 @@
 Burst Detection
 ===============
 
-This notebook shows how to use `neurodsp.burst` to determine the portions
-of a signal that contain a burst of an oscillation of interest.
+Analyze neural signals for bursts of oscillations of interest.
+
+This tutorial primarily covers :mod:`neurodsp.burst`.
 
 
 
@@ -277,7 +278,7 @@ sphinx_gallery_thumbnail_number = 2
 
 
 
-**Total running time of the script:** ( 0 minutes  0.121 seconds)
+**Total running time of the script:** ( 0 minutes  0.123 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_5-BurstDetection.py:
