+++ _sources/auto_tutorials/plot_1-Filtering.rst.txt
@@ -10,13 +10,22 @@
 Filtering
 =========
 
-This tutorial covers using digital filters, and is broken down into the following sections:
+Using digital filters on neural signals, including highpass, lowpass, bandpass & bandstop.
+
+This tutorial primarily covers :mod:`neurodsp.filt`.
+
+
+Sections
+--------
+
+This tutorial is broken down into the following sections:
 
 1. Bandpass filter: extract a single oscillator from your signal
 2. Highpass, lowpass, and bandstop filters: remove power in unwanted frequency ranges
 3. Time-frequency resolution tradeoff: Change the filter length
 4. Infinite-impulse-response (IIR) filter option.
 5. Beta bandpass filter on a neural signal
+
 
 
 
@@ -643,7 +652,7 @@ sphinx_gallery_thumbnail_number = 12
 
 
 
-**Total running time of the script:** ( 0 minutes  0.241 seconds)
+**Total running time of the script:** ( 0 minutes  0.301 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_1-Filtering.py:
