+++ _sources/auto_tutorials/plot_2-InstantaneousMeasures.rst.txt
@@ -10,8 +10,9 @@
 Time-frequency analysis
 =======================
 
-In this tutorial, we will show how to estimate instantaneous measures of
-phase, amplitude, and frequency.
+Estimate instantaneous measures of phase, amplitude, and frequency.
+
+This tutorial primarily covers :mod:`neurodsp.timefrequency`.
 
 
 
@@ -240,7 +241,7 @@ sphinx_gallery_thumbnail_number = 3
 
 
 
-**Total running time of the script:** ( 0 minutes  0.134 seconds)
+**Total running time of the script:** ( 0 minutes  0.151 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_2-InstantaneousMeasures.py:
