+++ _sources/auto_tutorials/plot_7-SimulatingSignals.rst.txt
@@ -10,7 +10,9 @@
 Simulating Signals
 ==================
 
-Simulate periodic (oscillatory) and aperiodic (1/f-like) signals with neurodsp.sim.
+Simulate periodic (oscillatory) and aperiodic (1/f-like) signals.
+
+This tutorial primarily covers :mod:`neurodsp.sim`.
 
 
 
@@ -695,7 +697,7 @@ sphinx_gallery_thumbnail_number = 15
 
 
 
-**Total running time of the script:** ( 0 minutes  0.880 seconds)
+**Total running time of the script:** ( 0 minutes  0.936 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_7-SimulatingSignals.py:

[…]
