### 2019-02-22 #1272 @whedon
[REVIEW]: NeuroDSP: A package for neural digital signal processing
**Submitting author:** @voytek (<a href="http://orcid.org/0000-0003-1640-2525">Bradley Voytek</a>)
**Repository:** <a href="https://github.com/neurodsp-tools/neurodsp" target ="_blank">https://github.com/neurodsp-tools/neurodsp</a>
**Version:** 2.0.0
