### 2019-02-22 #1272 @TomDLT
## General review

This Python package implements a collection of more-or-less standard signal processing functions used to analyze univariate neurophysiological time-series.

This package is well packaged, documented, licensed, and tested.
It comes with installation instructions, contributing guidelines, and detailed tutorials.
The API is clear, with consistent parameter names and orderings, and well documented.

I successfully installed the package and ran the tutorials with no effort. After a shallow code review and some personal testing, the implemented functions seems correct. The code is clear and complies with standard Python conventions.

## Package content

The package first provides standard signal processing functions, such as spectral analysis, filtering, and Morlet decomposition. For these functions, it relies heavily on scipy.signal, and extends it nicely in different ways, such as simpler interfaces, user-friendly parameters checks, edge artifacts considerations, or bootstrapping options.

It also implements less standard signal processing techniques, such as lagged-coherence (Fransen et al, 2015), sliding window matching (Gips et al, 2017), or burst detection (reference needed).

Finally, it proposes a number of functions to simulate neurophysiological signals, with different signal and noise models. Some comments might be added to emphasize that these are only models and could lead to oversimplistic signals, compared to empirical neurophysiological signals.

## Minor remarks

1. The authors state in the paper that this package "can easily be integrated with frameworks such as MNE". As this package only considers univariate time-series, one would need a bit of work to integrate it properly with MNE. A small tutorial would maybe help the user start in the right direction.

2. Andrew J Washington seems to be a significant contributor. Have you considered adding him in the author list ?

3. Some guidelines could be added, for instance linked in the README.md, for third parties wishing to report issues or seek support.

Other than these minor remarks, I consider that this submission does meet the JOSS criteria.
@arokem @voytek
