+++ README.md
@@ -30,9 +30,13 @@ We recommend using the [Anaconda](https://www.continuum.io/downloads) distributi
 
 ## Install
 
+**Stable Version**
+
 To install the latest release of neurodsp, you can install from pip:
 
 `$ pip install neurodsp`
+
+**Development Version**
 
 To get the development version (updates that are not yet published to pip), you can clone this repo.
 
@@ -41,6 +45,16 @@ To get the development version (updates that are not yet published to pip), you 
 To install this cloned copy of neurodsp, move into the directory you just cloned, and run:
 
 `$ pip install .`
+
+**Editable Version**
+
+If you want to install an editable version, for making contributions, download the development version as above, and run:
+
+`$ pip install -e .`
+
+## Bug Reports
+
+Please use the [Github issue tracker](https://github.com/neurodsp-tools/neurodsp/issues) to file bug reports and/or ask questions about this project.
 
 ## Modules
 
@@ -54,3 +68,11 @@ NeuroDSP includes the following modules, each of which have dedicated [tutorials
 - ```timefrequency``` : Estimate instantaneous measures of oscillatory activity
 - ```sim``` : Simulate periodic and aperiodic signal components
 - ```plts``` : Plotting functions
+
+## Contribute
+
+`NeuroDSP` welcomes and encourages contributions from the community, as long as contributors follow the [Code of Conduct](https://github.com/neurodsp-tools/neurodsp/blob/add-code-of-conduct-1/CODE_OF_CONDUCT.md).
+
+If you have an idea of something to add to NeuroDSP, please start by opening an [issue](https://github.com/neurodsp-tools/neurodsp/issues).
+
+When writing code to add to NeuroDSP, please follow the [Contribution Guidelines](https://github.com/neurodsp-tools/neurodsp/blob/master/CONTRIBUTING.md).
