### 2019-02-23 804b5f1
Update contributing file
