### 2019-02-23 #1272 @arokem
Thanks for the quick review!

@voytek : one potential typo on my reread of the manuscript. Should "1/f life components" be "1/f-like components"? Or is that a technical term I'm just not familiar with?
