### 2019-02-24 ba8cede
Continue refactor
