+++ neurodsp/filt.py
@@ -34,7 +34,7 @@ def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
         Should be a tuple of 2 floats for bandpass and bandstop.
         Can be a tuple of 2 floats, or a single float, for low/highpass.
         If float, it's taken as the cutoff frequency. If tuple, it's assumed
-        as (None,f_hi) for LP, and (f_lo,None) for HP.
+        as (None, f_hi) for LP, and (f_lo, None) for HP.
     n_cycles : float, optional, default: 3
         Length of filter in terms of number of cycles at 'f_lo' frequency.
         This parameter is overwritten by 'n_seconds', if provided.
@@ -70,38 +70,30 @@ def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
 
     # Check inputs & compute the nyquist frequency
     f_lo, f_hi = check_filter_definitions(pass_type, fc)
-    f_nyq = fs / 2.
 
     # Remove any NaN on the edges of 'sig'
-    first_nonan = np.where(~np.isnan(sig))[0][0]
-    last_nonan = np.where(~np.isnan(sig))[0][-1] + 1
-    sig_old = np.copy(sig)
-    sig = sig[first_nonan:last_nonan]
+    sig, sig_nans = remove_nans(sig)
 
     if iir:
         iir_checks(pass_type, n_seconds, butterworth_order, remove_edge_artifacts, verbose)
-        sig_filt, b_vals, a_vals = iir_filt(sig, pass_type, f_lo, f_hi, f_nyq,
+        sig_filt, b_vals, a_vals = iir_filt(sig, fs, pass_type, f_lo, f_hi,
                                             butterworth_order, plot_freq_response)
     else:
-        sig_filt, filt_len, kernel = fir_filt(sig, fs, pass_type, f_lo, f_hi, f_nyq, n_cycles,
+        sig_filt, filt_len, kernel = fir_filt(sig, fs, pass_type, f_lo, f_hi, n_cycles,
                                               n_seconds, plot_freq_response)
 
     # Compute transition bandwidth
     if compute_transition_band and verbose:
         if not iir:
             a_vals, b_vals = 1, kernel
-        calc_transition_band(b_vals, a_vals, fs, f_lo, f_hi, f_nyq, pass_type)
+        calc_transition_band(b_vals, a_vals, fs, f_lo, f_hi, pass_type)
 
     # Remove edge artifacts
     if not iir and remove_edge_artifacts:
-        n_rmv = int(np.ceil(filt_len / 2))
-        sig_filt[:n_rmv] = np.nan
-        sig_filt[-n_rmv:] = np.nan
+        sig_filt = drop_edge_artifacts(sig_filt, filt_len)
 
-    # Add NaN back on the edges of 'x', if there were any at the beginning
-    sig_filt_full = np.ones(len(sig_old)) * np.nan
-    sig_filt_full[first_nonan:last_nonan] = sig_filt
-    sig_filt = sig_filt_full
+    # Add NaN back on the edges of 'sig', if there were any at the beginning
+    sig_filt = restore_nans(sig_filt, sig_nans)
 
     # Return kernel if desired
     if return_kernel:
@@ -111,6 +103,39 @@ def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
             return sig_filt, kernel
     else:
         return sig_filt
+
+
+def remove_nans(sig):
+    """Words, words, words."""
+
+    sig_nans = np.isnan(sig)
+    sig_removed = sig[np.where(~np.isnan(sig))]
+
+    return sig_removed, sig_nans
+
+def restore_nans(sig, sig_nans):
+    """Words, words, words."""
+
[…]
