### 2019-02-24 ab950cb
Update API & organization to make FIR & IIR directly callable - more refactoring
