+++ neurodsp/filt.py
@@ -68,74 +68,97 @@ def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
         Filter kernel. Only returned if 'return_kernel' is True.
     """
 
-    # Check inputs & compute the nyquist frequency
-    f_lo, f_hi = check_filter_definitions(pass_type, fc)
-
-    # Remove any NaN on the edges of 'sig'
-    sig, sig_nans = remove_nans(sig)
-
     if iir:
-        iir_checks(pass_type, n_seconds, butterworth_order, remove_edge_artifacts, verbose)
-        sig_filt, b_vals, a_vals = iir_filt(sig, fs, pass_type, f_lo, f_hi,
-                                            butterworth_order, plot_freq_response)
+        _iir_checks(pass_type, n_seconds, butterworth_order, remove_edge_artifacts)
+        return iir_filt(sig, fs, pass_type, fc, butterworth_order, plot_freq_response, return_kernel, compute_transition_band)
     else:
-        sig_filt, filt_len, kernel = fir_filt(sig, fs, pass_type, f_lo, f_hi, n_cycles,
-                                              n_seconds, plot_freq_response)
+        return fir_filt(sig, fs, pass_type, fc, n_cycles, n_seconds, plot_freq_response, return_kernel, compute_transition_band, remove_edge_artifacts)
+
+
+def fir_filt(sig, fs, pass_type, fc, n_cycles, n_seconds, plot_freq_response, return_kernel, compute_transition_band, remove_edge_artifacts):
+    """Words, words, words."""
+
+    # Check filter definition
+    f_lo, f_hi = check_filter_definitions(pass_type, fc)
+    filt_len = _fir_checks(sig, fs, pass_type, f_lo, f_hi, n_cycles, n_seconds)
+
+    # Design filter
+    f_nyq = compute_nyquist(fs)
+    if pass_type == 'bandpass':
+        kernel = signal.firwin(filt_len, (f_lo, f_hi), pass_zero=False, nyq=f_nyq)
+    elif pass_type == 'bandstop':
+        kernel = signal.firwin(filt_len, (f_lo, f_hi), nyq=f_nyq)
+    elif pass_type == 'highpass':
+        kernel = signal.firwin(filt_len, f_lo, pass_zero=False, nyq=f_nyq)
+    elif pass_type == 'lowpass':
+        kernel = signal.firwin(filt_len, f_hi, nyq=f_nyq)
 
     # Compute transition bandwidth
-    if compute_transition_band and verbose:
-        if not iir:
-            a_vals, b_vals = 1, kernel
-        calc_transition_band(b_vals, a_vals, fs, f_lo, f_hi, pass_type)
+    if compute_transition_band:
+        a_vals, b_vals = 1, kernel
+        check_filter_properties(b_vals, a_vals, fs, pass_type, f_lo, f_hi)
+
+    # Remove any NaN on the edges of 'sig'
+    sig, sig_nans = _remove_nans(sig)
+
+    # Apply filter
+    sig_filt = np.convolve(kernel, sig, 'same')
 
     # Remove edge artifacts
-    if not iir and remove_edge_artifacts:
-        sig_filt = drop_edge_artifacts(sig_filt, filt_len)
+    if remove_edge_artifacts:
+        sig_filt = _drop_edge_artifacts(sig_filt, filt_len)
 
     # Add NaN back on the edges of 'sig', if there were any at the beginning
-    sig_filt = restore_nans(sig_filt, sig_nans)
+    sig_filt = _restore_nans(sig_filt, sig_nans)
 
-    # Return kernel if desired
+    # Plot frequency response, if desired
+    if plot_freq_response:
[…]
