### 2019-02-24 04d19c3
Refactor forever: split out filter design
