+++ neurodsp/filt.py
@@ -69,34 +69,28 @@ def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
     """
 
     if iir:
-        _iir_checks(pass_type, n_seconds, butterworth_order, remove_edge_artifacts)
-        return iir_filt(sig, fs, pass_type, fc, butterworth_order, plot_freq_response, return_kernel, compute_transition_band)
+        _iir_checks(n_seconds, butterworth_order, remove_edge_artifacts)
+        return filter_signal_iir(sig, fs, pass_type, fc, butterworth_order, plot_freq_response,
+                                 return_kernel, compute_transition_band)
     else:
-        return fir_filt(sig, fs, pass_type, fc, n_cycles, n_seconds, plot_freq_response, return_kernel, compute_transition_band, remove_edge_artifacts)
+        return filter_signal_fir(sig, fs, pass_type, fc, n_cycles, n_seconds, plot_freq_response,
+                                 return_kernel, compute_transition_band, remove_edge_artifacts)
 
 
-def fir_filt(sig, fs, pass_type, fc, n_cycles, n_seconds, plot_freq_response, return_kernel, compute_transition_band, remove_edge_artifacts):
+
+
+
+def filter_signal_fir(sig, fs, pass_type, fc, n_cycles, n_seconds, plot_freq_response,
+                      return_kernel, compute_transition_band, remove_edge_artifacts):
     """Words, words, words."""
 
-    # Check filter definition
-    f_lo, f_hi = check_filter_definitions(pass_type, fc)
-    filt_len = _fir_checks(sig, fs, pass_type, f_lo, f_hi, n_cycles, n_seconds)
-
     # Design filter
-    f_nyq = compute_nyquist(fs)
-    if pass_type == 'bandpass':
-        kernel = signal.firwin(filt_len, (f_lo, f_hi), pass_zero=False, nyq=f_nyq)
-    elif pass_type == 'bandstop':
-        kernel = signal.firwin(filt_len, (f_lo, f_hi), nyq=f_nyq)
-    elif pass_type == 'highpass':
-        kernel = signal.firwin(filt_len, f_lo, pass_zero=False, nyq=f_nyq)
-    elif pass_type == 'lowpass':
-        kernel = signal.firwin(filt_len, f_hi, nyq=f_nyq)
+    kernel = design_fir_filter(pass_type, fc, n_cycles, n_seconds, fs, len(sig))
 
     # Compute transition bandwidth
     if compute_transition_band:
         a_vals, b_vals = 1, kernel
-        check_filter_properties(b_vals, a_vals, fs, pass_type, f_lo, f_hi)
+        check_filter_properties(b_vals, a_vals, fs, pass_type, fc)
 
     # Remove any NaN on the edges of 'sig'
     sig, sig_nans = _remove_nans(sig)
@@ -106,7 +100,7 @@ def fir_filt(sig, fs, pass_type, fc, n_cycles, n_seconds, plot_freq_response, re
 
     # Remove edge artifacts
     if remove_edge_artifacts:
-        sig_filt = _drop_edge_artifacts(sig_filt, filt_len)
+        sig_filt = _drop_edge_artifacts(sig_filt, len(kernel))
 
     # Add NaN back on the edges of 'sig', if there were any at the beginning
     sig_filt = _restore_nans(sig_filt, sig_nans)
@@ -121,26 +115,16 @@ def fir_filt(sig, fs, pass_type, fc, n_cycles, n_seconds, plot_freq_response, re
         return sig_filt
 
 
-def iir_filt(sig, fs, pass_type, fc, butterworth_order, plot_freq_response, return_kernel, compute_transition_band):
+def filter_signal_iir(sig, fs, pass_type, fc, butterworth_order, plot_freq_response,
[…]
