+++ neurodsp/filt.py
@@ -19,9 +19,9 @@ def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
     Parameters
     ----------
     sig : 1d array
-        Time series.
+        Time series to be filtered.
     fs : float
-        The sampling rate, in Hz.
+        Sampling rate, in Hz.
     pass_type : {'bandpass', 'bandstop', 'lowpass', 'highpass'}
         Which kind of filter to apply:
 
@@ -29,43 +29,40 @@ def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
         * 'bandstop': apply a bandstop (notch) filter
         * 'lowpass': apply a lowpass filter
         * 'highpass' : apply a highpass filter
-    fc : tuple or float
-        Cutoff frequency(ies) used for filter.
-        Should be a tuple of 2 floats for bandpass and bandstop.
-        Can be a tuple of 2 floats, or a single float, for low/highpass.
-        If float, it's taken as the cutoff frequency. If tuple, it's assumed
-        as (None, f_hi) for LP, and (f_lo, None) for HP.
+    fc : tuple of (float, float) or float
+        Cutoff frequency(ies) used for filter, specified as f_lo & f_hi.
+        For 'bandpass' & 'bandstop', must be a tuple.
+        For 'lowpass' or 'highpass', can be a float that specifies pass frequency, or can be
+        a tuple and is assumed to be (None, f_hi) for 'lowpass', and (f_lo, None) for 'highpass'.
     n_cycles : float, optional, default: 3
-        Length of filter in terms of number of cycles at 'f_lo' frequency.
-        This parameter is overwritten by 'n_seconds', if provided.
+        Length of filter, in number of cycles, defined at the 'f_lo' frequency.
+        This parameter is overwritten by `n_seconds`, if provided.
     n_seconds : float, optional
         Length of filter, in seconds.
+        This parameter overwrites `n_cycles`.
     iir : bool, optional
         If True, use an infinite-impulse response (IIR) filter.
-        The only IIR filter to be used is a butterworth filter.
+        The only IIR filter offered is a butterworth filter.
     butterworth_order : int, optional
-        Order of the butterworth filter.
+        Order of the butterworth filter, if using an IIR filter.
         See input 'N' in scipy.signal.butter.
     plot_freq_response : bool, optional
         If True, plot the frequency response of the filter
     return_kernel : bool, optional
-        If True, return the complex filter kernel
+        If True, return the complex filter kernel.
     verbose : bool, optional
-        If True, print optional information
+        If True, print optional information.
     compute_transition_band : bool, optional
-        If True, the function computes the transition bandwidth,
-        defined as the frequency range between -20dB and -3dB attenuation,
-        and warns the user if this band is longer than the frequency bandwidth.
+        If True, computes the transition bandwidth, and prints this information.
     remove_edge_artifacts : bool, optional
-        If True, replace the samples that are within half a kernel's length to
-        the signal edge with np.nan.
+        If True, replace samples within half the kernel length to be np.nan.
 
     Returns
     -------
     sig_filt : 1d array
         Filtered time series.
     kernel : length-2 tuple of arrays
-        Filter kernel. Only returned if 'return_kernel' is True.
+        Filter kernel. Only returned if `return_kernel` is True.
     """
 
     if iir:
[…]
