### 2019-02-24 3380dfc
Lint & small fixes
