### 2019-02-24 bf6ddf4
Fill in default values and associated fixes
