+++ neurodsp/filt.py
@@ -13,7 +13,7 @@ from neurodsp.plts.filt import plot_frequency_response
 def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
                   iir=False, butterworth_order=None,
                   plot_freq_response=False, return_kernel=False,
-                  verbose=True, compute_transition_band=True, remove_edge_artifacts=True):
+                  verbose=False, compute_transition_band=True, remove_edge_artifacts=True):
     """Apply a bandpass, bandstop, highpass, or lowpass filter to a neural signal.
 
     Parameters
@@ -46,15 +46,15 @@ def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
     butterworth_order : int, optional
         Order of the butterworth filter, if using an IIR filter.
         See input 'N' in scipy.signal.butter.
-    plot_freq_response : bool, optional
+    plot_freq_response : bool, optional, default: False
         If True, plot the frequency response of the filter
-    return_kernel : bool, optional
+    return_kernel : bool, optional, default: False
         If True, return the complex filter kernel.
-    verbose : bool, optional
+    verbose : bool, optional, default: False
         If True, print optional information.
-    compute_transition_band : bool, optional
+    compute_transition_band : bool, optional, default: True
         If True, computes the transition bandwidth, and prints this information.
-    remove_edge_artifacts : bool, optional
+    remove_edge_artifacts : bool, optional, default: True
         If True, replace samples within half the kernel length to be np.nan.
 
     Returns
@@ -74,8 +74,8 @@ def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
                                  return_kernel, compute_transition_band, remove_edge_artifacts)
 
 
-def filter_signal_fir(sig, fs, pass_type, fc, n_cycles, n_seconds, plot_freq_response,
-                      return_kernel, compute_transition_band, remove_edge_artifacts):
+def filter_signal_fir(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None, plot_freq_response=False,
+                      return_kernel=False, compute_transition_band=True, remove_edge_artifacts=True):
     """Apply an FIR filter to a signal.
 
     Parameters
@@ -102,9 +102,9 @@ def filter_signal_fir(sig, fs, pass_type, fc, n_cycles, n_seconds, plot_freq_res
     n_seconds : float, optional
         Length of filter, in seconds.
         This parameter overwrites `n_cycles`.
-    plot_freq_response : bool, optional
+    plot_freq_response : bool, optional, default: False
         If True, plot the frequency response of the filter
-    return_kernel : bool, optional
+    return_kernel : bool, optional, default: False
         If True, return the complex filter kernel.
     compute_transition_band : bool, optional
         If True, computes the transition bandwidth, and prints this information.
@@ -120,7 +120,7 @@ def filter_signal_fir(sig, fs, pass_type, fc, n_cycles, n_seconds, plot_freq_res
     """
 
     # Design filter
-    kernel = design_fir_filter(pass_type, fc, n_cycles, n_seconds, fs, len(sig))
+    kernel = design_fir_filter(len(sig), fs, pass_type, fc, n_cycles, n_seconds)
 
     # Compute transition bandwidth
     if compute_transition_band:
@@ -150,8 +150,8 @@ def filter_signal_fir(sig, fs, pass_type, fc, n_cycles, n_seconds, plot_freq_res
         return sig_filt
 
 
[…]
