### 2019-02-24 a27cc35
Add new filter function organization to docsite
