### 2019-02-25 e265bef
Add note of tutorials in documentation section
