### 2019-02-25 c8a2d3e
1/f life > 1/f-life
