+++ paper/paper.md
@@ -41,7 +41,7 @@ bibliography: paper.bib
 
 # Summary
 
-Populations of neurons exhibit time-varying fluctuations in their aggregate electrical electrophysiological methods, such as magneto or electroencephalography (M/EEG), intracranial EEG (iEEG) or electrocorticography (ECoG), and local field potential (LFP) recordings [@buzsaki_origin_2012]. While there are existing Python tools for digital signal processing (DSP), such as [scipy.signal](https://docs.scipy.org/doc/scipy/reference/signal.html), neural data exhibit specific properties that warrant specialized analysis tools focused on idiosyncrasies of neural data. Features of interest in neural data include periodic properties—such as band-limited oscillations [@buzsaki_neural_2004] and transient or 'bursty' events—as well as an aperiodic signal that is variously referred to as the 1/f-like background [@freeman_simulated_2009, @miller_power-law_2009], or noise [@voytek_age-related_2015], or scale-free activity [@he_scale-free_2014], and that may carry information about the current generators, such as the ratio of excitation and inhibition [@gao_inferring_2017]. ``NeuroDSP`` is a package specifically designed to be used by neuroscientists for analyzing neural time series data, in particular for examing their time-varying properties related to oscillatory and 1/f life components.
+Populations of neurons exhibit time-varying fluctuations in their aggregate electrical electrophysiological methods, such as magneto or electroencephalography (M/EEG), intracranial EEG (iEEG) or electrocorticography (ECoG), and local field potential (LFP) recordings [@buzsaki_origin_2012]. While there are existing Python tools for digital signal processing (DSP), such as [scipy.signal](https://docs.scipy.org/doc/scipy/reference/signal.html), neural data exhibit specific properties that warrant specialized analysis tools focused on idiosyncrasies of neural data. Features of interest in neural data include periodic properties—such as band-limited oscillations [@buzsaki_neural_2004] and transient or 'bursty' events—as well as an aperiodic signal that is variously referred to as the 1/f-like background [@freeman_simulated_2009, @miller_power-law_2009], or noise [@voytek_age-related_2015], or scale-free activity [@he_scale-free_2014], and that may carry information about the current generators, such as the ratio of excitation and inhibition [@gao_inferring_2017]. ``NeuroDSP`` is a package specifically designed to be used by neuroscientists for analyzing neural time series data, in particular for examing their time-varying properties related to oscillatory and 1/f-like components.
 
 ``NeuroDSP`` is accompanied by a [documentation site](https://neurodsp-tools.github.io/neurodsp/) that includes detailed [tutorials](https://neurodsp-tools.github.io/neurodsp/auto_tutorials/index.html#) for each of the modules, which are described below, as well as suggested workflows for combining them.
 
