### 2019-02-25 #1272 @voytek
> @voytek : one potential typo on my reread of the manuscript. Should "1/f life components" be "1/f-like components"? Or is that a technical term I'm just not familiar with?

Too funny. Living my 1/f life. Good catch, @arokem. Fixed.
