### 2019-02-26 8a06b4b
fixed spectral.compute_spectrum doc for noverlap
