### 2019-02-26 a93fd28
made a note in sim on floating point precision causing kernel length to be imprecise
