### 2019-02-27 0092b81
add feingold burst reference
