+++ paper/paper.bib
@@ -43,6 +43,20 @@
     month = apr
 }
 
+@article{feingold_bursts_2015,
+	title = {Bursts of beta oscillation differentiate postperformance activity in the striatum and motor cortex of monkeys performing movement tasks},
+	volume = {112},
+	issn = {0027-8424, 1091-6490},
+	url = {http://www.pnas.org/lookup/doi/10.1073/pnas.1517629112},
+	doi = {10.1073/pnas.1517629112},
+	pages = {13687--13692},
+	number = {44},
+	journaltitle = {Proceedings of the National Academy of Sciences},
+	author = {Feingold, Joseph and Gibson, Daniel J. and {DePasquale}, Brian and Graybiel, Ann M.},
+  year = {2015},
+  month = nov
+}
+
 @article{fransen_identifying_2015,
     title = {Identifying neuronal oscillations using rhythmicity},
     volume = {118},
@@ -69,16 +83,6 @@
     author = {Freeman, Walter J. and Zhai, Jian},
     year = {2009},
     month = mar
-}
-
-@article{haller_parameterizing_2018,
-    title = {Parameterizing neural power spectra},
-    url = {http://biorxiv.org/lookup/doi/10.1101/299859},
-    doi = {10.1101/299859},
-    journal = {{bioRxiv}},
-    author = {Haller, Matar and Donoghue, Thomas and Peterson, Erik and Varma, Paroma and Sebastian, Priyadarshini and Gao, Richard and Noto, Torben and Knight, Robert T. and Shestyuk, Avgusta and Voytek, Bradley},
-    year = {2018},
-    month = apr
 }
 
 @article{gao_inferring_2017,
@@ -120,6 +124,16 @@
     month = feb,
     year = {2014},
     pages = {446--460}
+}
+
+@article{haller_parameterizing_2018,
+    title = {Parameterizing neural power spectra},
+    url = {http://biorxiv.org/lookup/doi/10.1101/299859},
+    doi = {10.1101/299859},
+    journal = {{bioRxiv}},
+    author = {Haller, Matar and Donoghue, Thomas and Peterson, Erik and Varma, Paroma and Sebastian, Priyadarshini and Gao, Richard and Noto, Torben and Knight, Robert T. and Shestyuk, Avgusta and Voytek, Bradley},
+    year = {2018},
+    month = apr
 }
 
 @article{he_scale-free_2014,
