+++ paper/paper.md
@@ -58,7 +58,7 @@ The modules included in ``NeuroDSP`` are:
 
 # Statement of Need
 
-``NeuroDSP`` is an open-source Python package for time-series analyses of neural data, including implementations of relevant DSP utilities as well as implementations of a collection of algorithms and approaches that have been developed specifically for neural data analysis. By design, ``NeuroDSP`` offers a lightweight architecture in which functions take in time-series directly, thus offering a flexible toolbox for custom analysis of a broad range of neural data. This approach complements, and can be used in conjunction with, related toolboxes such as MNE [@gramfort_mne_2014] that are more focused on data management and multi-channel analyses. ``NeuroDSP`` offers implementations of a distinct set of methods, with different use cases from other tools, and can easily be integrated with frameworks such as MNE or with other custom workflows. ``NeuroDSP`` also offers a developed module for simulating realistic field potential data, which is used for testing the properties of methods against synthetic data for which ground truth parameters are known.
+``NeuroDSP`` is an open-source Python package for time-series analyses of neural data, including implementations of relevant DSP utilities as well as implementations of a collection of algorithms and approaches that have been developed specifically for neural data analysis. By design, ``NeuroDSP`` offers a lightweight architecture in which functions take in time-series directly, thus offering a flexible toolbox for custom analysis of a broad range of neural data. This approach complements, and can be used in conjunction with, related toolboxes such as MNE [@gramfort_mne_2014] that are more focused on data management and multi-channel analyses. ``NeuroDSP`` offers implementations of a distinct set of methods, with different use cases from other tools, and can easily be integrated with frameworks such as MNE or with other custom workflows. ``NeuroDSP`` also offers a developed module for simulating realistic field potential data, which is used for testing the properties of methods against synthetic data for which ground truth parameters are known. Note that these simulations are designed to mimic the statistics of electrophysiological data—and properties of transient, non-stationary, non-sinusoidal rhythms—but they should not be over-interpreted as biophysically realistic.
 
 # Acknowledgements
 
