### 2019-02-27 neurodsp-tools/neurodsp#143 @voytek
Add MNE integration tutorial
Per the code review during the _Journal of Open Source Software_ [review of our code](https://github.com/openjournals/joss-reviews/issues/1272), we need to add a tutorial showing how to integrate neurodsp with MNE.
