### 2019-02-27 #1272 @voytek
> The API is clear, with consistent parameter names and orderings, and well documented.

Thanks, @TomDLT! The lab's been very diligent about our software development.

> It also implements less standard signal processing techniques, such as lagged-coherence (Fransen et al, 2015), sliding window matching (Gips et al, 2017), or burst detection (reference needed).

We have now added a reference to the dual-threshold burst algorithm: Feingold et al., PNAS 2015: Bursts of beta oscillation differentiate post-performance activity in the striatum and motor cortex of monkeys performing movement tasks.

>Some comments might be added to emphasize that these are only models and could lead to oversimplistic signals, compared to empirical neurophysiological signals.

Excellent point. I've added the following comment to the paper: "Note that these simulations are designed to mimic the statistics of electrophysiological data—and properties of transient, non-stationary, non-sinusoidal rhythms—but they should not be over-interpreted as biophysically realistic."

> 1. The authors state in the paper that this package "can easily be integrated with frameworks such as MNE". As this package only considers univariate time-series, one would need a bit of work to integrate it properly with MNE. A small tutorial would maybe help the user start in the right direction.

This is an excellent suggestion. I've [opened an issue](https://github.com/neurodsp-tools/neurodsp/issues/143) to ensure that this gets addressed soon.

> 2. Andrew J Washington seems to be a significant contributor. Have you considered adding him in the author list ?

Andrew was indeed helpful, but did not contribute significantly enough to the intellectual aspects of this project to warrant co-authorship, but we now acknowledge his help. We added the following to the Acknowledgements section: "We thank Andrew J. Washington for his early code contributions to this project." Thank you for catching this!

> 3. Some guidelines could be added, for instance linked in the README.md, for third parties wishing to report issues or seek support.

We have added explicit mention in the project README of how to ask questions and file bug reports (using Github issues), as well as a note on how to contribute, with associated updates to the CONTRIBUTING file, explaining the ideal workflow - see pull request #137 on the neurodsp repository.
