### 2019-02-28 342ce42
Add outline of glossary
