+++ doc/glossary.rst
@@ -1,0 +1,65 @@
+NeuroDSP Glossary
+=================
+
+General
+-------
+
+.. glossary::
+
+    time domain
+        xx
+    frequency domain
+        xx
+    convolution
+        xx
+    sampling rate
+        xx
+    coherence
+        xx
+
+Time Frequency
+--------------
+
+.. glossary::
+
+    frequency
+        xx
+    phase
+        xx
+    amplitude
+        xx
+
+Filters
+-------
+
+.. glossary::
+
+    FIR
+        xx
+    IIR
+        xx
+    pass type (bandpass, bandstop, lowpass, highpass)
+        xx
+    pass band
+        xx
+    transition band
+        xx
+    kernel
+        xx
+    frequency response
+        xx
+    kernel
+        xx
+    b_vals / a_vals
+        xx
+
+Spectral
+--------
+.. glossary::
+
+    fourier transform
+        xx
+    power spectrum
+        xx
+    wavelet
+        xx
