### 2019-02-28 1ba97b6
Separate out settings checking as a sub-function
