### 2019-02-28 76db565
Split out outlier subfunction - fix 1d issue in compute spectrum
