### 2019-02-28 8b5c968
Refactor compute spectrum
