+++ neurodsp/spectral.py
@@ -8,8 +8,7 @@ from scipy import signal
 
 def compute_spectrum(sig, fs, method='mean', window='hann', nperseg=None,
                      noverlap=None, filt_len=1., f_lim=None, spg_outlier_pct=0.):
-    """Estimate the power spectral density (PSD) of a time series from short-time Fourier
-    Transform (mean, median), or the entire signal's FFT smoothed (medfilt).
+    """Estimate the power spectral density (PSD) of a time series.
 
     Parameters
     -----------
@@ -58,46 +57,117 @@ def compute_spectrum(sig, fs, method='mean', window='hann', nperseg=None,
     https://docs.scipy.org/doc/scipy/reference/generated/scipy.signal.spectrogram.html
     """
 
-    # welch-style spectrum (mean/median of STFT)
-    if method in ('mean', 'median'):
-
-        # Calculate the short time fourier transform with signal.spectrogram
-        nperseg, noverlap = _check_settings(fs, window, nperseg, noverlap)
-        freqs, t_axis, spg = signal.spectrogram(sig, fs, window, nperseg, noverlap)
-
-        # Pad data to 2D
-        if len(sig.shape) == 1:
-            sig = sig[None, :]
-
-        numchan = sig.shape[0]
-
-        # Throw out outliers if indicated
-        if spg_outlier_pct > 0.:
-            spg = discard_outliers(spg, spg_outlier_pct)
-
-        if method == 'mean':
-            spectrum = np.mean(spg, axis=-1)
-        elif method == 'median':
-            spectrum = np.median(spg, axis=-1)
-
-    elif method == 'medfilt':
-
-        # Take the positive half of the spectrum since it's symmetrical
-        FT = np.fft.fft(sig)[:int(np.ceil(len(sig) / 2.))]
-        freqs = np.fft.fftfreq(len(sig), 1. / fs)[:int(np.ceil(len(sig) / 2.))]  # get freq axis
-
-        # Convert median filter length from Hz to samples
-        filt_len_samp = int(int(filt_len / (freqs[1] - freqs[0])) / 2 * 2 + 1)
-        spectrum = signal.medfilt(np.abs(FT)**2. / (fs * len(sig)), filt_len_samp)
-
-    else:
+    if method not in ('mean', 'median', 'medfilt'):
         raise ValueError('Unknown power spectrum method: %s' % method)
 
-    if f_lim is not None:
-        f_lim_ind = np.where(freqs > f_lim)[0][0]
-        return freqs[:f_lim_ind], spectrum[..., :f_lim_ind]
-    else:
-        return freqs, spectrum
+    if method in ('mean', 'median'):
+        return compute_spectrum_welch(sig, fs, method, window, nperseg, noverlap, f_lim, spg_outlier_pct)
+
+    elif method == 'medfilt':
+        return compute_spectrum_medfilt(sig, fs, filt_len, f_lim)
+
+
+def compute_spectrum_welch(sig, fs, method='mean', window='hann', nperseg=None,
+                           noverlap=None, f_lim=None, spg_outlier_pct=0.):
+    """Estimate the power spectral density using Welch's method.
+
+    Parameters
+    -----------
+    sig : 1d or 2d array
+        Time series of measurement values.
+    fs : float
+        Sampling rate, in Hz.
+    method : {'mean', 'median'}, optional
+        Method to average across the windows:
+
+        * 'mean' is the same as Welch's method (mean of STFT).
+        * 'median' uses median of STFT instead of mean to minimize outlier effect.
+    window : str or tuple or array_like, optional, default: 'hann'
+        Desired window to use. Defaults to a Hann window.
+        See scipy.signal.get_window for a list of windows and required parameters.
[…]
