+++ neurodsp/spectral.py
@@ -25,7 +25,7 @@ def compute_spectrum(sig, fs, method='mean', window='hann', nperseg=None,
     window : str or tuple or array_like, optional, default: 'hann'
         Desired window to use. Defaults to a Hann window.
         See scipy.signal.get_window for a list of windows and required parameters.
-        If window is array_like, this array will be used as the window and its length must be nperseg.
+        If array_like, the array will be used as the window and its length must be nperseg.
     nperseg : int, optional
         Length of each segment, in number of samples.
         If None, and window is str or tuple, is set to 1 second of data.
@@ -36,7 +36,7 @@ def compute_spectrum(sig, fs, method='mean', window='hann', nperseg=None,
     filt_len : float, optional
         Length of the median filter, in Hz.
         Only used with the 'medfilt' method.
-    f_lim : float, optional
+    f_lim : float, optional, default: 1.
         Maximum frequency to keep, in Hz.
         If None, keeps up to Nyquist.
     spg_outlier_pct : float, optional, default: 0.
@@ -61,7 +61,8 @@ def compute_spectrum(sig, fs, method='mean', window='hann', nperseg=None,
         raise ValueError('Unknown power spectrum method: %s' % method)
 
     if method in ('mean', 'median'):
-        return compute_spectrum_welch(sig, fs, method, window, nperseg, noverlap, f_lim, spg_outlier_pct)
+        return compute_spectrum_welch(sig, fs, method, window,
+                                      nperseg, noverlap, f_lim, spg_outlier_pct)
 
     elif method == 'medfilt':
         return compute_spectrum_medfilt(sig, fs, filt_len, f_lim)
@@ -85,7 +86,7 @@ def compute_spectrum_welch(sig, fs, method='mean', window='hann', nperseg=None,
     window : str or tuple or array_like, optional, default: 'hann'
         Desired window to use. Defaults to a Hann window.
         See scipy.signal.get_window for a list of windows and required parameters.
-        If window is array_like, this array will be used as the window and its length must be nperseg.
+        If array_like, the array will be used as the window and its length must be nperseg.
     nperseg : int, optional
         Length of each segment, in number of samples.
         If None, and window is str or tuple, is set to 1 second of data.
@@ -110,8 +111,8 @@ def compute_spectrum_welch(sig, fs, method='mean', window='hann', nperseg=None,
     """
 
     # Calculate the short time fourier transform with signal.spectrogram
-    nperseg, noverlap = _check_settings(fs, window, nperseg, noverlap)
-    freqs, t_axis, spg = signal.spectrogram(sig, fs, window, nperseg, noverlap)
+    nperseg, noverlap = _check_spg_settings(fs, window, nperseg, noverlap)
+    freqs, _, spg = signal.spectrogram(sig, fs, window, nperseg, noverlap)
 
     # Pad data to 2D
     if len(sig.shape) == 1:
@@ -119,7 +120,7 @@ def compute_spectrum_welch(sig, fs, method='mean', window='hann', nperseg=None,
 
     # Throw out outliers if indicated
     if spg_outlier_pct > 0.:
-        spg = discard_outliers(spg, spg_outlier_pct)
+        spg = _discard_outliers(spg, spg_outlier_pct)
 
     if method == 'mean':
         spectrum = np.mean(spg, axis=-1)
@@ -132,7 +133,7 @@ def compute_spectrum_welch(sig, fs, method='mean', window='hann', nperseg=None,
     return freqs, spectrum
 
 
-def compute_spectrum_medfilt(sig, fs, filt_len, f_lim=None):
+def compute_spectrum_medfilt(sig, fs, filt_len=1., f_lim=None):
     """Estimate the power spectral densitry as a smoothed FFT.
 
[…]
