### 2019-03-02 cd073cc
Add examples link to navbar
