### 2019-03-02 b635385
Add draft of MNE example
