### 2019-03-03 b56a871
WIP rework filter tests
