+++ neurodsp/tests/test_filt.py
@@ -1,122 +1,173 @@
 """Test filtering functions."""
 
-from pytest import raises
+from pytest import raises, warns
 
 import numpy as np
+from numpy.testing import assert_equal
 
-from neurodsp.filt import filter_signal
-from .util import load_example_data, generate_random_signal
+from neurodsp.filt import *
+from neurodsp.filt import _remove_nans, _restore_nans, _drop_edge_artifacts, _fir_checks, _iir_checks
 
 ###################################################################################################
 ###################################################################################################
 
-def test_bandpass_filter_consistent():
+def test_filter_signal():
+    # TODO
+    pass
 
-    # Load data and ground-truth filtered signal
-    sig, sig_filt_true = load_example_data(data_idx=1, filtered=True)
+def test_filter_signal_fir():
+    # TODO
+    pass
 
-    # filter data
-    fs = 1000
-    fc = (13, 30)
-    sig_filt = filter_signal(sig, fs, 'bandpass', fc=fc, n_cycles=3)
+def test_filter_signal_iir():
+    # TODO
+    pass
 
-    # Compute difference between current and past filtered signals
-    signal_diff = sig_filt[~np.isnan(sig_filt)] - sig_filt_true[~np.isnan(sig_filt_true)]
-    assert np.allclose(np.sum(np.abs(signal_diff)), 0, atol=10 ** -5)
+def test_design_fir_filter():
+    # TODO
+    pass
 
+def test_design_iir_filter():
+    # TODO
+    pass
 
-def test_edge_nan():
-    """
-    Confirm that the appropriate amount of edge artifact has been removed for FIR filters
-    and that edge artifacts are not removed for IIR filters
-    """
+def test_check_filter_definition():
 
-    # Apply a 4-8Hz bandpass filter to random noise
-    sig = generate_random_signal()
-    sig_filt, kernel = filter_signal(sig, 1000, 'bandpass', fc=(4, 8), return_kernel=True)
-
-    # Check if the correct edge artifacts have been removed
-    n_rmv = int(np.ceil(len(kernel) / 2))
-    assert all(np.isnan(sig_filt[:n_rmv]))
-    assert all(np.isnan(sig_filt[-n_rmv:]))
-    assert all(~np.isnan(sig_filt[n_rmv:-n_rmv]))
-
-    # Check that no edge artifacts are removed for IIR filters
-    sig_filt = filter_signal(sig, 1000, 'bandpass', fc=(4, 8), iir=True, butterworth_order=3)
-    assert all(~np.isnan(sig_filt))
-
-
-def test_filter_length_error():
-    """Check error is raised if the filter designed is longer than the signal."""
-
-    n_seconds = 2
-    fs = 1000
-    sig = np.random.randn(n_seconds * fs)
-    with raises(ValueError) as excinfo:
-        sig_filt = filter_signal(sig, fs, 'bandpass', fc=(1, 10))
-    assert 'The filter needs to be shortened by decreasing the n_cycles' in str(excinfo.value)
-
-
-def test_frequency_input_errors():
-    """Check that errors are properly raised when incorrectly enter frequency information."""
-
-    # Generate a random signal
-    sig = generate_random_signal()
-
-    # Check that a bandpass filter cannot be completed without proper frequency limits
+    # Check that error catching works for bad pass_type definition
     with raises(ValueError):
-        sig_filt = filter_signal(sig, 1000, 'bandpass', fc=8)
-    with raises(ValueError):
-        sig_filt = filter_signal(sig, 1000, 'bandpass', fc=(8, 4))
[…]
