### 2019-03-03 9fe2eb5
Update tf to use same util funcs as developed in filt
