### 2019-03-04 #1272 @arokem
OK. Looks like all the boxes are checked and looks like the one remaining concern (MNE tutorial) is currently in progress here: https://github.com/neurodsp-tools/neurodsp/pull/146

Once that's merged, please ping @TomDLT to take another look and confirm that the review is complete.
