### 2019-03-04 d4c9269
Move & rename 'infer_passtype to filt & use in tf
