+++ neurodsp/filt.py
@@ -546,6 +546,37 @@ def compute_nyquist(fs):
 
     return fs / 2.
 
+
+def infer_passtype(fc):
+    """Given frequency definition of a filter, infer the passtype.
+
+    Parameters
+    ----------
+    fc : tuple of (float, float)
+        Cutoff frequency(ies) used for filter, specified as f_lo & f_hi.
+
+    Returns
+    -------
+    pass_type : str
+        Which kind of filter pass_type is consistent with the frequency definition provided.
+
+    Notes
+    -----
+    Assumes that a definition with two frequencies is a 'bandpass' (not 'bandstop').
+    """
+
+    if fc[0] is None:
+        pass_type = 'lowpass'
+    elif fc[1] is None:
+        pass_type = 'highpass'
+    else:
+        pass_type = 'bandpass'
+
+    # Check the inferred passtype & frequency definition is valid
+    _ = check_filter_definition(pass_type, fc)
+
+    return pass_type
+
 ###################################################################################################
 ###################################################################################################
 
