+++ neurodsp/timefrequency.py
@@ -6,7 +6,7 @@ import numpy as np
 from scipy import signal
 
 import neurodsp
-from neurodsp.filt import filter_signal
+from neurodsp.filt import filter_signal, infer_passtype
 from neurodsp.filt import _drop_edge_artifacts, _remove_nans, _restore_nans
 
 ###################################################################################################
@@ -45,7 +45,7 @@ def phase_by_time(sig, fs, f_range, filter_kwargs={}, hilbert_increase_n=False,
     if filter_kwargs is None:
         filter_kwargs = {}
 
-    pass_type = _get_filt_passtype(f_range)
+    pass_type = infer_passtype(f_range)
 
     # Filter signal
     sig_filt, kernel = filter_signal(sig, fs, pass_type, fc=f_range,
@@ -96,7 +96,7 @@ def amp_by_time(sig, fs, f_range, filter_kwargs=None, hilbert_increase_n=False,
     if filter_kwargs is None:
         filter_kwargs = {}
 
-    pass_type = _get_filt_passtype(f_range)
+    pass_type = infer_passtype(f_range)
 
     # Filter signal
     sig_filt, kernel = filter_signal(sig, fs, pass_type, fc=f_range,
@@ -180,21 +180,3 @@ def _hilbert_ignore_nan(sig, hilbert_increase_n=False):
     sig_hilb = _restore_nans(sig_hilb_nonan, sig_nans, dtype=complex)
 
     return sig_hilb
-
-
-def _get_filt_passtype(f_range):
-    """Given frequency range of filter, check for Nones to return
-    appropriate filter pass_type to filt.filter_signal.
-    """
-
-    if f_range[0] is None:
-        pass_type = 'lowpass'
-    elif f_range[1] is None:
-        pass_type = 'highpass'
-    else:
-        if f_range[0] >= f_range[1]:
-            raise ValueError('Second cutoff frequency must be greater than first.')
-        else:
-            pass_type = 'bandpass'
-
-    return pass_type
