### 2019-03-04 66ca4b8
Rework layout for tf tests
