+++ neurodsp/tests/test_timefreq.py
@@ -1,16 +1,41 @@
-"""Test functions in the time-frequency analysis module."""
-
-import os
+"""Test functions for the time-frequency analysis module."""
 
 import numpy as np
 
-import neurodsp
-from neurodsp.filt import filter_signal
-from neurodsp.timefrequency import amp_by_time, phase_by_time, freq_by_time
-from .util import load_example_data
+from neurodsp.timefrequency import *
+from neurodsp.timefrequency import _hilbert_ignore_nan
 
 ###################################################################################################
 ###################################################################################################
+
+def test_phase_by_time():
+    pass
+
+def test_amp_by_time():
+    pass
+
+def test_freq_by_time():
+    pass
+
+def test_hilbert_ignore_nan():
+    """
+    Assure that time-resolved timefrequency functions do not return all NaN
+    if one of the elements in the input array is NaN.
+    Do this by replacing edge artifacts with NaN for a lowpass filter
+    """
+
+    # Generate a signal with NaNs
+    fs, n_points, n_nans = 100, 1000, 10
+    sig = np.random.randn(n_points)
+    sig[0:n_nans] = np.nan
+
+    # Check has correct number of nans (not all nan), without increase_n
+    hilb_sig = _hilbert_ignore_nan(sig)
+    assert sum(np.isnan(hilb_sig)) == n_nans
+
+    # Check has correct number of nans (not all nan), with increase_n
+    hilb_sig = _hilbert_ignore_nan(sig, True)
+    assert sum(np.isnan(hilb_sig)) == n_nans
 
 def test_timefreq_consistent():
     """
@@ -18,76 +43,33 @@ def test_timefreq_consistent():
     with computations in previous versions
     """
 
-    # Load data
-    data_idx = 1
-    sig = load_example_data(data_idx=data_idx)
-    fs = 1000
-    f_range = (13, 30)
+    # # Load data
+    # data_idx = 1
+    # sig = load_example_data(data_idx=data_idx)
+    # fs = 1000
+    # f_range = (13, 30)
 
-    # Load ground truth phase time series
-    pha_true = np.load(os.path.dirname(neurodsp.__file__) +
-                       '/tests/data/sample_data_' + str(data_idx) + '_pha.npy')
-    # Load ground truth amplitude time series
-    amp_true = np.load(os.path.dirname(neurodsp.__file__) +
-                       '/tests/data/sample_data_' + str(data_idx) + '_amp.npy')
-    # Load ground truth frequency time series
-    i_f_true = np.load(os.path.dirname(neurodsp.__file__) +
-                       '/tests/data/sample_data_' + str(data_idx) + '_i_f.npy')
+    # # Load ground truth phase time series
+    # pha_true = np.load(os.path.dirname(neurodsp.__file__) +
+    #                    '/tests/data/sample_data_' + str(data_idx) + '_pha.npy')
+    # # Load ground truth amplitude time series
+    # amp_true = np.load(os.path.dirname(neurodsp.__file__) +
+    #                    '/tests/data/sample_data_' + str(data_idx) + '_amp.npy')
+    # # Load ground truth frequency time series
+    # i_f_true = np.load(os.path.dirname(neurodsp.__file__) +
+    #                    '/tests/data/sample_data_' + str(data_idx) + '_i_f.npy')
 
-    # Compute phase time series
-    pha = phase_by_time(sig, fs, f_range)
-    # Compute amplitude time series
[…]
