+++ neurodsp/sim/sim_transients.py
@@ -1,0 +1,7 @@
+"""Simulating time series, with transients."""
+
+
+###################################################################################################
+###################################################################################################
+
+pass
