+++ neurodsp/sim/sim_periodic.py
@@ -1,0 +1,356 @@
+"""Simulating time series, with periodic activity."""
+
+import warnings
+
+import numpy as np
+import pandas as pd
+from scipy import signal
+
+###################################################################################################
+###################################################################################################
+
+def sim_oscillator(n_seconds, fs, freq, rdsym=.5):
+    """Simulate an oscillation.
+
+    Parameters
+    ----------
+    n_seconds : float
+        Signal duration, in seconds.
+    fs : float
+        Signal sampling rate, in Hz.
+    freq : float
+        Oscillator frequency.
+    rdsym : float
+        Rise-decay symmetry of the oscillator, as fraction of the period in the rise time, where:
+        = 0.5 - symmetric (sine wave)
+        < 0.5 - shorter rise, longer decay
+        > 0.5 - longer rise, shorter decay
+
+    Returns
+    -------
+    osc : 1d array
+        Oscillating time series.
+    """
+
+    # Compute number of samples per cycle and number of cycles
+    n_samples_cycle = int(np.ceil(fs / freq))
+    n_samples = int(fs * n_seconds)
+    n_cycles = int(np.ceil(n_seconds * freq))
+
+    # Determine number of samples in rise and decay periods
+    rise_samples = int(np.round(n_samples_cycle * rdsym))
+    decay_samples = n_samples_cycle - rise_samples
+
+    # Make phase array for a single cycle, then repeat it
+    pha_one_cycle = np.hstack([np.linspace(
+        0, np.pi, decay_samples + 1), np.linspace(-np.pi, 0, rise_samples + 1)[1:-1]])
+    phase_t = np.tile(pha_one_cycle, n_cycles)
+    phase_t = phase_t[:n_samples]
+
+    # Transform phase into an oscillator
+    osc = np.cos(phase_t)
+
+    return osc
+
+
+def sim_bursty_oscillator(n_seconds, fs, freq, rdsym=.5, prob_enter_burst=.2,
+                          prob_leave_burst=.2, cycle_features=None,
+                          return_cycle_df=False, n_tries=5):
+    """Simulate a bursty oscillation.
+
+    Parameters
+    ----------
+    n_seconds : float
+        Simulation time, in seconds.
+    fs : float
+        Sampling rate of simulated signal, in Hz
+    freq : float
+        Oscillator frequency, in Hz.
+    rdsym : float
+        Rise-decay symmetry of the oscillator, as fraction of the period in the rise time:
+
+        - = 0.5: symmetric (sine wave)
+        - < 0.5: shorter rise, longer decay
+        - > 0.5: longer rise, shorter decay
+    prob_enter_burst : float
+        Probability of a cycle being oscillating given the last cycle is not oscillating.
+    prob_leave_burst : float
+        Probability of a cycle not being oscillating given the last cycle is oscillating.
+    cycle_features : dict
+        Specifies the mean and standard deviations (within and across bursts) of each cycle's
+        amplitude, period, and rise-decay symmetry. This can include a complete or incomplete
+        set (using defaults) of the following keys:
+
+        * amp_mean: mean cycle amplitude
+        * amp_std: standard deviation of cycle amplitude
+        * amp_burst_std: standard deviation of mean amplitude for each burst
+        * period_mean: mean period (computed from `freq`)
+        * period_std: standard deviation of period (samples)
+        * period_burst_std: standard deviation of mean period for each burst
+        * rdsym_mean: mean rise-decay symmetry
+        * rdsym_std: standard deviation of rdsym
+        * rdsym_burst_std: standard deviation of mean rdsym for each burst
+    return_cycle_df : bool
+        If True, return the dataframe that contains the simulation parameters for each cycle.
[…]

[…]
