### 2019-03-11 caa9d57
Update names & misc (non-breaking) clean-ups & refactors
