+++ neurodsp/sim/periodic.py
@@ -1,6 +1,7 @@
 """Simulating time series, with periodic activity."""
 
 import warnings
+from copy import deepcopy
 
 import numpy as np
 import pandas as pd
@@ -42,8 +43,8 @@ def sim_oscillator(n_seconds, fs, freq, rdsym=.5):
     decay_samples = n_samples_cycle - rise_samples
 
     # Make phase array for a single cycle, then repeat it
-    pha_one_cycle = np.hstack([np.linspace(
-        0, np.pi, decay_samples + 1), np.linspace(-np.pi, 0, rise_samples + 1)[1:-1]])
+    pha_one_cycle = np.hstack([np.linspace(0, np.pi, decay_samples + 1),
+                               np.linspace(-np.pi, 0, rise_samples + 1)[1:-1]])
     phase_t = np.tile(pha_one_cycle, n_cycles)
     phase_t = phase_t[:n_samples]
 
@@ -122,126 +123,19 @@ def sim_bursty_oscillator(n_seconds, fs, freq, rdsym=.5, prob_enter_burst=.2,
     n_cycles_overestimate = int(np.ceil(n_samples / mean_period_samples * 2))
 
     # Simulate if a series of cycles are oscillating or not oscillating
-    is_oscillating = [False]
-    n_cycles_current = 1
-    while n_cycles_current < n_cycles_overestimate:
-        rand_num = np.random.rand()
-        if is_oscillating[-1]:
-            is_oscillating.append(rand_num > prob_leave_burst)
-        else:
-            is_oscillating.append(rand_num < prob_enter_burst)
-        n_cycles_current += 1
+    is_oscillating = _make_is_osc(n_cycles_overestimate, prob_enter_burst, prob_leave_burst)
 
-    # Determine period, amp, and rdsym for each cycle
-    periods = []
-    amps = []
-    rdsyms = []
-    for is_osc in is_oscillating:
-        if is_osc is False:
-            period = cycle_features_use['period_mean'] + \
-                np.random.randn() * cycle_features_use['period_std']
-            periods.append(int(period))
-            amps.append(np.nan)
-            rdsyms.append(np.nan)
+    # Determine what the cycle properties shall be for each cycle
+    periods, amps, rdsyms = _determine_cycle_properties(is_oscillating, cycle_features_use, n_tries)
 
-            current_burst_period_mean = np.nan
-            current_burst_amp_mean = np.nan
-            current_burst_rdsym_mean = np.nan
-        else:
-            if np.isnan(current_burst_period_mean):
-                current_burst_period_mean = cycle_features_use['period_mean'] + \
-                    np.random.randn() * cycle_features_use['period_burst_std']
-                current_burst_amp_mean = cycle_features_use['amp_mean'] + \
-                    np.random.randn() * cycle_features_use['amp_burst_std']
-                current_burst_rdsym_mean = cycle_features_use['rdsym_mean'] + \
-                    np.random.randn() * cycle_features_use['rdsym_burst_std']
-
-            # simulate for n_tries, if any params are still negative, raise error
-            tries_left = n_tries
-            features_valid = False
-            while features_valid is False and tries_left >= 0:
-                tries_left = tries_left - 1
-                period = current_burst_period_mean + \
-                    np.random.randn() * cycle_features_use['period_std']
-                amp = current_burst_amp_mean + \
-                    np.random.randn() * cycle_features_use['amp_std']
-                rdsym = current_burst_rdsym_mean + \
-                    np.random.randn() * cycle_features_use['rdsym_std']
-
-                if period > 0 and amp > 0 and rdsym > 0 and rdsym < 1:
-                    features_valid = True
-
-            if features_valid is False:
-                # features are still invalid after n_tries, give up.
[…]
