+++ neurodsp/sim/__init__.py
@@ -1,3 +1,3 @@
-from .sim_periodic import *
-from .sim_aperiodic import *
-from .sim_combined import *
+from .periodic import *
+from .aperiodic import *
+from .combined import *
