### 2019-03-14 9b81e39
Continue - small refactors / reorgs / updates
