+++ neurodsp/sim/aperiodic.py
@@ -3,8 +3,8 @@
 import warnings
 
 import numpy as np
-from scipy.stats import zscore
 from scipy import signal
+from scipy.stats import zscore
 
 from neurodsp.filt import filter_signal_fir
 from neurodsp.spectral import rotate_powerlaw
@@ -14,15 +14,6 @@ from neurodsp.spectral import rotate_powerlaw
 
 def sim_poisson_pop(n_seconds, fs, n_neurons, firing_rate):
     """Simulates a poisson population.
-
-    It is essentially white noise, but satisfies the Poisson property, i.e. mean(X) = var(X).
-
-    The lambda parameter of the Poisson process (total rate) is determined as
-    firing rate * number of neurons, i.e. summation of poisson processes is still
-    a poisson processes.
-
-    Note that the Gaussian approximation for a sum of Poisson processes is only
-    a good approximation for large lambdas.
 
     Parameters
     ----------
@@ -39,17 +30,26 @@ def sim_poisson_pop(n_seconds, fs, n_neurons, firing_rate):
     -------
     sig : 1d array
         Simulated population activity.
+
+    Notes
+    -----
+    It is essentially white noise, but satisfies the Poisson property, i.e. mean(X) = var(X).
+
+    The lambda parameter of the Poisson process (total rate) is determined as
+    firing rate * number of neurons, i.e. summation of poisson processes is still
+    a poisson processes.
+
+    Note that the Gaussian approximation for a sum of Poisson processes is only
+    a good approximation for large lambdas.
     """
 
-    n_samples = int(n_seconds * fs)
-
-    # poisson population rate signal scales with # of neurons and individual rate
+    # Poisson population rate signal scales with # of neurons and individual rate
     lam = n_neurons * firing_rate
 
-    # variance is equal to the mean
-    sig = np.random.normal(loc=lam, scale=lam**0.5, size=n_samples)
+    # Variance is equal to the mean
+    sig = np.random.normal(loc=lam, scale=lam**0.5, size=int(n_seconds * fs))
 
-    # enforce that sig is non-negative in cases of low firing rate
+    # Enforce that sig is non-negative in cases of low firing rate
     sig[np.where(sig < 0.)] = 0.
 
     return sig
@@ -57,11 +57,6 @@ def sim_poisson_pop(n_seconds, fs, n_neurons, firing_rate):
 
 def make_synaptic_kernel(t_ker, fs, tau_r, tau_d):
     """Creates synaptic kernels that with specified time constants.
-
-    3 types of kernels are available, based on combinations of time constants:
-    - tau_r == tau_d  : alpha (function) synapse
-    - tau_r = 0       : instantaneous rise, (single) exponential decay
-    - tau_r!=tau_d!=0 : double-exponential (rise and decay)
 
     Parameters
     ----------
@@ -78,11 +73,18 @@ def make_synaptic_kernel(t_ker, fs, tau_r, tau_d):
     -------
     kernel : 1d array
         Computed synaptic kernel with length equal to t
+
+    Notes
+    -----
+    3 types of kernels are available, based on combinations of time constants:
+    - tau_r == tau_d  : alpha (function) synapse
+    - tau_r = 0       : instantaneous rise, (single) exponential decay
+    - tau_r!=tau_d!=0 : double-exponential (rise and decay)
     """
 
-    ### NOTE: sometimes t_ker is not exact, resulting in a slightly longer or
-    ###     shorter times vector, which will affect final signal length
[…]
