+++ neurodsp/sim/periodic.py
@@ -4,6 +4,7 @@ import warnings
 from copy import deepcopy
 
 import numpy as np
+from numpy.random import rand, randn, randint
 import pandas as pd
 from scipy import signal
 
@@ -196,7 +197,7 @@ def sim_jittered_oscillator(n_seconds, fs, freq, jitter=0, cycle=('gaussian', 0.
     # Add jitter to "spike" indices
     if jitter != 0:
         spk_indices = spk_indices + \
-            np.random.randint(low=-int(fs * jitter), high=int(fs * jitter), size=len(spk_indices))
+            randint(low=-int(fs * jitter), high=int(fs * jitter), size=len(spk_indices))
 
     spks[spk_indices] = 1
     sig = np.convolve(spks, osc_cycle, 'valid')
@@ -256,7 +257,7 @@ def _make_is_osc(n_cycles, prob_enter_burst, prob_leave_burst):
 
     for ii in range(1, n_cycles):
 
-        rand_num = np.random.rand()
+        rand_num = rand()
 
         if is_oscillating[ii-1]:
             is_oscillating[ii] = rand_num > prob_leave_burst
@@ -275,43 +276,39 @@ def _determine_cycle_properties(is_oscillating, cycle_features, n_tries):
     for ind, is_osc in enumerate(is_oscillating):
 
         if is_osc is False:
-            period = cycle_features['period_mean'] + \
-                np.random.randn() * cycle_features['period_std']
+            period = cycle_features['period_mean'] + randn() * cycle_features['period_std']
 
             periods[ind] = int(period)
             amps[ind] = np.nan
             rdsyms[ind] = np.nan
 
-            current_burst_period_mean = np.nan
-            current_burst_amp_mean = np.nan
-            current_burst_rdsym_mean = np.nan
+            cur_burst = {'period_mean' : np.nan, 'amp_mean' : np.nan, 'rdsym_mean' : np.nan}
 
         else:
-            if np.isnan(current_burst_period_mean):
-                current_burst_period_mean = cycle_features['period_mean'] + \
-                    np.random.randn() * cycle_features['period_burst_std']
-                current_burst_amp_mean = cycle_features['amp_mean'] + \
-                    np.random.randn() * cycle_features['amp_burst_std']
-                current_burst_rdsym_mean = cycle_features['rdsym_mean'] + \
-                    np.random.randn() * cycle_features['rdsym_burst_std']
+
+            if np.isnan(cur_burst['period_mean']):
+
+                cur_burst['period_mean'] = cycle_features['period_mean'] + randn() * cycle_features['period_burst_std']
+                cur_burst['amp_mean'] = cycle_features['amp_mean'] + randn() * cycle_features['amp_burst_std']
+                cur_burst['rdsym_mean'] = cycle_features['rdsym_mean'] + randn() * cycle_features['rdsym_burst_std']
 
             # simulate for n_tries, if any params are still negative, raise error
             tries_left = n_tries
             features_valid = False
 
             while features_valid is False and tries_left >= 0:
+
                 tries_left = tries_left - 1
-                period = current_burst_period_mean + \
-                    np.random.randn() * cycle_features['period_std']
-                amp = current_burst_amp_mean + \
-                    np.random.randn() * cycle_features['amp_std']
-                rdsym = current_burst_rdsym_mean + \
-                    np.random.randn() * cycle_features['rdsym_std']
+
+                period = cur_burst['period_mean'] + randn() * cycle_features['period_std']
+                amp = cur_burst['amp_mean'] + randn() * cycle_features['amp_std']
+                rdsym = cur_burst['rdsym_mean'] + randn() * cycle_features['rdsym_std']
 
[…]
