### 2019-03-16 370f307
Move some stuff to transients, and small refactors
