### 2019-03-16 ced7fa0
Finish updates of tests for 1.1
