+++ neurodsp/tests/util.py
@@ -24,11 +24,13 @@ def load_example_data(data_idx=1, filtered=False):
     """Load an example voltage time series collected experimentally"""
 
     # Load time series
-    sig = np.load(os.path.dirname(neurodsp.__file__) + '/tests/data/sample_data_'+str(data_idx)+'.npy')
+    sig = np.load(os.path.dirname(neurodsp.__file__) + \
+                  '/tests/data/sample_data_'+str(data_idx)+'.npy')
 
     # Load ground-truth filtered data
     if filtered:
-        sig_filt_true = np.load(os.path.dirname(neurodsp.__file__) + '/tests/data/sample_data_'+str(data_idx)+'_filt.npy')
+        sig_filt_true = np.load(os.path.dirname(neurodsp.__file__) + \
+            '/tests/data/sample_data_'+str(data_idx)+'_filt.npy')
         return sig, sig_filt_true
     else:
         return sig
