+++ neurodsp/tests/test_timefreq.py
@@ -1,25 +1,19 @@
 """Test functions for the time-frequency analysis module."""
 
+import os
+
 import numpy as np
 
 from neurodsp.timefrequency import *
 from neurodsp.timefrequency import _hilbert_ignore_nan
 
+from .util import load_example_data
+
 ###################################################################################################
 ###################################################################################################
-
-def test_phase_by_time():
-    pass
-
-def test_amp_by_time():
-    pass
-
-def test_freq_by_time():
-    pass
 
 def test_hilbert_ignore_nan():
-    """
-    Assure that time-resolved timefrequency functions do not return all NaN
+    """Check that time-resolved timefrequency functions do not return all NaN
     if one of the elements in the input array is NaN.
     Do this by replacing edge artifacts with NaN for a lowpass filter
     """
@@ -38,38 +32,37 @@ def test_hilbert_ignore_nan():
     assert sum(np.isnan(hilb_sig)) == n_nans
 
 def test_timefreq_consistent():
-    """
-    Confirm consistency in estimation of instantaneous phase, amp, and frequency
+    """Confirm consistency in estimation of instantaneous phase, amp, and frequency
     with computations in previous versions
     """
 
-    # # Load data
-    # data_idx = 1
-    # sig = load_example_data(data_idx=data_idx)
-    # fs = 1000
-    # f_range = (13, 30)
+    # Load data
+    data_idx = 1
+    sig = load_example_data(data_idx=data_idx)
+    fs = 1000
+    f_range = (13, 30)
 
-    # # Load ground truth phase time series
-    # pha_true = np.load(os.path.dirname(neurodsp.__file__) +
-    #                    '/tests/data/sample_data_' + str(data_idx) + '_pha.npy')
-    # # Load ground truth amplitude time series
-    # amp_true = np.load(os.path.dirname(neurodsp.__file__) +
-    #                    '/tests/data/sample_data_' + str(data_idx) + '_amp.npy')
-    # # Load ground truth frequency time series
-    # i_f_true = np.load(os.path.dirname(neurodsp.__file__) +
-    #                    '/tests/data/sample_data_' + str(data_idx) + '_i_f.npy')
+    # Load ground truth phase time series
+    pha_true = np.load(os.path.dirname(neurodsp.__file__) +
+                       '/tests/data/sample_data_' + str(data_idx) + '_pha.npy')
+    # Load ground truth amplitude time series
+    amp_true = np.load(os.path.dirname(neurodsp.__file__) +
+                       '/tests/data/sample_data_' + str(data_idx) + '_amp.npy')
+    # Load ground truth frequency time series
+    i_f_true = np.load(os.path.dirname(neurodsp.__file__) +
+                       '/tests/data/sample_data_' + str(data_idx) + '_i_f.npy')
 
-    # # Compute phase time series
-    # pha = phase_by_time(sig, fs, f_range)
-    # # Compute amplitude time series
-    # amp = amp_by_time(sig, fs, f_range)
-    # # Compute frequency time series
-    # i_f = freq_by_time(sig, fs, f_range)
+    # Compute phase time series
+    pha = phase_by_time(sig, fs, f_range)
+    # Compute amplitude time series
+    amp = amp_by_time(sig, fs, f_range)
+    # Compute frequency time series
+    i_f = freq_by_time(sig, fs, f_range)
 
-    # # Compute difference between current and past signals
-    # assert np.allclose(
[…]
