### 2019-03-18 cfeb650
Update filt API
