+++ neurodsp/filt.py
@@ -5,15 +5,16 @@ import warnings
 import numpy as np
 from scipy import signal
 
-from neurodsp.plts.filt import plot_frequency_response
+from neurodsp.utils import remove_nans, restore_nans
+from neurodsp.plts.filt import plot_filter_properties, plot_frequency_response
 
 ###################################################################################################
 ###################################################################################################
 
 def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
-                  iir=False, butterworth_order=None,
-                  plot_freq_response=False, return_kernel=False,
-                  verbose=False, compute_transition_band=True, remove_edge_artifacts=True):
+                  filt_type='fir', remove_edge_artifacts=True,
+                  butterworth_order=None, print_transitions=True,
+                  plot_properties=False, return_kernel=False):
     """Apply a bandpass, bandstop, highpass, or lowpass filter to a neural signal.
 
     Parameters
@@ -38,24 +39,22 @@ def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
         Length of filter, in number of cycles, defined at the 'f_lo' frequency.
         This parameter is overwritten by `n_seconds`, if provided.
     n_seconds : float, optional
-        Length of filter, in seconds.
-        This parameter overwrites `n_cycles`.
-    iir : bool, optional
-        If True, use an infinite-impulse response (IIR) filter.
+        Length of filter, in seconds. This parameter overwrites `n_cycles`.
+    filt_type : {'fir', 'iir'}, optional
+        Whether to use an FIR or IIR filter.
         The only IIR filter offered is a butterworth filter.
+    remove_edge_artifacts : bool, optional, default: True
+        If True, replace samples within half the kernel length to be np.nan.
+        Only used for FIR filters.
     butterworth_order : int, optional
         Order of the butterworth filter, if using an IIR filter.
         See input 'N' in scipy.signal.butter.
-    plot_freq_response : bool, optional, default: False
-        If True, plot the frequency response of the filter
+    print_transitions : bool, optional, default: True
+        If True, computes the transition bandwidth(s), and prints this information.
+    plot_properties : bool, optional, default: False
+        If True, plot the properties of the filter, including frequency response and/or kernel.
     return_kernel : bool, optional, default: False
         If True, return the complex filter kernel.
-    verbose : bool, optional, default: False
-        If True, print optional information.
-    compute_transition_band : bool, optional, default: True
-        If True, computes the transition bandwidth, and prints this information.
-    remove_edge_artifacts : bool, optional, default: True
-        If True, replace samples within half the kernel length to be np.nan.
 
     Returns
     -------
@@ -65,17 +64,22 @@ def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
         Filter kernel. Only returned if `return_kernel` is True.
     """
 
-    if iir:
+    if filt_type == 'fir':
+        return filter_signal_fir(sig, fs, pass_type, fc, n_cycles, n_seconds,
+                                 remove_edge_artifacts, print_transitions,
+                                 plot_properties, return_kernel)
+    elif filt_type == 'iir':
         _iir_checks(n_seconds, butterworth_order, remove_edge_artifacts)
-        return filter_signal_iir(sig, fs, pass_type, fc, butterworth_order, plot_freq_response,
-                                 return_kernel, compute_transition_band)
[…]
