### 2019-03-18 5196910
Update / refactor fitler plotting funcs
