### 2019-03-18 8afa55e
Add separate utils file
