### 2019-03-18 9aef4fe
Lint & last clean ups
