### 2019-03-18 6bc2f99
Finish org, & test update, of utils & remove filt edges
