+++ neurodsp/tests/test_plts_spectral.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 
-from neurodsp.spectral import spectral_hist
+from neurodsp.spectral import compute_spectral_hist
 from neurodsp.plts.spectral import plot_spectral_hist
 from .util import plot_test
 
@@ -17,7 +17,7 @@ def test_plot_spectral_hist():
     fs = 1000
 
     # Compute spectral histogram
-    freqs, power_bins, spect_hist = spectral_hist(sig, fs)
+    freqs, power_bins, spect_hist = compute_spectral_hist(sig, fs)
 
     # Test plotting function runs without error
     plot_spectral_hist(freqs, power_bins, spect_hist)
