### 2019-03-18 5a2517c
Udpate sim_filtered_noise to use general ndsp filter - currently not really tested
