+++ neurodsp/tests/test_sim_aperiodic.py
@@ -18,13 +18,19 @@ filter_order = 1501
 
 def test_sim_filtered_noise():
 
-    np.random.seed(0)
-    noise = sim_filtered_noise(n_seconds, fs, exponent, f_range_filter, filter_order)
-    # np.save(os.path.dirname(neurodsp.__file__) + '/tests/data/noise_filt.npy', noise)
-    noise_true = np.load(os.path.dirname(
-        neurodsp.__file__) + '/tests/data/noise_filt.npy')
+    noise = sim_filtered_noise(n_seconds, fs, exponent, f_range_filter)
+    assert np.all(noise)
 
-    assert np.allclose(np.sum(np.abs(noise - noise_true)), 0, atol=10 ** -5)
+    # Note: old consistency test turned off after moving to using filter_signal which
+    #   is slightly different than the filter defined directly in `sim_filtered_noise`
+
+    #np.random.seed(0)
+    #noise = sim_filtered_noise(n_seconds, fs, exponent, f_range_filter, filter_order)
+    # np.save(os.path.dirname(neurodsp.__file__) + '/tests/data/noise_filt.npy', noise)
+    #noise_true = np.load(os.path.dirname(
+    #    neurodsp.__file__) + '/tests/data/noise_filt.npy')
+
+    #assert np.allclose(np.sum(np.abs(noise - noise_true)), 0, atol=10 ** -5)
 
 
 def test_sim_poisson_pop():
