### 2019-03-18 a1aa6d7
Consolidate on **filter_kwargs, and small refactors / lints
