### 2019-03-18 6432d83
Lints & small refactors
