### 2019-03-18 ea483b1
Update to new rotate_spectrum
