### 2019-03-19 fcadcf4
Small refactors
