### 2019-03-19 3d045ac
WIP updates, some refactors & internal renamings, and make utils file
