### 2019-03-21 20950d7
Consolidate on 'f_range'
