+++ neurodsp/filt.py
@@ -11,7 +11,7 @@ from neurodsp.plts.filt import plot_filter_properties, plot_frequency_response
 ###################################################################################################
 ###################################################################################################
 
-def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
+def filter_signal(sig, fs, pass_type, f_range, n_cycles=3, n_seconds=None,
                   filt_type='fir', remove_edges=True, butterworth_order=None,
                   print_transitions=True, plot_properties=False, return_kernel=False):
     """Apply a bandpass, bandstop, highpass, or lowpass filter to a neural signal.
@@ -29,7 +29,7 @@ def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
         * 'bandstop': apply a bandstop (notch) filter
         * 'lowpass': apply a lowpass filter
         * 'highpass' : apply a highpass filter
-    fc : tuple of (float, float) or float
+    f_range : tuple of (float, float) or float
         Cutoff frequency(ies) used for filter, specified as f_lo & f_hi.
         For 'bandpass' & 'bandstop', must be a tuple.
         For 'lowpass' or 'highpass', can be a float that specifies pass frequency, or can be
@@ -64,19 +64,19 @@ def filter_signal(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None,
     """
 
     if filt_type == 'fir':
-        return filter_signal_fir(sig, fs, pass_type, fc, n_cycles, n_seconds,
+        return filter_signal_fir(sig, fs, pass_type, f_range, n_cycles, n_seconds,
                                  remove_edges, print_transitions,
                                  plot_properties, return_kernel)
     elif filt_type == 'iir':
         _iir_checks(n_seconds, butterworth_order, remove_edges)
-        return filter_signal_iir(sig, fs, pass_type, fc, butterworth_order,
+        return filter_signal_iir(sig, fs, pass_type, f_range, butterworth_order,
                                  print_transitions, plot_properties,
                                  return_kernel)
     else:
         raise ValueError('Filter type not understood.')
 
 
-def filter_signal_fir(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None, remove_edges=True,
+def filter_signal_fir(sig, fs, pass_type, f_range, n_cycles=3, n_seconds=None, remove_edges=True,
                       print_transitions=True, plot_properties=False, return_kernel=False):
     """Apply an FIR filter to a signal.
 
@@ -93,7 +93,7 @@ def filter_signal_fir(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None, remove
         * 'bandstop': apply a bandstop (notch) filter
         * 'lowpass': apply a lowpass filter
         * 'highpass' : apply a highpass filter
-    fc : tuple of (float, float) or float
+    f_range : tuple of (float, float) or float
         Cutoff frequency(ies) used for filter, specified as f_lo & f_hi.
         For 'bandpass' & 'bandstop', must be a tuple.
         For 'lowpass' or 'highpass', can be a float that specifies pass frequency, or can be
@@ -121,11 +121,11 @@ def filter_signal_fir(sig, fs, pass_type, fc, n_cycles=3, n_seconds=None, remove
     """
 
     # Design filter
-    kernel = design_fir_filter(len(sig), fs, pass_type, fc, n_cycles, n_seconds)
[…]
