### 2019-03-21 8466dc6
Clean up filt API
