+++ neurodsp/filt.py
@@ -13,7 +13,7 @@ from neurodsp.plts.filt import plot_filter_properties, plot_frequency_response
 
 def filter_signal(sig, fs, pass_type, f_range, n_cycles=3, n_seconds=None,
                   filt_type='fir', remove_edges=True, butterworth_order=None,
-                  print_transitions=True, plot_properties=False, return_kernel=False):
+                  print_transitions=True, plot_properties=False, return_filter=False):
     """Apply a bandpass, bandstop, highpass, or lowpass filter to a neural signal.
 
     Parameters
@@ -52,32 +52,32 @@ def filter_signal(sig, fs, pass_type, f_range, n_cycles=3, n_seconds=None,
         If True, computes the transition bandwidth(s), and prints this information.
     plot_properties : bool, optional, default: False
         If True, plot the properties of the filter, including frequency response and/or kernel.
-    return_kernel : bool, optional, default: False
-        If True, return the complex filter kernel.
+    return_filter : bool, optional, default: False
+        If True, return the filter coefficients.
 
     Returns
     -------
     sig_filt : 1d array
         Filtered time series.
-    kernel : length-2 tuple of arrays
-        Filter kernel. Only returned if `return_kernel` is True.
+    kernel : 1d array or tuple of (1d array, 1d array)
+        Filter coefficients. Only returned if `return_filter` is True.
     """
 
     if filt_type == 'fir':
         return filter_signal_fir(sig, fs, pass_type, f_range, n_cycles, n_seconds,
                                  remove_edges, print_transitions,
-                                 plot_properties, return_kernel)
+                                 plot_properties, return_filter)
     elif filt_type == 'iir':
         _iir_checks(n_seconds, butterworth_order, remove_edges)
         return filter_signal_iir(sig, fs, pass_type, f_range, butterworth_order,
                                  print_transitions, plot_properties,
-                                 return_kernel)
+                                 return_filter)
     else:
         raise ValueError('Filter type not understood.')
 
 
 def filter_signal_fir(sig, fs, pass_type, f_range, n_cycles=3, n_seconds=None, remove_edges=True,
-                      print_transitions=True, plot_properties=False, return_kernel=False):
+                      print_transitions=True, plot_properties=False, return_filter=False):
     """Apply an FIR filter to a signal.
 
     Parameters
@@ -109,50 +109,50 @@ def filter_signal_fir(sig, fs, pass_type, f_range, n_cycles=3, n_seconds=None, r
         If True, computes the transition bandwidth, and prints this information.
     plot_properties : bool, optional, default: False
         If True, plot the properties of the filter, including frequency response and/or kernel.
-    return_kernel : bool, optional, default: False
-        If True, return the complex filter kernel.
+    return_filter : bool, optional, default: False
+        If True, return the filter coefficients of the FIR filter.
 
     Returns
     -------
     sig_filt : 1d array
         Filtered time series.
-    kernel : length-2 tuple of arrays
-        Filter kernel. Only returned if `return_kernel` is True.
+    filter_coefs : 1d array
+        Filter coefficients of the FIR filter. Only returned if `return_filter` is True.
     """
 
     # Design filter
-    kernel = design_fir_filter(len(sig), fs, pass_type, f_range, n_cycles, n_seconds)
[…]
