### 2019-03-21 d71b12a
Extend variance normalization to deal with non-continuous signals
