### 2019-03-22 3ae944e
Add info file, to manage getting sim funcs
