### 2019-03-22 441bab4
Extend set of util funcs
