+++ neurodsp/sim/combined.py
@@ -1,17 +1,18 @@
 """Simulating time series, with combinations of periodic, aperiodic and transient components."""
 
-import numpy as np
+from itertools import repeat
 
-from neurodsp.sim.periodic import sim_oscillation, sim_bursty_oscillation
-from neurodsp.sim.aperiodic import get_aperiodic_sim
-from neurodsp.sim.utils import normalized_sum
+from neurodsp.sim.info import get_sim_func
+from neurodsp.sim.utils import normalized_sum, proportional_sum
 
 ###################################################################################################
 ###################################################################################################
 
-def sim_noisy_oscillation(n_seconds, fs, freq, noise_generator='synaptic',
-                          noise_args={}, rdsym=.5, ratio_osc_var=1):
-    """Simulate an oscillation embedded in background 1/f noise.
+def sim_combined(n_seconds, fs, freq,
+                 sim_periodic='oscillation', sim_aperiodic='powerlaw',
+                 periodic_args={}, aperiodc_args={},
+                 ratio=1., select_nonzero=True):
+    """Simulate a combined signal, with periodic and aperiodic components.
 
     Parameters
     ----------
@@ -19,157 +20,64 @@ def sim_noisy_oscillation(n_seconds, fs, freq, noise_generator='synaptic',
         Signal duration, in seconds.
     fs : float
         Signal sampling rate, in Hz.
-    freq : float
-        Oscillation frequency.
-    noise_generator: str, optional, default='synaptic'
-        Noise model, can be one of the simulators in neurodsp.sim specificed as a string.
-
-        Possible models (see respective documentation):
-
-        - 'synaptic' or 'lorentzian': sim.sim_synaptic_noise()
-          Defaults: n_neurons=1000, firing_rate=2, tau_r=0, tau_d=0.01
-        - 'filtered_powerlaw': sim.sim_filtered_noise()
-          Defaults: exponent=-2., f_range=(0.5, None), filter_order=None
-        - 'powerlaw': sim.sim_variable_powerlaw()
-          Defaults: exponent=-2.0
-        - 'ou_process': sim.sim_ou_process()
-          Defaults: theta=1., mu=0., sigma=5.
-    noise_args: dict('argname':argval, ...)
-        Function arguments for the neurodsp.sim noise generaters. See API for arg names.
-        All args are optional, defaults for each noise generator are listed above.
-    rdsym : float
-        Rise-decay symmetry of the oscillation, as fraction of the period in the rise time:
-
-        - = 0.5 - symmetric (i.e., sine wave, default)
-        - < 0.5 - shorter rise, longer decay
-        - > 0.5 - longer rise, shorter decay
-    ratio_osc_var : float, optional, default=1
-        Ratio of oscillation variance to noise variance.
-        If >1 - oscillation is stronger, if <1 - noise is stronger.
+    sim_periodic : str or executable
+        Periodic simulation function to use.
+    sim_aperiodic : str or executable
+        Aperiodic simulation function to use.
+    periodic_args : dictionary
+        Parameters for the periodic sim function.
+    aperiodic_args : dictionary
+        Parameters for the aperiodic sim function.
+    ratio : float, optional, default: 1.
+        Ratio of signal variance. If >1, periodic is stronger, if <1 - aperiodic is stronger.
+    select_nonzero : bool, optional, default: True
+        Whether to select only non-zero samples when normalizing variance of the periodic signal.
 
     Returns
     -------
     sig : 1d array
-        Combined signal with an oscillation and an aperiodic component.
+        Simulated combined signal.
     """
 
-    # Generate & demean aperiodic signal
[…]
