### 2019-03-23 5cfb870
Add decorator to apply normalization
