### 2019-03-23 c766ed7
Switch up combined approach
