+++ neurodsp/sim/utils.py
@@ -5,13 +5,15 @@ import numpy as np
 ###################################################################################################
 ###################################################################################################
 
-def demean(array):
-    """Demean an array.
+def demean(array, mean=0.):
+    """Demean an array, updating to specified mean.
 
     Parameters
     ----------
     array : 1d array
         Data to demean.
+    mean : float, optional, default: 0
+        New mean for data to have.
 
     Returns
     -------
@@ -19,11 +21,11 @@ def demean(array):
         Demeaned data.
     """
 
-    return array - array.mean()
+    return array - array.mean() + mean
 
 
 def normalize_variance(array, variance=1.):
-    """Normalize the variance of an array.
+    """Normalize the variance of an array, updating to specified variance.
 
     Parameters
     ----------
@@ -60,58 +62,58 @@ def proportional_sum(signals, proportions):
     return sum([normalize_variance(sig, prop) for prop, sig in zip(proportions, signals)])
 
 
-def normalized_sum(sig1, sig2, ratio, select_nonzero=True):
-    """Combine two signals, after transforming to have the desired variance ratio.
+# def normalized_sum(sig1, sig2, ratio, select_nonzero=True):
+#     """Combine two signals, after transforming to have the desired variance ratio.
 
-    Parameters
-    ----------
-    sig1, sig2 : 1d array
-        Arrays of data to normalize variance with respect to each other.
-    ratio : float
-        Desired ratio of sig1 variance to sig2 variance.
-        If > 1 - sig1 is stronger, if < 1 - sig2 is stronger.
-    select_nonzero : boolean, optional, default=True
-        Whether to calculate the variance of sig1 across only non-zero data points.
+#     Parameters
+#     ----------
+#     sig1, sig2 : 1d array
+#         Arrays of data to normalize variance with respect to each other.
+#     ratio : float
+#         Desired ratio of sig1 variance to sig2 variance.
+#         If > 1 - sig1 is stronger, if < 1 - sig2 is stronger.
+#     select_nonzero : boolean, optional, default=True
+#         Whether to calculate the variance of sig1 across only non-zero data points.
 
-    Returns
-    -------
-    1d array
-        New array with the combined input signals.
-    """
+#     Returns
+#     -------
+#     1d array
+#         New array with the combined input signals.
+#     """
 
-    return sum(variance_ratio(sig1, sig2, ratio, select_nonzero))
+#     return sum(variance_ratio(sig1, sig2, ratio, select_nonzero))
 
 
-def variance_ratio(sig1, sig2, ratio, select_nonzero=True):
-    """Normalize the variance across two signals to reflect a specified ratio.
+# def variance_ratio(sig1, sig2, ratio, select_nonzero=True):
+#     """Normalize the variance across two signals to reflect a specified ratio.
 
-    Parameters
-    ----------
-    sig1, sig2 : 1d array
-        Vectors of data to normalize variance with respect to each other.
-    ratio : float
-        Desired ratio of sig1 variance to sig2 variance.
-        If > 1 - sig1 is stronger, if < 1 - sig2 is stronger.
-    select_nonzero : boolean, optional, default=True
-        Whether to calculate the variance of sig1 across only non-zero data points.
+#     Parameters
+#     ----------
+#     sig1, sig2 : 1d array
+#         Vectors of data to normalize variance with respect to each other.
+#     ratio : float
+#         Desired ratio of sig1 variance to sig2 variance.
+#         If > 1 - sig1 is stronger, if < 1 - sig2 is stronger.
+#     select_nonzero : boolean, optional, default=True
[…]
