+++ neurodsp/tests/test_sim_utils.py
@@ -1,0 +1,15 @@
+"""Tests for sim utility functions."""
+
+from neurodsp.sim.utils import *
+
+###################################################################################################
+###################################################################################################
+
+def test_demean():
+    pass
+
+def test_normalize_variance():
+    pass
+
+def test_proportional_sum():
+    pass
