+++ neurodsp/tests/test_sim_decorators.py
@@ -1,0 +1,9 @@
+"""Tests for sim decorator functions."""
+
+from neurodsp.sim.decorators import *
+
+###################################################################################################
+###################################################################################################
+
+def test_normalize():
+    pass
