+++ neurodsp/tests/test_sim_info.py
@@ -1,0 +1,15 @@
+"""Tests for sim info functions."""
+
+from neurodsp.sim.info import *
+
+###################################################################################################
+###################################################################################################
+
+def test_get_sim_funcs():
+    pass
+
+def test_get_sim_names():
+    pass
+
+def test_get_sim_func():
+    pass
