### 2019-03-23 bde69a2
Clear out the leftovers of consistency tests
