+++ neurodsp/tests/test_timefreq.py
@@ -1,64 +1,24 @@
 """Test functions for the time-frequency analysis module."""
 
-import os
-
 import numpy as np
 
-import neurodsp
 from neurodsp.timefrequency import *
-
-from .util import load_example_data
 
 ###################################################################################################
 ###################################################################################################
 
 def test_phase_by_time():
-
-    data_idx = 1
-    sig = load_example_data(data_idx=data_idx)
-    fs = 1000
-    f_range = (13, 30)
-
-    pha_true = np.load(os.path.dirname(neurodsp.__file__) + '/tests/data/sample_data_' + str(data_idx) + '_pha.npy')
-
-    pha = phase_by_time(sig, fs, f_range)
-
-    assert np.allclose(np.sum(np.abs(pha[~np.isnan(pha)] - pha_true[~np.isnan(pha)])), 0, atol=10 ** -5)
-
+    pass
 
 def test_amp_by_time():
-
-    data_idx = 1
-    sig = load_example_data(data_idx=data_idx)
-    fs = 1000
-    f_range = (13, 30)
-
-    amp_true = np.load(os.path.dirname(neurodsp.__file__) + '/tests/data/sample_data_' + str(data_idx) + '_amp.npy')
-
-    amp = amp_by_time(sig, fs, f_range)
-
-    assert np.allclose(np.sum(np.abs(amp[~np.isnan(amp)] - amp_true[~np.isnan(amp)])), 0, atol=10 ** -5)
+    pass
 
 def test_freq_by_time():
+    pass
 
-    data_idx = 1
-    sig = load_example_data(data_idx=data_idx)
-    fs = 1000
-    f_range = (13, 30)
-
-    i_f_true = np.load(os.path.dirname(neurodsp.__file__) +
-                       '/tests/data/sample_data_' + str(data_idx) + '_i_f.npy')
-
-    i_f = freq_by_time(sig, fs, f_range)
-
-    assert np.allclose(np.sum(np.abs(i_f[~np.isnan(i_f)] - i_f_true[~np.isnan(i_f_true)])), 0, atol=10 ** -5)
-
+## PRIVATE FUNCTIONS
 
 def test_robust_hilbert():
-    """Check that time-resolved timefrequency functions do not return all NaN
-    if one of the elements in the input array is NaN.
-    Do this by replacing edge artifacts with NaN for a lowpass filter
-    """
 
     # Generate a signal with NaNs
     fs, n_points, n_nans = 100, 1000, 10
