### 2019-03-23 f7fca8a
Update test configs & utils
