+++ neurodsp/tests/conftest.py
@@ -1,0 +1,10 @@
+"""Configuration file for pytest for NDSP."""
+
+from fooof.tests.utils import get_random_signal
+
+###################################################################################################
+###################################################################################################
+
+@pytest.fixture(scope='session')
+def tsig():
+    yield get_random_signal()
