### 2019-03-23 3e8f07f
Fill in bunch of smoke tests
