+++ neurodsp/tests/test_plts_spectral.py
@@ -1,9 +1,7 @@
 """Test spectral plots."""
 
-import numpy as np
-
 from neurodsp.spectral import compute_spectral_hist
-from neurodsp.tests.util import plot_test
+from neurodsp.tests.utils import plot_test
 
 from neurodsp.plts.spectral import *
 
@@ -11,11 +9,7 @@ from neurodsp.plts.spectral import *
 ###################################################################################################
 
 @plot_test
-def test_plot_spectral_hist():
+def test_plot_spectral_hist(tsig):
 
-    # Generate random signal & compute spectral histogram
-    sig = np.random.randn(1000)
-    freqs, power_bins, spect_hist = compute_spectral_hist(sig, fs=1000)
-
-    # Test plotting function runs without error
+    freqs, power_bins, spect_hist = compute_spectral_hist(tsig, fs=1000)
     plot_spectral_hist(freqs, power_bins, spect_hist)
