+++ neurodsp/tests/test_sim_periodic.py
@@ -1,6 +1,7 @@
 """Test periodic simulation functions."""
 
 from neurodsp.sim.periodic import *
+from neurodsp.sim.periodic import _make_is_osc, _determine_cycle_properties, _sim_cycles
 
 ###################################################################################################
 ###################################################################################################
@@ -23,3 +24,14 @@ def test_sim_jittered_oscillation():
 
     sig = sim_jittered_oscillation(N_SECONDS, FS, FREQ)
     assert True
+
+## PRIVATE FUNCTIONS
+
+def test_make_is_osc():
+    pass
+
+def test_determine_cycle_properties():
+    pass
+
+def test_sim_cycles():
+    pass
