### 2019-03-24 e20352c
Fix inputs for plots
