### 2019-03-24 7c51e1c
Update tutorials for new API
