+++ tutorials/plot_1-Filtering.py
@@ -51,8 +51,8 @@ fs = 1000
 ###################################################################################################
 
 # Filter the data, across a frequency band of interest
-fc = (4, 8)
-sig_filt = filt.filter_signal(sig, fs, 'bandpass', fc)
+f_range = (4, 8)
+sig_filt = filt.filter_signal(sig, fs, 'bandpass', f_range)
 
 ###################################################################################################
 
@@ -92,8 +92,8 @@ sig = np.random.randn(len(times)) + 5 * np.sin(times*2*np.pi*3) + 4 * np.sin(tim
 ###################################################################################################
 
 # Filter the data
-fc = 2
-sig_filt = filt.filter_signal(sig, fs, 'highpass', (fc, None))
+f_range = (2, None)
+sig_filt = filt.filter_signal(sig, fs, 'highpass', f_range)
 
 ###################################################################################################
 
@@ -120,8 +120,8 @@ sig = np.random.randn(len(times)) + 5 * np.sin(times*2*np.pi*3) + 4 * np.sin(tim
 ###################################################################################################
 
 # Filter the data
-fc = 20
-sig_filt = filt.filter_signal(sig, fs, 'lowpass', (None, fc))
+f_range = (None, 20)
+sig_filt = filt.filter_signal(sig, fs, 'lowpass', f_range)
 
 ###################################################################################################
 
@@ -152,8 +152,8 @@ sig = 5 * np.sin(times*2*np.pi*5) + 2 * np.sin(times*2*np.pi*60)
 ###################################################################################################
 
 # Filter the data
-fc = (58, 62)
-sig_filt = filt.filter_signal(sig, fs, 'bandstop', fc=fc, n_seconds=0.5)
+f_range = (58, 62)
+sig_filt = filt.filter_signal(sig, fs, 'bandstop', f_range, n_seconds=0.5)
 
 ###################################################################################################
 
@@ -168,14 +168,14 @@ plt.legend(loc='best')
 # Note the user warning above.
 # This is because in the computed frequency response (below),
 # the attenuation in the stopband does not go below 20dB.
-sig_filt = filt.filter_signal(sig, fs, 'bandstop', fc=fc, n_seconds=0.25,
-                              plot_freq_response=True)
+sig_filt = filt.filter_signal(sig, fs, 'bandstop', f_range, n_seconds=0.25,
+                              plot_properties=True)
 
 ###################################################################################################v
 
 # This user warning disappears if we elongate the filter
-sig_filt = filt.filter_signal(sig, fs, 'bandstop', fc=fc, n_seconds=1,
-                              plot_freq_response=True)
+sig_filt = filt.filter_signal(sig, fs, 'bandstop', f_range, n_seconds=1,
+                              plot_properties=True)
 
 ###################################################################################################
 #
@@ -208,13 +208,13 @@ fs = 100
 sig[:fs] = 0
 
 # Define the frequency band of interest
-fc = (4, 8)
+f_range = (4, 8)
 
 ###################################################################################################
 
 # Filter the data
-sig_filt_short = filt.filter_signal(sig, fs, 'bandpass', fc=fc, n_seconds=.1)
-sig_filt_long = filt.filter_signal(sig, fs, 'bandpass', fc=fc, n_seconds=1)
+sig_filt_short = filt.filter_signal(sig, fs, 'bandpass', f_range, n_seconds=.1)
+sig_filt_long = filt.filter_signal(sig, fs, 'bandpass', f_range, n_seconds=1)
 
 ###################################################################################################
 
@@ -229,11 +229,11 @@ plt.legend(loc='best')
 
 # Visualize the kernels and frequency responses
 print('Short filter')
[…]
