### 2019-04-03 edadb36
Extent & finish exanple, adding lagged coherence
