+++ examples/plot_mne_example.py
@@ -2,20 +2,55 @@
 Using NeuroDSP with MNE
 =======================
 
-This example explores some example analyses using NeuroDSP, integrated with MNE.
+This example explores some example analyses using NeuroDSP in combination with `MNE
+<https://mne-tools.github.io/>`_.
+
+NeuroDSP is a library designed to offer a large selection of methods and analyses
+that can be applied to neural time series. NeuroDSP itself does not offering
+functionality for managing multi-channel data and metadata, but can be used with
+tools that do, such as MNE. Here, we explore an example of how analyses from NeuroDSP
+can be applied to data that is managed and processed by another module, such as the
+MNE module for managing and processing M/EEG data.
+
+In particular, it explores applying some methods for finding rhythmic properties
+in an example electroencephalography (EEG) dataset, including calculating power spectra,
+running burst detection, and applying lagged coherence.
+
+This tutorial does require that you have MNE installed. If you don't already have
+MNE, you can follow instructions to get it `here
+<https://mne-tools.github.io/stable/getting_started.html>`_.
 """
 
 ###################################################################################################
 
+# General imports
 import numpy as np
 import matplotlib.pyplot as plt
+from matplotlib import cm, colors, colorbar
 
+# Import MNE, as well as the MNE sample dataset
 import mne
 from mne import io
 from mne.datasets import sample
+from mne.viz import plot_topomap
 
+# Import some NeuroDSP functions to use with MNE
 from neurodsp.spectral import compute_spectrum, trim_spectrum
 from neurodsp.burst import detect_bursts_dual_threshold
+from neurodsp.laggedcoherence import lagged_coherence
+
+###################################################################################################
+# Load & Check MNE Data
+# ---------------------
+#
+# First, we will load the example dataset from MNE, and have a quick look at the data.
+#
+# The MNE sample dataset is a combined MEG/EEG recording with an audiovisual task, as
+# described `here <https://martinos.org/mne/stable/manual/sample_dataset.html>`_.
+#
+# For the current example, we are going to sub-select only the EEG data,
+# and analyze it as continuous (non-epoched) data.
+#
 
 ###################################################################################################
 
@@ -28,14 +63,14 @@ raw = io.read_raw_fif(raw_fname, preload=True, verbose=False)
 ###################################################################################################
 
 # Select EEG channels from the dataset
-raw = raw.pick_types(meg=False, eeg=True, eog=True, exclude='bads')
+raw = raw.pick_types(meg=False, eeg=True, eog=False, exclude='bads')
 
 # Grab the sampling rate from the data
 fs = raw.info['sfreq']
 
 ###################################################################################################
 
-# Settings
+# Settings for exploring an example channel of data
 ch_label = 'EEG 058'
 t_start = 20000
 t_stop = int(t_start + (10 * fs))
@@ -43,7 +78,8 @@ t_stop = int(t_start + (10 * fs))
 ###################################################################################################
 
 # Extract an example channel to explore
-sig, times = raw.get_data(mne.pick_channels(raw.ch_names, [ch_label]), start=t_start, stop=t_stop, return_times=True)
+sig, times = raw.get_data(mne.pick_channels(raw.ch_names, [ch_label]),
+                          start=t_start, stop=t_stop, return_times=True)
 sig = np.squeeze(sig)
 
 ###################################################################################################
@@ -51,6 +87,14 @@ sig = np.squeeze(sig)
 # Plot a segment of the extracted time series data
 plt.figure(figsize=(16, 3))
 plt.plot(times, sig, 'k')
+
+###################################################################################################
[…]
