+++ neurodsp/timefrequency/tf.py
@@ -38,7 +38,7 @@ def phase_by_time(sig, fs, f_range, hilbert_increase_n=False, remove_edges=True,
     sig_filt, kernel = filter_signal(sig, fs, infer_passtype(f_range), f_range=f_range,
                                      remove_edges=False, return_filter=True, **filter_kwargs)
 
-    pha = np.angle(robust_hilbert(sig_filt, hilbert_increase_n=hilbert_increase_n))
+    pha = np.angle(robust_hilbert(sig_filt, increase_n=hilbert_increase_n))
 
     if remove_edges:
         pha = remove_filter_edges(pha, len(kernel))
@@ -75,7 +75,7 @@ def amp_by_time(sig, fs, f_range, hilbert_increase_n=False, remove_edges=True, *
     sig_filt, kernel = filter_signal(sig, fs, infer_passtype(f_range), f_range=f_range,
                                      remove_edges=False, return_filter=True, **filter_kwargs)
 
-    amp = np.abs(robust_hilbert(sig_filt, hilbert_increase_n=hilbert_increase_n))
+    amp = np.abs(robust_hilbert(sig_filt, increase_n=hilbert_increase_n))
 
     if remove_edges:
         amp = remove_filter_edges(amp, len(kernel))
