### 2019-04-03 a6515a5
Move decorators & some utils, and add multidim
