### 2019-04-03 abfefb1
Fix / move tests for decorator & utils updates
