### 2019-04-04 7b51271
Fix up spectral org, wavelets -> tf, add avg_func selection
