### 2019-04-04 3faea68
Sort out inits, adding docs & aliasing funcs to directory names
