+++ neurodsp/utils/__init__.py
@@ -1,2 +1,4 @@
+"""Utility functions."""
 
+from .utils import demean, normalize_variance
 from .utils import remove_nans, restore_nans, discard_outliers
