+++ neurodsp/filt/__init__.py
@@ -1,5 +1,6 @@
+"""Filtering functions."""
 
 from .filter import filter_signal
-from .utils import infer_passtype
 from .fir import filter_signal_fir
 from .iir import filter_signal_iir
+from .utils import infer_passtype
