+++ neurodsp/sim/__init__.py
@@ -1,3 +1,4 @@
+"""Simulate neural time series, with periodic, aperiodic, and transient components."""
 
 from .periodic import *
 from .aperiodic import *
