+++ neurodsp/timefrequency/__init__.py
@@ -1,3 +1,4 @@
+"""Time-frequency analyse of neural time series."""
 
 from .hilbert import robust_hilbert
 from .tf import phase_by_time, amp_by_time, freq_by_time
