+++ neurodsp/rhythm/__init__.py
@@ -1,3 +1,4 @@
+"""Metrics and algorithms for evaluating rhythmic components of neural time series."""
 
 from .lc import lagged_coherence
 from .swm import sliding_window_matching
