### 2019-04-04 2bafab6
Work on tests - add a lot of smoke tests, to up coverage
