+++ neurodsp/tests/test_filt_fir.py
@@ -1,9 +1,8 @@
-"""Tests fir FIR filters."""
+"""Tests for FIR filters."""
 
 from pytest import raises
 
 from neurodsp.filt.fir import *
-from neurodsp.filt.fir import _fir_checks
 
 ###################################################################################################
 ###################################################################################################
@@ -16,13 +15,14 @@ def test_filter_signal_fir(tsig):
 def test_design_fir_filter():
 
     sig_length, fs = 1000, 100
-    test_filts = {'bandpass' : (8, 12), 'bandstop' : (58, 60),
-                  'lowpass' : (None, 10), 'highpass' : (10, None)}
+    test_filts = {'bandpass' : (5, 10), 'bandstop' : (5, 6),
+                  'lowpass' : (None, 5), 'highpass' : (5, None)}
 
-#    for pass_type, f_range in test_filts.items():
-#        filter_coefs = design_fir_filter()
+    for pass_type, f_range in test_filts.items():
+        filter_coefs = design_fir_filter(sig_length, fs, pass_type, f_range)
+    assert True
 
-def test_fir_checks():
+def test_compute_filt_len():
 
     # Settings for checks
     fs = 500
@@ -32,24 +32,24 @@ def test_fir_checks():
     # Check filt_len, if defined using n_seconds
     n_seconds = 1.75 # Number chosen to create odd expected filt_len (not needing rounding up)
     expected_filt_len = n_seconds * fs
-    filt_len = _fir_checks('bandpass', f_lo, f_hi, n_cycles=None, n_seconds=n_seconds,
-                           fs=fs, sig_length=sig_length)
+    filt_len = compute_filt_len(sig_length, fs, 'bandpass', f_lo, f_hi,
+                                n_cycles=None, n_seconds=n_seconds)
     assert filt_len == expected_filt_len
 
     # Check filt_len, if defined using n_cycles
     n_cycles = 5
     expected_filt_len = int(np.ceil(fs * n_cycles / f_lo))
-    filt_len = _fir_checks('bandpass', f_lo, f_hi, n_cycles=n_cycles, n_seconds=None,
-                           fs=fs, sig_length=sig_length)
+    filt_len = compute_filt_len(sig_length, fs, 'bandpass', f_lo, f_hi,
+                                n_cycles=n_cycles, n_seconds=None)
     assert filt_len == expected_filt_len
 
     # Check filt_len, if expected to be rounded up to be odd
     n_cycles = 4
     expected_filt_len = int(np.ceil(fs * n_cycles / f_lo)) + 1
-    filt_len = _fir_checks('bandpass', f_lo, f_hi, n_cycles=n_cycles, n_seconds=None,
-                           fs=fs, sig_length=sig_length)
+    filt_len = compute_filt_len(sig_length, fs, 'bandpass', f_lo, f_hi,
+                                n_cycles=n_cycles, n_seconds=None)
     assert filt_len == expected_filt_len
 
     # Check error is raised if the filter designed is longer than the signal
     with raises(ValueError):
-        _fir_checks('bandpass', f_lo, f_hi, None, 3, fs, sig_length=1000)
+        compute_filt_len(1000, fs, 'bandpass', f_lo, f_hi, None, 3)
