+++ neurodsp/tests/test_filt_iir.py
@@ -1,4 +1,4 @@
-""".  """
+"""Tests for IIR filters."""
 
 from neurodsp.filt.iir import *
 
@@ -11,4 +11,11 @@ def test_filter_signal_iir(tsig):
     assert True
 
 def test_design_iir_filter():
-    pass
+
+    sig_length, butter_order = 1000, 3
+    test_filts = {'bandpass' : (5, 10), 'bandstop' : (5, 6),
+                  'lowpass' : (None, 5), 'highpass' : (5, None)}
+
+    for pass_type, f_range in test_filts.items():
+        filter_coefs = design_iir_filter(sig_length, pass_type, f_range, butter_order)
+    assert True
