+++ neurodsp/timefrequency/hilbert.py
@@ -1,9 +1,11 @@
-"""Hilbert transforms."""
+"""Tools for estimating time frequency properties, using hilbert methods."""
 
 import numpy as np
 from scipy.signal import hilbert
 
-from neurodsp.utils import remove_nans, restore_nans
+from neurodsp.filt import filter_signal
+from neurodsp.utils.utils import remove_nans, restore_nans
+from neurodsp.filt.utils import infer_passtype, remove_filter_edges
 
 ###################################################################################################
 ###################################################################################################
@@ -13,15 +15,16 @@ def robust_hilbert(sig, increase_n=False):
 
     Parameters
     ----------
-    sig : xx
-        xx
-    increase_n : x
-        xx
+    sig : 1d array
+        Time series.
+    increase_n : bool, optional, default: False
+        If True, zeropad the signal to length the next power of 2 when doing the hilbert transform.
+        This is because scipy.signal.hilbert can be very slow for some lengths of x.
 
     Returns
     -------
-    sig_hilb : x
-        xx
+    sig_hilb : 1d array
+        The hilbert transform of the input signal.
     """
 
     # Extract the signal that is not nan
@@ -39,3 +42,120 @@ def robust_hilbert(sig, increase_n=False):
     sig_hilb = restore_nans(sig_hilb_nonan, sig_nans, dtype=complex)
 
     return sig_hilb
+
+
+def phase_by_time(sig, fs, f_range, hilbert_increase_n=False, remove_edges=True, **filter_kwargs):
+    """Calculate the phase time series of a neural oscillation.
+
+    Parameters
+    ----------
+    sig : 1d array
+        Time series.
+    fs : float
+        Sampling rate, in Hz.
+    f_range : tuple of float
+        Frequency range, in Hz, as (low, high).
+    hilbert_increase_n : bool, optional, default: False
+        If True, zeropad the signal to length the next power of 2 when doing the hilbert transform.
+        This is because scipy.signal.hilbert can be very slow for some lengths of x.
+    remove_edges : bool, optional, default: True
+        If True, replace the samples that are within half a kernel's length to
+        the signal edge with np.nan.
+    **filter_kwargs
+        Keyword parameters to pass to `filter_signal`.
+
+    Returns
+    -------
+    pha : 1d array
+        Time series of phase.
+    """
+
+    sig_filt, kernel = filter_signal(sig, fs, infer_passtype(f_range), f_range=f_range,
+                                     remove_edges=False, return_filter=True, **filter_kwargs)
+
+    pha = np.angle(robust_hilbert(sig_filt, increase_n=hilbert_increase_n))
+
+    if remove_edges:
+        pha = remove_filter_edges(pha, len(kernel))
+
+    return pha
+
+
+def amp_by_time(sig, fs, f_range, hilbert_increase_n=False, remove_edges=True, **filter_kwargs):
+    """Calculate the amplitude time series.
+
+    Parameters
+    ----------
+    sig : 1d array
+        Time series.
+    fs : float
+        Sampling rate, in Hz.
+    f_range : tuple of float
+        The frequency filtering range, in Hz, as (low, high).
+    hilbert_increase_n : bool, optional, default: False
+        If True, zeropad the signal to length the next power of 2 when doing the hilbert transform.
+        This is because scipy.signal.hilbert can be very slow for some lengths of sig.
+    remove_edges : bool, optional, default: True
+        If True, replace the samples that are within half a kernel's length to
+        the signal edge with np.nan.
+    **filter_kwargs
+        Keyword parameters to pass to `filter_signal`.
+
+    Returns
+    -------
[…]
