+++ neurodsp/filt/__init__.py
@@ -1,6 +1,6 @@
 """Filtering functions."""
 
 from .filter import filter_signal
-from .fir import filter_signal_fir
-from .iir import filter_signal_iir
+from .fir import filter_signal_fir, design_fir_filter
+from .iir import filter_signal_iir, design_iir_filter
 from .utils import infer_passtype

[…]
