+++ doc/api.rst
@@ -16,66 +16,135 @@ Table of Contents
 Filtering
 ---------
 
-.. currentmodule:: neurodsp.filt
+General Filter Function
+~~~~~~~~~~~~~~~~~~~~~~~
 
-Design & Apply Filters
-~~~~~~~~~~~~~~~~~~~~~~
+.. currentmodule:: neurodsp.filt.filter
 
 .. autosummary::
   :toctree: generated/
 
   filter_signal
+
+
+FIR Filters
+~~~~~~~~~~~
+
+.. currentmodule:: neurodsp.filt.fir
+
+.. autosummary::
+  :toctree: generated/
+
   filter_signal_fir
-  filter_signal_iir
   design_fir_filter
+
+IIR Filters
+~~~~~~~~~~~
+
+.. currentmodule:: neurodsp.filt.iir
+
+.. autosummary::
+  :toctree: generated/
+
+  filter_signal_iir
   design_iir_filter
 
-Filter Properties
-~~~~~~~~~~~~~~~~~
+Check Filter Properties
+~~~~~~~~~~~~~~~~~~~~~~~
+
+.. currentmodule:: neurodsp.filt.checks
 
 .. autosummary::
   :toctree: generated/
 
   check_filter_definition
   check_filter_properties
-  compute_frequency_response
-  compute_pass_band
-  compute_trans_band
-  compute_nyquist
 
-Time-Frequency analysis
------------------------
+Filter Utilities
+~~~~~~~~~~~~~~~~
 
-.. currentmodule:: neurodsp.timefrequency
+.. currentmodule:: neurodsp.filt.utils
 
 .. autosummary::
   :toctree: generated/
 
+  compute_frequency_response
+  compute_pass_band
+  compute_transition_band
+  compute_nyquist
+  remove_filter_edges
+
+Time-Frequency analysis
+-----------------------
+
+Hilbert Methods
+~~~~~~~~~~~~~~~
+
+.. currentmodule:: neurodsp.timefrequency.hilbert
+
+.. autosummary::
+  :toctree: generated/
+
+  robust_hilbert
   phase_by_time
   amp_by_time
   freq_by_time
 
+Wavelet Methods
+~~~~~~~~~~~~~~~
+
+.. currentmodule:: neurodsp.timefrequency.wavelets
+
+.. autosummary::
+  :toctree: generated/
+
+  morlet_transform
+  morlet_convolve
+
 Spectral analysis
 -----------------
 
-.. currentmodule:: neurodsp.spectral
+Spectral Power Measures
+~~~~~~~~~~~~~~~~~~~~~~~
+
+.. currentmodule:: neurodsp.spectral.spectral
 
 .. autosummary::
    :toctree: generated/
 
-   compute_spectrum
-   compute_scv
-   compute_scv_rs
-   spectral_hist
-   morlet_transform
-   morlet_convolve
-   rotate_powerlaw
+  compute_spectrum
+  compute_spectrum_welch
+  compute_spectrum_wavelet
+  compute_spectrum_medfilt
 
-Rhythmic analysis
------------------
+Spectral Variance Measures
+~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. currentmodule:: neurodsp.spectral.spectral
+
+.. autosummary::
+   :toctree: generated/
+
+  compute_scv
+  compute_scv_rs
+  compute_spectral_hist
+
+Spectral Utilities
+~~~~~~~~~~~~~~~~~~
+
+.. currentmodule:: neurodsp.spectral.utils
+
+.. autosummary::
+   :toctree: generated/
+
[…]
