### 2019-04-04 6e6ee9c
Add 'docstring' type sentences to API listings
