### 2019-04-04 neurodsp-tools/neurodsp#143 @TomDonoghue
Done with #146
