+++ _sources/generated/neurodsp.filt.filter_signal_fir.rst.txt
@@ -1,0 +1,6 @@
+neurodsp.filt.filter\_signal\_fir
+=================================
+
+.. currentmodule:: neurodsp.filt
+
+.. autofunction:: filter_signal_fir
