+++ _sources/auto_tutorials/plot_1-Filtering.rst.txt
@@ -96,7 +96,8 @@ Extract signal within a specific frequency range (e.g. theta, 4-8 Hz).
 
  .. code-block:: none
 
-    Transition bandwidth is 2.0 Hz. Pass/stop bandwidth is 4 Hz
+    Transition bandwidth is 2.0 Hz.
+    Pass/stop bandwidth is 4.0 Hz.
 
 
 
@@ -169,7 +170,8 @@ Remove low frequency drift from the data
 
  .. code-block:: none
 
-    Transition bandwidth is 1.0 Hz. Pass/stop bandwidth is 498.0 Hz
+    Transition bandwidth is 1.0 Hz.
+    Pass/stop bandwidth is 498.0 Hz.
 
 
 
@@ -230,7 +232,8 @@ Remove high frequency activity from the data.
 
  .. code-block:: none
 
-    Transition bandwidth is 8.8 Hz. Pass/stop bandwidth is 20 Hz
+    Transition bandwidth is 8.8 Hz.
+    Pass/stop bandwidth is 20.0 Hz.
 
 
 
@@ -350,7 +353,8 @@ the 60Hz oscillation much (try this yourself!).
 
  .. code-block:: none
 
-    Transition bandwidth is 2.0 Hz. Pass/stop bandwidth is 4 Hz
+    Transition bandwidth is 2.0 Hz.
+    Pass/stop bandwidth is 4.0 Hz.
 
 
 3. Time-frequency resolution tradeoff
@@ -419,8 +423,10 @@ but the short filter did not (i.e. better frequency resolution).
 
  .. code-block:: none
 
-    Transition bandwidth is 9.9 Hz. Pass/stop bandwidth is 4 Hz
-    Transition bandwidth is 1.4 Hz. Pass/stop bandwidth is 4 Hz
+    Transition bandwidth is 9.9 Hz.
+    Pass/stop bandwidth is 4.0 Hz.
+    Transition bandwidth is 1.4 Hz.
+    Pass/stop bandwidth is 4.0 Hz.
 
 
 
@@ -479,11 +485,13 @@ but the short filter did not (i.e. better frequency resolution).
  .. code-block:: none
 
     Short filter
-    Transition bandwidth is 9.9 Hz. Pass/stop bandwidth is 4 Hz
+    Transition bandwidth is 9.9 Hz.
+    Pass/stop bandwidth is 4.0 Hz.
 
 
     Long filter
-    Transition bandwidth is 1.4 Hz. Pass/stop bandwidth is 4 Hz
+    Transition bandwidth is 1.4 Hz.
+    Pass/stop bandwidth is 4.0 Hz.
 
 
 4. Infinite impulse response (IIR) filter option
@@ -535,7 +543,8 @@ Notice that some edge artifacts remain.
 
  .. code-block:: none
 
-    Transition bandwidth is 43.9 Hz. Pass/stop bandwidth is 100 Hz
+    Transition bandwidth is 43.9 Hz.
+    Pass/stop bandwidth is 100.0 Hz.
 
 
 
@@ -556,7 +565,8 @@ Notice that some edge artifacts remain.
 
  .. code-block:: none
 
-    Transition bandwidth is 1.0 Hz. Pass/stop bandwidth is 4 Hz
+    Transition bandwidth is 1.0 Hz.
+    Pass/stop bandwidth is 4.0 Hz.
 
 
 
@@ -616,6 +626,14 @@ Notice that some edge artifacts remain.
     :class: sphx-glr-single-img
 
 
+.. rst-class:: sphx-glr-script-out
+
+ Out:
+
+ .. code-block:: none
+
+    Transition bandwidth is 5.9 Hz.
[…]
