+++ _sources/generated/neurodsp.filt.design_iir_filter.rst.txt
@@ -1,0 +1,6 @@
+neurodsp.filt.design\_iir\_filter
+=================================
+
+.. currentmodule:: neurodsp.filt
+
+.. autofunction:: design_iir_filter
