### 2019-04-04 #1272 @TomDonoghue
Hey @TomDLT & @arokem

I've finished up a version of an example using MNE, as was suggested in the review, and the relevant PR has now been merged:
https://github.com/neurodsp-tools/neurodsp/pull/146
You can also see the hosted version of this example here:
https://neurodsp-tools.github.io/neurodsp/auto_examples/plot_mne_example.html#sphx-glr-auto-examples-plot-mne-example-py

One more thing - we have been working more generally on  some updates for the module (see: https://github.com/neurodsp-tools/neurodsp/pull/150). Related to this, and to make sure the paper will still accurately reflect the module after these changes go through, we have some proposed updates to the JOSS paper, which are here:
https://github.com/neurodsp-tools/neurodsp/pull/151

Basically: the upcoming changes include some code re-organizations / refactors (no big changes to most implementations), and the updates in that PR make the paper match a little closer to what the module will look like soon. That PR for the paper edits is basically ready to be merged - I just wasn't quite sure the protocol for us initiating changes on the paper through the review process, and wanted to mention it here first before merging!

If the MNE example looks good to y'all, and if the paper edits we recommend look okay, then I think we're done everything we have noted on our end for the JOSS paper - so please let us know if there is anything else we should do. Thanks!

cc: @voytek
