### 2019-04-04 #1272 @TomDLT
Nice MNE example, and moving swm and lagged coherence in the same module makes sense.
Everything is ok on my side.
