### 2019-04-05 dda7cfd
Add bunch of definitions to glossary
