+++ doc/glossary.rst
@@ -1,65 +1,117 @@
 NeuroDSP Glossary
 =================
 
+The following is a glossary of neuroscience and digital processing related terms that are used in NeuroDSP.
+
 General
 -------
-
 .. glossary::
 
     time domain
-        xx
+        Signals that are represented as variations over time, and analyses of such signals.
     frequency domain
-        xx
-    convolution
-        xx
+        Signals that are represented in terms of frequencies, and Analyses of such signals.
     sampling rate
-        xx
-    coherence
-        xx
+        The rate at which samples are taken.
+    periodic
+        Properties or components of a signal that are rhythmic.
+    aperiodic
+        Properties or components of a signal that are arrhythmic, with no characteristic frequency.
+    temporal resolution
+        The precision of a measurement, in the time domain.
+    frequency resolution
+        The precision of a measurement, in the frequency domain.
 
-Time Frequency
---------------
-
+Units
+-----
 .. glossary::
 
-    frequency
-        xx
-    phase
-        xx
-    amplitude
-        xx
+    Hertz (Hz)
+        A unit of frequency, as the number of cycles per second.
+    Decibels (dB)
+        A unit of intensity, on a logarithmic scale.
+    Volts (V)
+        A unit of voltage, typically in the microvolt (uV) range for neural time series.
 
 Filters
 -------
-
 .. glossary::
 
+    Impulse Response
+        The response of a fitler when presented with an impulse; a single, brief input.
     FIR
-        xx
+        A Finite Impulse Response filter, meaning the response to a single input is finite, settling to zero.
     IIR
-        xx
-    pass type (bandpass, bandstop, lowpass, highpass)
-        xx
-    pass band
-        xx
+        An Infinite Impulse Response filter, meaning the response to a single input continue infinitely.
+    passband
+        The range (band) of frequencies that can pass through a filter.
+    stopband
+        The range (band) of frequencies that are attenuated (stopped) by a filter.
+    passtype
+        The type of filter, defined in terms of what frequency bands or ranges it passes through, or filters out.
+
+        * bandpass: a filter whose passband is a specific frequency band, bound by a low and high frequency point.
+        * bandstop: a filter that passes through all frequencies except a band region that is attenuated.
+        * lowpass: a filter whose passband is all frequencies below a filter frequency (low frequencies pass through).
+        * highpass: a filter whose passband is all frequencies above a filter frequency (high frequencies pass through).
     transition band
-        xx
-    kernel
-        xx
+        The range of frequencies that are in the transition region between the passband and the stopband.
     frequency response
-        xx
-    kernel
-        xx
-    b_vals / a_vals
-        xx
+        The response profile of a filter, specifying the gain and phase shift applied by the filter at each frequency.
+
+Rhythms & Bursts
+----------------
+.. glossary::
+
+    burst
+        Periodic activity that lasts for a finite time , as in a 'burst of oscillatory activity'.
+
+Time Frequency
+--------------
+.. glossary::
+
+    frequency
+        The number of occurences over a unit of time.
+    phase
+        The position, at a point in time, on a waveform cycle.
+    amplitude
+        The magnitude of a signal, as the peak-to-trough.
+    period
+        A single cycle of a rhythm, defined as the time between two consecutive troughs (or peaks).
+    hilbert transform
+        A mathematical transform that derives the analyic representation of a signal, where the
+        analytic representation is complex-valued representation that can be used to find the
+        analytic amplitude and phase of a signal.
+    wavelet
+        A wave-like signal, or 'brief oscillation', that starts at zero amplitude, increases
+        in amplitude to some value, and then decays back to zero.
 
 Spectral
 --------
 .. glossary::
 
     fourier transform
-        xx
[…]
