### 2019-04-05 #1272 @TomDonoghue
Awesome - thanks for doing the review @TomDLT!

@arokem is it okay if I merge paper updates? And is there anything else you need from us?
