### 2019-04-07 8fb736b
Update combined to set variance & add - remove old sim utils
