+++ neurodsp/tests/test_sim_utils.py
@@ -1,11 +1,0 @@
-"""Tests for sim utility functions."""
-
-import numpy as np
-
-from neurodsp.sim.utils import *
-
-###################################################################################################
-###################################################################################################
-
-def test_proportional_sum():
-    pass
