+++ neurodsp/sim/combined.py
@@ -10,7 +10,7 @@ from neurodsp.sim.utils import proportional_sum
 ###################################################################################################
 
 @normalize
-def sim_combined(n_seconds, fs, simulations, proportions=1):
+def sim_combined(n_seconds, fs, simulations, variances=1):
     """Sim multiple component signals and combine them.
 
     Parameters
@@ -21,8 +21,8 @@ def sim_combined(n_seconds, fs, simulations, proportions=1):
         Signal sampling rate, in Hz.
     simulations : dictionary
         A dictionary of simulation functions to run, with their desired parameters.
-    proportions : float of list of float
-        Variance proportions to use to sum across signals.
+    variances : list of float or 1
+        Specified variance for each component of the signal.
 
     Returns
     -------
@@ -30,18 +30,17 @@ def sim_combined(n_seconds, fs, simulations, proportions=1):
         Simulated combined signal.
     """
 
-    if not (proportions == 1 or len(proportions) == len(simulations)):
+    if not (variances == 1 or len(variances) == len(simulations)):
         raise ValueError('Simulations and proportions lengths do not match.')
 
+    # Collect the sim function to use, and repeat variance if is set to 1
     simulations = {(get_sim_func(name) if isinstance(name, str) else name) : params \
                    for name, params in simulations.items()}
+    variances = repeat(variances) if isinstance(variances, int) else variances
 
-    proportions = repeat(proportions) if isinstance(proportions, int) else proportions
-
-    #components = [func(n_seconds, fs, **params) for func, params in simulations.items()]
+    # Simulate each component, specifying variance, and combine them
     components = [func(n_seconds, fs, **params, variance=variance) for \
-        (func, params), variance in zip(simulations.items(), proportions)]
-
-    sig = proportional_sum(components, proportions)
+        (func, params), variance in zip(simulations.items(), variances)]
+    sig = np.sum(components, axis=0)
 
     return sig
