### 2019-04-07 #1272 @arokem
Yes. Please do merge the paper updates and the updates. I think that it would be better for the paper to represent a more up-to-date version of the software. Once you do that, I'll take another look at the rendered PDF to make sure that everything is in order. Then, I'll ask you to deposit an archive with the accepted version of the software and article, and then, we can move on to ask the EIC for a final look and approval. Does that all make sense?
