### 2019-04-08 ea59b52
Small updates & clean ups of spectral
