+++ neurodsp/spectral/spectral.py
@@ -30,7 +30,9 @@ def compute_spectrum(sig, fs, method='welch', avg_type='mean', **kwargs):
     method : {'welch', 'wavelet', 'medfilt'}
         Method to use to estimate the power spectrum.
     avg_type : {'mean', 'median'}, optional
-        If relevant, the method to average across windows to create the spectrum:
+        If relevant, the method to average across windows to create the spectrum.
+    **kwargs
+        Keyword arguments to pass through to function that calculates the spectrum.
 
     Returns
     -------
@@ -43,14 +45,14 @@ def compute_spectrum(sig, fs, method='welch', avg_type='mean', **kwargs):
     if method not in ('welch', 'medfilt', 'wavelet'):
         raise ValueError('Unknown power spectrum method: %s' % method)
 
-    if method in ('welch'):
+    if method == 'welch':
         return compute_spectrum_welch(sig, fs, avg_type=avg_type, **kwargs)
 
     elif method == 'wavelet':
         return compute_spectrum_wavelet(sig, fs, avg_type=avg_type, **kwargs)
 
     elif method == 'medfilt':
-        return compute_spectrum_medfilt(sig, fs, filt_len, f_lim)
+        return compute_spectrum_medfilt(sig, fs, **kwargs)
 
 
 def compute_spectrum_wavelet(sig, fs, freqs, avg_type='mean', **kwargs):
@@ -81,8 +83,9 @@ def compute_spectrum_wavelet(sig, fs, freqs, avg_type='mean', **kwargs):
     return freqs, spectrum
 
 
-def compute_spectrum_welch(sig, fs, avg_type='mean', window='hann', nperseg=None,
-                           noverlap=None, f_lim=None, spg_outlier_pct=0.):
+def compute_spectrum_welch(sig, fs, avg_type='mean', window='hann',
+                           nperseg=None, noverlap=None,
+                           f_range=None, outlier_pct=None):
     """Estimate the power spectral density using Welch's method.
 
     Parameters
@@ -107,12 +110,10 @@ def compute_spectrum_welch(sig, fs, avg_type='mean', window='hann', nperseg=None
     noverlap : int, optional
         Number of points to overlap between segments.
         If None, noverlap = nperseg // 8.
-    f_lim : float, optional
-        Maximum frequency to keep, in Hz.
-        If None, keeps up to Nyquist.
-    spg_outlier_pct : float, optional, default: 0.
-        Percentage of spectrogram windows with the highest powers to discard prior to averaging.
-        Useful for quickly eliminating potential outliers to compute spectrum.
+    f_range : list of [float, float] optional
+        Frequency range to sub-select from the power spectrum.
+    outlier_pct : float, optional
+        Percentage of the windows with the lowest and highest total log power to discard.
         Must be between 0 and 100.
 
     Returns
@@ -132,20 +133,20 @@ def compute_spectrum_welch(sig, fs, avg_type='mean', window='hann', nperseg=None
         sig = sig[np.newaxis :]
 
     # Throw out outliers if indicated
-    if spg_outlier_pct > 0.:
-        spg = discard_outliers(spg, spg_outlier_pct)
+    if outlier_pct is not None:
+        spg = discard_outliers(spg, outlier_pct)
 
     # Average across windows
     spectrum = get_avg_func(avg_type)(spg, axis=-1)
 
     # Trim spectrum, if requested
-    if f_lim:
-        freqs, spectrum = trim_spectrum(freqs, spectrum, [freqs[0], f_lim])
+    if f_range:
+        freqs, spectrum = trim_spectrum(freqs, spectrum, f_range)
 
[…]
