### 2019-04-08 2269c7b
Update tutorials for code updates
