+++ tutorials/plot_3-LaggedCoherence.py
@@ -22,7 +22,7 @@ This tutorial primarily covers :mod:`neurodsp.laggedcoherence`.
 import numpy as np
 import matplotlib.pyplot as plt
 
-from neurodsp.laggedcoherence import lagged_coherence
+from neurodsp.rhythm import lagged_coherence
 
 ###################################################################################################
 
