+++ neurodsp/sim/periodic.py
@@ -36,16 +36,18 @@ def sim_oscillation(n_seconds, fs, freq, rdsym=.5):
 
     # Compute number of samples per cycle and number of cycles
     n_samples_cycle = int(np.ceil(fs / freq))
-    n_samples = int(fs * n_seconds)
     n_cycles = int(np.ceil(n_seconds * freq))
+    n_samples = int(fs * n_seconds)
 
     # Determine number of samples in rise and decay periods
     rise_samples = int(np.round(n_samples_cycle * rdsym))
     decay_samples = n_samples_cycle - rise_samples
 
-    # Make phase array for a single cycle, then repeat it
+    # Make phase array for a single cycle
     pha_one_cycle = np.hstack([np.linspace(0, np.pi, decay_samples + 1),
                                np.linspace(-np.pi, 0, rise_samples + 1)[1:-1]])
+
+    # Repeat the phase to make the full signal
     phase_t = np.tile(pha_one_cycle, n_cycles)
     phase_t = phase_t[:n_samples]
 
@@ -125,24 +127,22 @@ def sim_bursty_oscillation(n_seconds, fs, freq, rdsym=.5, prob_enter_burst=.2,
     n_samples = int(n_seconds * fs)
     n_cycles_overestimate = int(np.ceil(n_samples / mean_period_samples * 2))
 
-    # Simulate if a series of cycles are oscillating or not oscillating
+    # Determine which periods will be oscillating and the cycle properties for each cycle
     is_oscillating = _make_is_osc(n_cycles_overestimate, prob_enter_burst, prob_leave_burst)
-
-    # Determine what the cycle properties shall be for each cycle
     periods, amps, rdsyms = _determine_cycle_properties(is_oscillating, cycle_features_use, n_tries)
 
     # Set up the dataframe of parameters
-    df = pd.DataFrame({'is_cycle': is_oscillating, 'period': periods, 'amp': amps, 'rdsym': rdsyms})
+    df = pd.DataFrame({'is_cycle': is_oscillating, 'period': periods,
+                       'amp': amps, 'rdsym': rdsyms})
     df['start_sample'] = np.insert(df['period'].cumsum().values[:-1], 0, 0)
     df = df[df['start_sample'] < n_samples]
-    # Shorten df to only cycles that are included in the data
 
     # Create the signal
     sig = _sim_cycles(df)
     sig = sig[:n_samples]
 
     if return_cycle_df:
-        # Remove last row of df
+        # Shorten df to only cycles that are included in the data
         df.drop(df.index[len(df) - 1], inplace=True)
         return sig, df
     else:
@@ -151,8 +151,7 @@ def sim_bursty_oscillation(n_seconds, fs, freq, rdsym=.5, prob_enter_burst=.2,
 
 @normalize
 def sim_jittered_oscillation(n_seconds, fs, freq, jitter=0, cycle=('gaussian', 0.01)):
-    """Simulate a jittered oscillation, as defined by the oscillation frequency,
-    the oscillation cycle, and how much (in time) to jitter each period.
+    """Simulate a jittered oscillation.
 
     Parameters
     ----------
@@ -182,8 +181,8 @@ def sim_jittered_oscillation(n_seconds, fs, freq, jitter=0, cycle=('gaussian', 0
     # If cycle is a tuple, generate the window with given params,
     if isinstance(cycle, tuple):
 
-        # defaults to 1 second window for a cycle, which is more than enough
-        # if interested in longer period oscillations, just pass in premade cycle
+        # Defaults to 1 second window for a cycle
+        #   If interested in longer period oscillations, pass in premade cycle
         osc_cycle = sim_osc_cycle(1, fs, cycle)
 
[…]
