### 2019-04-08 39083a5
Fix up n_samps for sim_powerlaw - note: bit of a hack right now
