+++ neurodsp/tests/test_sim_combined.py
@@ -5,8 +5,8 @@ from neurodsp.sim.combined import *
 ###################################################################################################
 ###################################################################################################
 
-FS = 1000
-N_SECONDS = 10
+FS = 100
+N_SECONDS = 1
 
 def test_sim_combined():
 
