+++ neurodsp/tests/test_sim_transients.py
@@ -5,7 +5,7 @@ from neurodsp.sim.transients import *
 ###################################################################################################
 ###################################################################################################
 
-FS = 1000
+FS = 100
 
 def test_sim_make_synaptic_kernel():
 
