### 2019-04-08 0f0778e
Extend sim_osc_cycle to have more kernels / consolidate approach
