### 2019-04-08 cb774bd
Doc updates & add create_times helper func
