### 2019-04-08 2c065a6
Refactor some periodic stuff to use osc cycle kernels
