+++ neurodsp/tests/test_sim_periodic.py
@@ -23,8 +23,8 @@ def test_sim_bursty_oscillation():
 def test_sim_jittered_oscillation():
     pass
 
-    #sig = sim_jittered_oscillation(N_SECONDS, FS, FREQ, 'gaussian', std=0.01)
-    #assert True
+    sig = sim_jittered_oscillation(N_SECONDS, FS, FREQ)
+    assert True
 
 ## PRIVATE FUNCTIONS
 
