### 2019-04-08 b197fb0
Update normalization appraoches to deal with zeros in bursts
