+++ neurodsp/tests/test_utils.py
@@ -10,19 +10,37 @@ from neurodsp.utils.utils import *
 
 def test_demean():
 
-    dat = np.array([2, 2, 2])
-    dat = demean(dat)
+    d1 = np.array([1, 2, 3])
+    d2 = np.array([0, 1, 2, 3, 0])
 
-    # TODO: Add isclose checking here
-    assert True
+    # Check default - demean to 0
+    out1 = demean(d1)
+    assert np.isclose(out1.mean(), 0.)
+
+    # Check demeaning and adding specific mean
+    out2 = demean(d1, mean=1.)
+    assert np.isclose(out2.mean(), 1.)
+
+    # Check dealing with zero entries
+    out3 = demean(d2)
+    assert np.isclose(out3[np.nonzero(out3)].mean(), 0)
+
+    # Check turning of non-zero selection
+    out3 = demean(d2, mean=1, select_nonzero = False)
+    assert np.isclose(out3.mean(), 1)
 
 def test_normalize_variance():
 
-    dat = np.array([2, 2, 2])
-    dat = normalize_variance(dat)
+    d1 = np.array([1, 2, 3])
+    d2 = np.array([0, 1, 2, 3, 0])
 
-    # TODO: Add isclose checking here
-    assert True
+    # Check default - normalize variance to 1
+    out1 = normalize_variance(d1)
+    np.isclose(out1.var(), 1.)
+
+    # Check normalizing and add specific variance
+    out2 = normalize_variance(d1, 2.)
+    np.isclose(out2.var(), 2.)
 
 def test_remove_nans():
 
