### 2019-04-08 13c0c1c
Add bursting func that uses osc_cycle
