### 2019-04-08 0fc16b9
Update doc/tutorials for updates
