### 2019-04-08 657ce89
Add and use check_ax helper function
