+++ neurodsp/plts/utils.py
@@ -1,0 +1,25 @@
+"""Utility functions for NeuroDSP plots."""
+
+import matplotlib.pyplot as plt
+
+###################################################################################################
+###################################################################################################
+
+def check_ax(ax, figsize=None):
+    """Check whether a figure axes object is defined, define if not.
+
+    Parameters
+    ----------
+    ax : matplotlib.Axes or None
+        Axes object to check if is defined.
+
+    Returns
+    -------
+    ax : matplotlib.Axes
+        Figure axes object to use.
+    """
+
+    if not ax:
+        _, ax = plt.subplots(figsize=figsize)
+
+    return ax
