### 2019-04-08 cf0b384
Add plots for rhythm
