### 2019-04-08 e70be1c
Add time series plts
