### 2019-04-08 5b1973c
Add plot tests, and associated fixes/updates
