+++ neurodsp/plts/rhythm.py
@@ -2,6 +2,8 @@
 
 import matplotlib.pyplot as plt
 
+from neurodsp.plts.utils import check_ax
+
 ###################################################################################################
 ###################################################################################################
 
