+++ neurodsp/plts/time_series.py
@@ -6,6 +6,8 @@ import numpy as np
 import numpy.ma as ma
 import matplotlib.pyplot as plt
 
+from neurodsp.plts.utils import check_ax
+
 ###################################################################################################
 ###################################################################################################
 
