### 2019-04-09 8184efd
Add spectral plts
