### 2019-04-09 fe362e1
Add decorator for saving plts
