+++ neurodsp/plts/utils.py
@@ -1,5 +1,8 @@
 """Utility functions for NeuroDSP plots."""
 
+from functools import wraps
+from os.path import join as pjoin
+
 import matplotlib.pyplot as plt
 
 ###################################################################################################
@@ -23,3 +26,21 @@ def check_ax(ax, figsize=None):
         _, ax = plt.subplots(figsize=figsize)
 
     return ax
+
+
+def savefig(func):
+
+    @wraps(func)
+    def decorated(*args, **kwargs):
+
+        save_fig = kwargs.pop('save_fig', False)
+        file_name = kwargs.pop('file_name', None)
+        file_path = kwargs.pop('file_path', None)
+
+        func(*args, **kwargs)
+
+        if save_fig:
+            full_path = pjoin(file_path, file_name) if file_path else file_name
+            plt.savefig(full_path)
+
+    return decorated
