### 2019-04-09 0bdceb3
Add decorators, small updates
