### 2019-04-09 63b03b3
Update tutorials to use plot funcs
