+++ tutorials/plot_1-Filtering.py
@@ -27,6 +27,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 from neurodsp import filt
+from neurodsp.plts.time_series import plot_time_series
 
 ###################################################################################################
 
@@ -57,10 +58,7 @@ sig_filt = filt.filter_signal(sig, fs, 'bandpass', f_range)
 ###################################################################################################
 
 # Plot filtered signal
-plt.figure(figsize=(15, 3))
-plt.plot(times, sig, 'k', label='raw')
-plt.plot(times, sig_filt, 'r', label='filtered')
-plt.legend(loc='best')
+plot_time_series(times, [sig, sig_filt], ['Raw', 'Filtered'])
 
 ###################################################################################################
 #
@@ -98,10 +96,7 @@ sig_filt = filt.filter_signal(sig, fs, 'highpass', f_range)
 ###################################################################################################
 
 # Plot filtered signal
-plt.figure(figsize=(15, 3))
-plt.plot(times, sig, 'k', label='raw')
-plt.plot(times, sig_filt, 'r', label='filtered')
-plt.legend(loc='best')
+plot_time_series(times, [sig, sig_filt], ['Raw', 'Filtered'])
 
 ###################################################################################################
 #
@@ -126,10 +121,7 @@ sig_filt = filt.filter_signal(sig, fs, 'lowpass', f_range)
 ###################################################################################################
 
 # Plot filtered signal
-plt.figure(figsize=(15, 3))
-plt.plot(times, sig, 'k', label='raw')
-plt.plot(times, sig_filt, 'r', label='filtered')
-plt.legend(loc='best')
+plot_time_series(times, [sig, sig_filt], ['Raw', 'Filtered'])
 
 ###################################################################################################
 #
@@ -158,10 +150,7 @@ sig_filt = filt.filter_signal(sig, fs, 'bandstop', f_range, n_seconds=0.5)
 ###################################################################################################
 
 # Plot filtered signal
-plt.figure(figsize=(15, 3))
-plt.plot(times, sig, 'k', label='raw')
-plt.plot(times, sig_filt, 'r', label='filtered')
-plt.legend(loc='best')
+plot_time_series(times, [sig, sig_filt], ['Raw', 'Filtered'])
 
 ###################################################################################################
 
@@ -219,11 +208,8 @@ sig_filt_long = filt.filter_signal(sig, fs, 'bandpass', f_range, n_seconds=1)
 ###################################################################################################
 
 # Plot filtered signal
-plt.figure(figsize=(15, 3))
-plt.plot(times, sig, 'k', label='raw')
-plt.plot(times, sig_filt_short, 'r', label='short filter')
-plt.plot(times, sig_filt_long, 'b', label='long filter')
-plt.legend(loc='best')
+plot_time_series(times, [sig, sig_filt_short, sig_filt_long],
+                 ['Raw', 'Short Filter', 'Long Filter'])
 
 ###################################################################################################
 
@@ -274,10 +260,7 @@ sig_filt = filt.filter_signal(sig, fs, 'bandstop', f_range, filt_type='iir', but
 ###################################################################################################
 
 # Plot filtered signal
-plt.figure(figsize=(15, 3))
-plt.plot(times, sig, 'k', label='raw')
-plt.plot(times, sig_filt, 'r', label='filtered')
-plt.legend(loc='best')
+plot_time_series(times, [sig, sig_filt], ['Raw', 'Filtered'])
 plt.xlim((0, .2))
 
 ###################################################################################################
@@ -303,11 +286,8 @@ sig_filt, kernel = filt.filter_signal(sig, fs, 'bandpass', f_range, n_cycles=3,
[…]
