+++ tutorials/plot_3-LaggedCoherence.py
@@ -23,6 +23,8 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 from neurodsp.rhythm import lagged_coherence
+from neurodsp.plts.time_series import plot_time_series
+from neurodsp.plts.rhythm import plot_lagged_coherence
 
 ###################################################################################################
 
@@ -60,8 +62,7 @@ for ind in burst_starts:
 ###################################################################################################
 
 # Plot example signal
-plt.figure(figsize=(12, 3))
-plt.plot(times, sig, 'k')
+plot_time_series(times, sig)
 
 ###################################################################################################
 #
@@ -85,11 +86,7 @@ print('Lagged coherence = ', lag_coh_alpha)
 lag_coh_by_f, freqs = lagged_coherence(sig, (1, 40), fs, return_spectrum=True)
 
 # Visualize lagged coherence as a function of frequency
-plt.figure(figsize=(6, 3))
-plt.plot(freqs, lag_coh_by_f, 'k.-')
-plt.xlabel('Frequency (Hz)')
-plt.ylabel('Lagged coherence')
-plt.tight_layout()
+plot_lagged_coherence(freqs, lag_coh_by_f)
 
 ###################################################################################################
 #
@@ -125,12 +122,7 @@ f_range = (13, 30)
 ###################################################################################################
 
 # Plot example signal
-plt.figure(figsize=(12, 3))
-plt.plot(times, sig, 'k')
-plt.xlim((0, 5))
-plt.xlabel('Time (s)')
-plt.ylabel('Voltage (uV)')
-plt.tight_layout()
+plot_time_series(times, sig)
 
 ###################################################################################################
 
