### 2019-04-09 a28ac41
Alias all plots to plts folder
