### 2019-04-09 3b6c14a
Fix typo & lint fixes
