### 2019-04-09 30afca8
Continue updating tutorials to use plts module
