+++ tutorials/plot_7-SimulatingSignals.py
@@ -14,6 +14,9 @@ import matplotlib.pyplot as plt
 
 from neurodsp import spectral, sim
 
+from neurodsp.plts.spectral import plot_power_spectra
+from neurodsp.plts.time_series import plot_time_series
+
 ###################################################################################################
 
 # Set the random seed, for consistency simulating data
@@ -41,19 +44,13 @@ br_noise = sim.sim_powerlaw(n_seconds, fs, exponent)
 ###################################################################################################
 
 # Plot the simulated data, in the time domain
-plt.figure(figsize=(12, 3))
-plt.plot(times, br_noise, 'k')
-plt.xlabel('Time (s)')
-plt.ylabel('Voltage')
+plot_time_series(times, br_noise)
 
 ###################################################################################################
 
 # Plot the simulated data, in the frequency domain
 freqs, psd = spectral.compute_spectrum(br_noise, fs)
-plt.figure(figsize=(5, 5))
-plt.loglog(freqs, psd, 'k')
-plt.xlabel('Frequency (Hz)')
-plt.ylabel('Power')
+plot_power_spectra(freqs, psd)
 
 ###################################################################################################
 #
@@ -79,19 +76,13 @@ brown_filt = sim.sim_powerlaw(n_seconds, fs, exponent, f_range=(f_hipass_brown, 
 ###################################################################################################
 
 # Plot the simulated data, in the time domain
-plt.figure(figsize=(12, 3))
-plt.plot(times, brown_filt, 'k')
-plt.xlabel('Time (s)')
-plt.ylabel('Voltage')
+plot_time_series(times, brown_filt)
 
 ###################################################################################################
 
 # Plot the simulated data, in the frequency domain
 freqs, psd = spectral.compute_spectrum(brown_filt, fs)
-plt.figure(figsize=(5, 5))
-plt.loglog(freqs, psd, 'k')
-plt.xlabel('Frequency (Hz)')
-plt.ylabel('Power')
+plot_power_spectra(freqs, psd)
 
 ###################################################################################################
 #
@@ -123,25 +114,18 @@ syn_noise = sim.sim_synaptic_current(n_seconds, fs)
 ###################################################################################################
 
 # Plot the simulated data, in the time domain
-plt.figure(figsize=(12, 6))
-plt.subplot(2, 1, 1, title='OU Process')
-plt.plot(times, ou_noise, 'k')
-plt.subplot(2, 1, 2, title='Synaptic Noise')
-plt.plot(times, syn_noise, 'r')
-plt.xlabel('Time (s)')
-plt.ylabel('Voltage')
+plot_time_series(times, ou_noise)
+plt.title('OU Process')
+plot_time_series(times, syn_noise)
+plt.title('Synaptic Noise')
 
 ###################################################################################################
 
 # Plot the simulated data, in the frequency domain
 freqs, ou_psd = spectral.compute_spectrum(ou_noise, fs)
 freqs, syn_psd = spectral.compute_spectrum(syn_noise, fs)
-plt.figure(figsize=(5, 5))
-plt.loglog(freqs, ou_psd, 'k', label='OU')
-plt.loglog(freqs, syn_psd, 'r', label='Synaptic')
-plt.legend()
-plt.xlabel('Frequency (Hz)')
-plt.ylabel('Power')
+
+plot_power_spectra(freqs, [ou_psd, syn_psd], ['OU', 'Synaptic'])
 
 ###################################################################################################
 #
@@ -155,34 +139,30 @@ plt.ylabel('Power')
 ###################################################################################################
 
 # Simulate symmetric oscillator
[…]
