### 2019-04-09 c28c88a
Fix bug with dealing with labels that stoppd some multi-plotting
