### 2019-04-09 e5768ff
Update style approach to catch and apply some plot args
