+++ neurodsp/plts/style.py
@@ -7,8 +7,13 @@ import matplotlib.pyplot as plt
 ###################################################################################################
 ###################################################################################################
 
-def plot_style(ax):
+PLOT_STYLE_ARGS = ['title', 'xlabel', 'ylabel', 'xlim', 'ylim']
+
+def plot_style(ax, **kwargs):
     """Define plot style."""
+
+    # Set any provided plot arguments
+    ax.set(**kwargs)
 
     # Aesthetics and axis labels
     ax.xaxis.label.set_size(16)
@@ -19,6 +24,10 @@ def plot_style(ax):
     if ax.get_legend_handles_labels()[0]:
         ax.legend(prop={'size': 12}, loc='best')
 
+    # If a title was provided, update the size
+    if ax.get_title():
+        ax.title.set_size(20)
+
     plt.tight_layout()
 
 
@@ -28,8 +37,10 @@ def style_plot(func, *args, **kwargs):
     @wraps(func)
     def decorated(*args, **kwargs):
 
+        style_kwargs = {key : kwargs.pop(key) for key in PLOT_STYLE_ARGS if key in kwargs}
+
         style_func = kwargs.pop('plot_style', plot_style)
         func(*args, **kwargs)
-        style_func(plt.gca())
+        style_func(plt.gca(), **style_kwargs)
 
     return decorated
