+++ neurodsp/tests/test_plts_time_series.py
@@ -1,5 +1,7 @@
 """Test time series plots."""
 
+from pytest import raises
+
 from neurodsp.tests.utils import plot_test
 
 from neurodsp.plts.time_series import *
@@ -12,6 +14,18 @@ def test_plot_time_series(tsig):
 
     times = np.arange(0, len(tsig), 1)
     plot_time_series(times, tsig)
+
+@plot_test
+def test_plot_instantaneous_measure(tsig):
+
+    times = np.arange(0, len(tsig), 1)
+    plot_instantaneous_measure(times, tsig, 'phase')
+    plot_instantaneous_measure(times, tsig, 'amplitude')
+    plot_instantaneous_measure(times, tsig, 'frequency')
+
+    # Check the error for bad measure
+    with raises(ValueError):
+        plot_instantaneous_measure(times, tsig, 'BAD')
 
 @plot_test
 def test_plot_bursts(tsig):
