### 2019-04-10 ab65204
Update to use styling args
