+++ tutorials/plot_7-SimulatingSignals.py
@@ -10,7 +10,6 @@ This tutorial primarily covers :mod:`neurodsp.sim`.
 ###################################################################################################
 
 import numpy as np
-import matplotlib.pyplot as plt
 
 from neurodsp import spectral, sim
 
@@ -114,10 +113,8 @@ syn_noise = sim.sim_synaptic_current(n_seconds, fs)
 ###################################################################################################
 
 # Plot the simulated data, in the time domain
-plot_time_series(times, ou_noise)
-plt.title('OU Process')
-plot_time_series(times, syn_noise)
-plt.title('Synaptic Noise')
+plot_time_series(times, ou_noise, title='OU Process')
+plot_time_series(times, syn_noise, title='Synaptic Noise')
 
 ###################################################################################################
 
@@ -204,8 +201,7 @@ signal = sim.combined.sim_combined(n_seconds, fs, components)
 
 # Plot the simulated data, in the time domain
 samps_plot = np.arange(1000)
-plot_time_series(times[samps_plot], signal[samps_plot])
-plt.xlim((0, times[samps_plot[-1]]))
+plot_time_series(times[samps_plot], signal[samps_plot], xlim=[0, times[samps_plot[-1]]])
 
 ###################################################################################################
 
@@ -241,8 +237,7 @@ times = np.arange(0, n_seconds, 1/fs)
 ###################################################################################################
 
 # Plot the simulated data, in the time domain
-plot_time_series(times, osc)
-plt.xlim((0, n_seconds))
+plot_time_series(times, osc, xlim=[0, n_seconds])
 
 ###################################################################################################
 #
@@ -261,8 +256,7 @@ times = np.arange(0, n_seconds, 1/fs)
 ###################################################################################################
 
 # Plot the simulated data, in the time domain
-plot_time_series(times, osc)
-plt.xlim((0, n_seconds))
+plot_time_series(times, osc, xlim=[0, n_seconds])
 
 ###################################################################################################
 #
@@ -281,8 +275,7 @@ times = np.arange(0, n_seconds, 1/fs)
 ###################################################################################################
 
 # Plot the simulated data, in the time domain
-plot_time_series(times, osc)
-plt.xlim((0, n_seconds))
+plot_time_series(times, osc, xlim=[0, n_seconds])
 
 ###################################################################################################
 #
@@ -320,8 +313,7 @@ times = np.arange(0, n_seconds, 1/fs)
 ###################################################################################################
 
 # Plot the simulated data, in the time domain
-plot_time_series(times, osc)
-plt.xlim((0, n_seconds))
+plot_time_series(times, osc, xlim=[0, n_seconds])
 
 ###################################################################################################
 #
@@ -348,8 +340,7 @@ osc = sim.combined.sim_combined(n_seconds, fs, components)
 ###################################################################################################
 
 # Plot the simulated data, in the time domain
-plot_time_series(times, osc)
-plt.xlim((0, n_seconds))
+plot_time_series(times, osc, xlim=[0, n_seconds])
 
 ###################################################################################################
 
