+++ _sources/generated/neurodsp.sim.periodic.sim_jittered_oscillation.rst.txt
@@ -1,0 +1,6 @@
+neurodsp.sim.periodic.sim\_jittered\_oscillation
+================================================
+
+.. currentmodule:: neurodsp.sim.periodic
+
+.. autofunction:: sim_jittered_oscillation
