+++ _sources/generated/neurodsp.sim.periodic.sim_bursty_oscillation_features.rst.txt
@@ -1,0 +1,6 @@
+neurodsp.sim.periodic.sim\_bursty\_oscillation\_features
+========================================================
+
+.. currentmodule:: neurodsp.sim.periodic
+
+.. autofunction:: sim_bursty_oscillation_features
