+++ _sources/generated/neurodsp.filt.checks.check_filter_definition.rst.txt
@@ -1,0 +1,6 @@
+neurodsp.filt.checks.check\_filter\_definition
+==============================================
+
+.. currentmodule:: neurodsp.filt.checks
+
+.. autofunction:: check_filter_definition

[…]
