+++ _sources/auto_tutorials/plot_5-BurstDetection.rst.txt
@@ -257,7 +257,7 @@ sphinx_gallery_thumbnail_number = 2
 
 
 
-**Total running time of the script:** ( 0 minutes  0.303 seconds)
+**Total running time of the script:** ( 0 minutes  0.321 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_5-BurstDetection.py:
