+++ _sources/auto_tutorials/plot_4-SpectralAnalysis.rst.txt
@@ -315,7 +315,7 @@ sphinx_gallery_thumbnail_number = 2
 
 
 
-**Total running time of the script:** ( 0 minutes  1.599 seconds)
+**Total running time of the script:** ( 0 minutes  1.817 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_4-SpectralAnalysis.py:
