### 2019-04-10 a0e7b3e
Fix style approach to work with subplots
