+++ _sources/auto_tutorials/plot_6-SlidingWindowMatching.rst.txt
@@ -135,7 +135,7 @@ sphinx_gallery_thumbnail_number = 2
 
 
 
-**Total running time of the script:** ( 0 minutes  1.185 seconds)
+**Total running time of the script:** ( 0 minutes  1.237 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_6-SlidingWindowMatching.py:
