+++ _sources/auto_tutorials/plot_3-LaggedCoherence.rst.txt
@@ -263,7 +263,7 @@ sphinx_gallery_thumbnail_number = 2
 
 
 
-**Total running time of the script:** ( 0 minutes  0.106 seconds)
+**Total running time of the script:** ( 0 minutes  0.114 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_3-LaggedCoherence.py:
