### 2019-04-10 aec0f85
Clear site before install, to make sure everything is rebuilt
