### 2019-04-10 04c635f
Update tf plots and udpates for tutorials
