+++ tutorials/plot_2-InstantaneousMeasures.py
@@ -13,7 +13,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 from neurodsp.timefrequency import amp_by_time, freq_by_time, phase_by_time
-from neurodsp.plts.time_series import plot_time_series
+from neurodsp.plts.time_series import plot_time_series, plot_instantaneous_measure
 
 ###################################################################################################
 #
@@ -46,17 +46,9 @@ pha = phase_by_time(sig, fs, f_range)
 ###################################################################################################
 
 # Plot example signal
-plt.figure(figsize=(12, 4))
-plt.subplot(2, 1, 1)
-plt.plot(times, sig, 'k')
-plt.xlim((4, 5))
-plt.ylabel('Voltage (uV)')
-plt.subplot(2, 1, 2)
-plt.plot(times, pha, 'k')
-plt.xlim((4, 5))
-plt.yticks([-np.pi, 0, np.pi], ['$\pi$', 0, '$\pi$'])
-plt.xlabel('Time (s)')
-plt.ylabel('Phase (rad)')
+_, axs = plt.subplots(2, 1, figsize=(15, 6))
+plot_time_series(times, sig, xlim=[4, 5], xlabel=None, ax=axs[0])
+plot_instantaneous_measure(times, pha, xlim=[4, 5], ax=axs[1])
 
 ###################################################################################################
 #
@@ -71,20 +63,13 @@ amp = amp_by_time(sig, fs, f_range)
 ###################################################################################################
 
 # Plot example signal
-plt.figure(figsize=(12, 4))
-plt.subplot(2, 1, 1)
-plt.plot(times, sig, 'k', label='raw voltage')
-plt.plot(times, amp, 'r', label='amplitude')
-plt.legend(loc='best')
-plt.xlim((4, 5))
-plt.ylabel('Voltage (uV)')
-plt.subplot(2, 1, 2)
-plt.plot(times, sig_filt_true, 'b', label='beta-filtered voltage')
-plt.plot(times, amp, 'r', label='amplitude')
-plt.legend(loc='best')
-plt.xlim((4, 5))
-plt.xlabel('Time (s)')
-plt.ylabel('Voltage (uV)')
+_, axs = plt.subplots(2, 1, figsize=(15, 6))
+plot_instantaneous_measure(times, [sig, amp], 'amplitude',
+                           labels=['Raw Voltage', 'Amplitude'],
+                           xlim=[4, 5], xlabel=None, ax=axs[0])
+plot_instantaneous_measure(times, [sig_filt_true, amp], 'amplitude',
+                           labels=['Raw Voltage', 'Amplitude'], colors=['b', 'r'],
+                           xlim=[4, 5], ax=axs[1])
 
 ###################################################################################################
 #
@@ -103,24 +88,13 @@ i_f = freq_by_time(sig, fs, f_range)
 ###################################################################################################
 
 # Plot example signal
-plt.figure(figsize=(12, 6))
-plt.subplot(3, 1, 1)
-plt.plot(times, sig, 'k', label='raw voltage')
-plt.legend(loc='best')
-plt.ylabel('Voltage (uV)')
-plt.xlim((4, 5))
-plt.subplot(3, 1, 2)
-plt.plot(times, sig_filt_true, 'b', label='beta-filtered voltage')
-plt.legend(loc='best')
-plt.ylabel('Voltage (uV)')
-plt.xlim((4, 5))
-plt.subplot(3, 1, 3)
-plt.plot(times, i_f, 'r')
-plt.legend(loc='best')
-plt.xlim((4, 5))
-plt.xlabel('Time (s)')
-plt.ylabel('Instantaneous\nFrequency (Hz)')
-plt.ylim((10, 30))
[…]
