### 2019-04-10 cfb0d47
Fix axis label handling, add color support
