+++ neurodsp/plts/time_series.py
@@ -1,6 +1,6 @@
 """Plots for time series, for NeuroDSP."""
 
-from itertools import repeat
+from itertools import repeat, cycle
 
 import numpy as np
 import numpy.ma as ma
@@ -14,7 +14,7 @@ from neurodsp.plts.utils import check_ax, savefig
 
 @savefig
 @style_plot
-def plot_time_series(times, sigs, labels=None, ax=None):
+def plot_time_series(times, sigs, labels=None, colors=None, ax=None):
     """Plot a neural time series.
 
     Parameters
@@ -25,6 +25,8 @@ def plot_time_series(times, sigs, labels=None, ax=None):
         Time series to plot.
     labels : list of str, optional
         Labels for each time series.
+    cols : str or list of str
+        Colors to use to plot lines.
     ax : matplotlib.Axes, optional
         Figure axes upon which to plot.
     """
@@ -33,22 +35,27 @@ def plot_time_series(times, sigs, labels=None, ax=None):
 
     times = repeat(times) if isinstance(times, np.ndarray) else times
     sigs = [sigs] if isinstance(sigs, np.ndarray) else sigs
+
     if labels is not None:
         labels = [labels] if not isinstance(labels, list) else labels
     else:
         labels = repeat(labels)
-    cols = ['k', 'r', 'b', 'g', 'm', 'c', 'y']
 
-    for time, sig, col, label in zip(times, sigs, cols, labels):
-        plt.plot(time, sig, col, label=label)
+    if colors is not None:
+        colors = repeat(colors) if not isinstance(colors, list) else cycle(colors)
+    else:
+        colors = cycle(['k', 'r', 'b', 'g', 'm', 'c'])
 
-    plt.xlabel('Time (s)')
-    plt.ylabel('Voltage (uV)')
+    for time, sig, color, label in zip(times, sigs, colors, labels):
+        ax.plot(time, sig, color, label=label)
+
+    ax.set_xlabel('Time (s)')
+    ax.set_ylabel('Voltage (uV)')
 
 
 @savefig
 @style_plot
-def plot_instantaneous_measure(times, sigs, measure='phase', ax=None):
+def plot_instantaneous_measure(times, sigs, measure='phase', ax=None, **plt_kwargs):
     """Plot an instantaneous measure, of phase, amplitude or frequency.
 
     Parameters
@@ -61,23 +68,25 @@ def plot_instantaneous_measure(times, sigs, measure='phase', ax=None):
         Which kind of measure is being plotted.
     ax : matplotlib.Axes, optional
         Figure axes upon which to plot.
+    **kwargs
+        Keyword arguments to pass into `plot_time_series`.
     """
 
     if measure not in ['phase', 'amplitude', 'frequency']:
         raise ValueError('Measure not understood.')
 
     if measure == 'phase':
-        plot_time_series(times, sigs, ax=ax, ylabel='Phase (rad)')
+        plot_time_series(times, sigs, ax=ax, ylabel='Phase (rad)', **plt_kwargs)
         plt.yticks([-np.pi, 0, np.pi], ['-$\pi$', 0, '$\pi$'])
     elif measure == 'amplitude':
-        plot_time_series(times, sigs, ax=ax, ylabel='Amplitude')
+        plot_time_series(times, sigs, ax=ax, ylabel='Amplitude', **plt_kwargs)
     elif measure == 'frequency':
-        plot_time_series(times, sigs, ax=ax, ylabel='Instantaneous\nFrequency (Hz)')
+        plot_time_series(times, sigs, ax=ax, ylabel='Instantaneous\nFrequency (Hz)', **plt_kwargs)
 
 
 @savefig
 @style_plot
-def plot_bursts(times, sig, bursting, labels=None, ax=None):
[…]
