+++ _downloads/6beb788292810fb612f365994abc96ea/plot_4-SpectralAnalysis.py
@@ -25,10 +25,10 @@ This tutorial primarily covers :mod:`neurodsp.spectral`.
 
 import numpy as np
 from scipy import io
+import matplotlib.pylab as plt
 
 from neurodsp import spectral
-from neurodsp.plts.time_series import plot_time_series
-from neurodsp.plts.spectral import *
+from neurodsp.plts.spectral import plot_spectral_hist
 
 ###################################################################################################
 
@@ -39,8 +39,10 @@ data = io.loadmat('./data/sample_data_2.mat', squeeze_me=True)
 sig = data['x']
 fs = data['fs']
 times = np.arange(len(sig))/fs
-
-plot_time_series(times, sig, xlim=[0, 3])
+plt.plot(times, sig)
+plt.xlim((0, 3))
+plt.xlabel('Time (s)')
+plt.ylabel('Voltage')
 
 # Plotting the data, we observe a strong theta oscillation (~6-8 Hz)
 
@@ -66,12 +68,10 @@ plot_time_series(times, sig, xlim=[0, 3])
 ###################################################################################################
 
 # Mean of spectrogram (Welch)
-freq_mean, psd_mean = spectral.compute_spectrum(sig, fs, method='welch',
-                                                avg_type='mean', nperseg=fs*2)
+freq_mean, psd_mean = spectral.compute_spectrum(sig, fs, method='mean', nperseg=fs*2)
 
 # Median of spectrogram ("median Welch")
-freq_med, psd_med = spectral.compute_spectrum(sig, fs, method='welch',
-                                              avg_type='median', nperseg=fs*2)
+freq_med, psd_med = spectral.compute_spectrum(sig, fs, method='median', nperseg=fs*2)
 
 # Median filtered spectrum
 freq_mf, psd_mf = spectral.compute_spectrum(sig, fs, method='medfilt')
@@ -79,9 +79,13 @@ freq_mf, psd_mf = spectral.compute_spectrum(sig, fs, method='medfilt')
 ###################################################################################################
 
 # Plot the power spectra
-plot_power_spectra([freq_mean[:200], freq_med[:200], freq_mf[100:10000]],
-                   [psd_mean[:200], psd_med[:200], psd_mf[100:10000]],
-                   ['Welch', 'Median Welch', 'Median Filter FFT'])
+plt.figure(figsize=(8, 8))
+plt.loglog(freq_mean[:200], psd_mean[:200], label='Welch')
+plt.loglog(freq_med[:200], psd_med[:200], label='Median Welch')
+plt.loglog(freq_mf[100:10000], psd_mf[100:10000], label='Median Filter FFT')
+plt.legend()
+plt.xlabel('Frequency (Hz)')
+plt.ylabel('Power (V^2/Hz)')
 
 ###################################################################################################
 # Aside: Fitting 1/f and oscillations in power spectra
@@ -122,8 +126,8 @@ plot_power_spectra([freq_mean[:200], freq_med[:200], freq_mf[100:10000]],
 ###################################################################################################
 
 # Calculate the spectral histogram
-freqs, bins, spect_hist = spectral.compute_spectral_hist(sig, fs, nbins=50, f_range=(0, 80),
-                                                         cut_pct=(0.1, 99.9))
+freqs, bins, spect_hist = spectral.spectral_hist(sig, fs, nbins=50, f_lim=(0, 80),
+                                                 cutpct=(0.1, 99.9))
 
 # Plot the spectral histogram
 plot_spectral_hist(freqs, bins, spect_hist, freq_med, psd_med)
[…]
