+++ _downloads/b52349e2ed4ed17a9944d2416e40393d/plot_7-SimulatingSignals.py
@@ -10,11 +10,9 @@ This tutorial primarily covers :mod:`neurodsp.sim`.
 ###################################################################################################
 
 import numpy as np
+import matplotlib.pyplot as plt
 
 from neurodsp import spectral, sim
-
-from neurodsp.plts.spectral import plot_power_spectra
-from neurodsp.plts.time_series import plot_time_series
 
 ###################################################################################################
 
@@ -38,18 +36,24 @@ n_seconds = 10
 fs = 1000
 exponent = -2
 times = np.arange(0, n_seconds, 1/fs)
-br_noise = sim.sim_powerlaw(n_seconds, fs, exponent)
+br_noise = sim.sim_variable_powerlaw(n_seconds, fs, exponent)
 
 ###################################################################################################
 
 # Plot the simulated data, in the time domain
-plot_time_series(times, br_noise)
+plt.figure(figsize=(12, 3))
+plt.plot(times, br_noise, 'k')
+plt.xlabel('Time (s)')
+plt.ylabel('Voltage')
 
 ###################################################################################################
 
 # Plot the simulated data, in the frequency domain
 freqs, psd = spectral.compute_spectrum(br_noise, fs)
-plot_power_spectra(freqs, psd)
+plt.figure(figsize=(5, 5))
+plt.loglog(freqs, psd, 'k')
+plt.xlabel('Frequency (Hz)')
+plt.ylabel('Power')
 
 ###################################################################################################
 #
@@ -70,18 +74,24 @@ plot_power_spectra(freqs, psd)
 
 # Simulate highpass-filtered brown noise with a 1Hz cutoff frequency
 f_hipass_brown = 1
-brown_filt = sim.sim_powerlaw(n_seconds, fs, exponent, f_range=(f_hipass_brown, None))
+brown_filt = sim.sim_filtered_noise(n_seconds, fs, exponent, (f_hipass_brown, None))
 
 ###################################################################################################
 
 # Plot the simulated data, in the time domain
-plot_time_series(times, brown_filt)
+plt.figure(figsize=(12, 3))
+plt.plot(times, brown_filt, 'k')
+plt.xlabel('Time (s)')
+plt.ylabel('Voltage')
 
 ###################################################################################################
 
 # Plot the simulated data, in the frequency domain
 freqs, psd = spectral.compute_spectrum(brown_filt, fs)
-plot_power_spectra(freqs, psd)
+plt.figure(figsize=(5, 5))
+plt.loglog(freqs, psd, 'k')
+plt.xlabel('Frequency (Hz)')
+plt.ylabel('Power')
 
 ###################################################################################################
 #
@@ -107,22 +117,31 @@ plot_power_spectra(freqs, psd)
 ###################################################################################################
 
 # Simulate aperiodic signals from OU process & synaptic noise model
-ou_noise = sim.sim_random_walk(n_seconds, fs)
-syn_noise = sim.sim_synaptic_current(n_seconds, fs)
+ou_noise = sim.sim_ou_process(n_seconds, fs)
+syn_noise = sim.sim_synaptic_noise(n_seconds, fs)
 
 ###################################################################################################
 
 # Plot the simulated data, in the time domain
-plot_time_series(times, ou_noise, title='OU Process')
-plot_time_series(times, syn_noise, title='Synaptic Noise')
+plt.figure(figsize=(12, 6))
+plt.subplot(2, 1, 1, title='OU Process')
+plt.plot(times, ou_noise, 'k')
+plt.subplot(2, 1, 2, title='Synaptic Noise')
+plt.plot(times, syn_noise, 'r')
+plt.xlabel('Time (s)')
[…]
