+++ _sources/generated/neurodsp.spectral.spectral.compute_spectrum_medfilt.rst.txt
@@ -1,6 +1,0 @@
-neurodsp.spectral.spectral.compute\_spectrum\_medfilt
-=====================================================
-
-.. currentmodule:: neurodsp.spectral.spectral
-
-.. autofunction:: compute_spectrum_medfilt
