+++ _sources/generated/neurodsp.sim.sim_noisy_bursty_oscillator.rst.txt
@@ -1,0 +1,6 @@
+neurodsp.sim.sim\_noisy\_bursty\_oscillator
+===========================================
+
+.. currentmodule:: neurodsp.sim
+
+.. autofunction:: sim_noisy_bursty_oscillator

[…]
