### 2019-04-10 9ce3f88
Add plot fucntios to API list
