### 2019-04-10 c408158
Update some var names & small doc updates to spectral
