+++ neurodsp/tests/test_plts_spectral.py
@@ -11,5 +11,5 @@ from neurodsp.plts.spectral import *
 @plot_test
 def test_plot_spectral_hist(tsig):
 
-    freqs, power_bins, spect_hist = compute_spectral_hist(tsig, fs=1000)
-    plot_spectral_hist(freqs, power_bins, spect_hist)
+    freqs, power_bins, spectral_hist = compute_spectral_hist(tsig, fs=1000)
+    plot_spectral_hist(freqs, power_bins, spectral_hist)
