### 2019-04-10 a4471b0
Update default to not necessarily print filter properties, but always run warning checks
