### 2019-04-10 8762c32
Update MNE examples for code updates
