### 2019-04-10 288ec93
Add a 'create_times' helper fucntion
