### 2019-04-11 d2faaa7
Fix up glossary based on comments from Richard
