### 2019-04-12 ab1f9d8
Update n_second managed in sim jittered oscillator
