+++ doc/conf.py
@@ -110,6 +110,7 @@ html_theme_options = {
     # A list of tuples containing pages or urls to link to.
     'navbar_links': [
         ("API", "api"),
+        ("Glossary", "glossary"),
         ("Tutorial", "auto_tutorials/index"),
         ("Examples", "auto_examples/index"),
         ("GitHub", "https://github.com/neurodsp-tools/neurodsp", True)
