+++ doc/glossary.rst
@@ -10,6 +10,7 @@ General
 
     periodic
         Properties or components of a signal that are rhythmic.
+
     aperiodic
         Properties or components of a signal that are arrhythmic, with no characteristic frequency.
 
@@ -24,13 +25,17 @@ by Jack Schaedler.
 
     time domain
         Signals that are represented as variations over time, and analyses of such signals.
+
     frequency domain
         Signals that are represented in terms of frequencies, and analyses of such signals.
+
     sampling rate
         The rate at which samples are taken.
+
     temporal resolution
         The precision of a measurement, in the time domain.
         This is set by the magnitude of time between successive measurements (e.g. 0.01 seconds between samples).
+
     frequency resolution
         The precision of a measurement, in the frequency domain.
         This is set by the magnitude of frequency between successive measurements (e.g. 0.5 Hz between measurements).
@@ -42,8 +47,10 @@ Units
 
     Hertz (Hz)
         A unit of frequency, as the number of cycles per second.
+
     Decibels (dB)
         A unit of intensity, on a logarithmic scale.
+
     Volts (V)
         A unit of voltage, typically in the microvolt (uV) range for neural time series.
 
@@ -60,14 +67,19 @@ For a more in depth tutorial, in code, check out the
 
     Impulse Response
         The response of a filter when presented with an impulse; a single, brief input.
+
     FIR
         A Finite Impulse Response filter, meaning its impulse response settles to zero in finite time.
+
     IIR
         An Infinite Impulse Response filter, meaning the filter is recursive, and its impulse response continues infinitely.
+
     passband
         The range (band) of frequencies that are unattenuated by a filter.
+
     stopband
         The range (band) of frequencies that are attenuated (stopped) by a filter.
+
     passtype
         The type of filter, defined in terms of what frequency bands or ranges it passes through, or filters out.
 
@@ -75,8 +87,10 @@ For a more in depth tutorial, in code, check out the
         * bandstop: a filter that passes through all frequencies except a band region that is attenuated.
         * lowpass: a filter whose passband is all frequencies below a filter frequency (low frequencies pass through).
         * highpass: a filter whose passband is all frequencies above a filter frequency (high frequencies pass through).
+
     transition band
         The range of frequencies that are in the transition region between the passband and the stopband.
+
     frequency response
         The response profile of a filter, specifying the gain and phase shift applied by the filter at each frequency.
 
@@ -106,17 +120,23 @@ We currently have two general approaches to time frequency analyses:
 
     frequency
         The number of occurences over a unit of time, typically referred to as cycles per second, and measured in Hz.
+
     phase
         The position, at a point in time, on a waveform cycle.
+
     amplitude
         The magnitude of a signal, as the peak-to-trough.
+
     power
         The squared magnitude of a signal.
+
     period
         A single cycle of a rhythm, defined as the time between two consecutive troughs (or peaks).
+
     hilbert transform
         A mathematical transform that computes the 'analytic signal', a complex-valued representation
         of a time-series (signal) that can be used to find its analytic amplitude and phase.
+
     wavelet
         A wave-like signal, or 'brief oscillation', that starts at zero amplitude, increases
         in amplitude to some value, and then decays back to zero.
@@ -134,10 +154,13 @@ by 3Blue1Brown.
 
     fourier transform
         A mathematical transformation to decompose a time series into its constituent frequencies.
+
     power spectrum
         A frequency domain representation, as an estimate of the power across frequencies in a signal.
+
     median filter
         A smoothing approach to replace each value in a signal with the median of the neighbouring entries.
+
     coefficient of variation
         A standardized measure of dispersion, as the ratio of the standard deviation to the mean.
 
[…]
