+++ _sources/auto_tutorials/plot_6-SlidingWindowMatching.rst.txt
@@ -144,7 +144,7 @@ sphinx_gallery_thumbnail_number = 2
 
 
 
-**Total running time of the script:** ( 0 minutes  1.257 seconds)
+**Total running time of the script:** ( 0 minutes  1.506 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_6-SlidingWindowMatching.py:
