+++ _sources/auto_tutorials/plot_7-SimulatingSignals.rst.txt
@@ -690,7 +690,7 @@ sphinx_gallery_thumbnail_number = 15
 
 
 
-**Total running time of the script:** ( 0 minutes  0.907 seconds)
+**Total running time of the script:** ( 0 minutes  0.885 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_7-SimulatingSignals.py:
