### 2019-04-16 3810375
Add multidim decorator to all relevant funcs
