+++ neurodsp/filt/filter.py
@@ -2,12 +2,14 @@
 
 from warnings import warn
 
+from neurodsp.utils.decorators import multidim
 from neurodsp.filt.fir import filter_signal_fir
 from neurodsp.filt.iir import filter_signal_iir
 
 ###################################################################################################
 ###################################################################################################
 
+@multidim
 def filter_signal(sig, fs, pass_type, f_range, filt_type='fir',
                   n_cycles=3, n_seconds=None, remove_edges=True, butterworth_order=None,
                   print_transitions=False, plot_properties=False, return_filter=False):
