### 2019-04-16 d45dbc4
Clean up how rotating deals with f=0
