+++ neurodsp/spectral/utils.py
@@ -63,7 +63,18 @@ def rotate_powerlaw(freqs, spectrum, delta_exponent, f_rotation=1):
         Rotated spectrum.
     """
 
+    if freqs[0] == 0:
+        skipped_zero = True
+        f_0, p_0 = freqs[0], spectrum[0]
+        freqs, spectrum = freqs[1:], spectrum[1:]
+    else:
+        skipped_zero = False
+
     mask = (np.abs(freqs) / f_rotation)**-delta_exponent
     rotated_spectrum = mask * spectrum
+
+    if skipped_zero:
+        freqs = np.insert(freqs, 0, f_0),
+        rotated_spectrum =  np.insert(rotated_spectrum, 0, p_0)
 
     return rotated_spectrum
