### 2019-04-16 9580515
Temporarily remove periodic sims that don't match new approach
