+++ neurodsp/sim/periodic.py
@@ -99,145 +99,6 @@ def sim_bursty_oscillation(n_seconds, fs, freq, enter_burst=.2, leave_burst=.2,
 
     return sig
 
-
-@normalize
-def sim_bursty_oscillation_features(n_seconds, fs, freq, rdsym=.5, enter_burst=.2, leave_burst=.2,
-                                    cycle_features=None, return_cycle_df=False, n_tries=5):
-    """Simulate a bursty oscillation, with defined cycle features.
-
-    Parameters
-    ----------
-    n_seconds : float
-        Simulation time, in seconds.
-    fs : float
-        Sampling rate of simulated signal, in Hz
-    freq : float
-        Oscillation frequency, in Hz.
-    rdsym : float
-        Rise-decay symmetry of the oscillation, as fraction of the period in the rise time:
-
-        - = 0.5: symmetric (sine wave)
-        - < 0.5: shorter rise, longer decay
-        - > 0.5: longer rise, shorter decay
-    enter_burst : float
-        Probability of a cycle being oscillating given the last cycle is not oscillating.
-    leave_burst : float
-        Probability of a cycle not being oscillating given the last cycle is oscillating.
-    cycle_features : dict
-        Specifies the mean and standard deviations (within and across bursts) of each cycle's
-        amplitude, period, and rise-decay symmetry. This can include a complete or incomplete
-        set (using defaults) of the following keys:
-
-        * amp_mean: mean cycle amplitude
-        * amp_std: standard deviation of cycle amplitude
-        * amp_burst_std: standard deviation of mean amplitude for each burst
-        * period_mean: mean period (computed from `freq`)
-        * period_std: standard deviation of period (samples)
-        * period_burst_std: standard deviation of mean period for each burst
-        * rdsym_mean: mean rise-decay symmetry
-        * rdsym_std: standard deviation of rdsym
-        * rdsym_burst_std: standard deviation of mean rdsym for each burst
-    return_cycle_df : bool
-        If True, return the dataframe that contains the simulation parameters for each cycle.
-        This may be useful for computing power, for example, as the power of the oscillation
-        should only be considered over the times where there are bursts.
-    n_tries : int, optional, default=5
-        Number of times to try to resimulate cycle features when an
-        invalid value is returned before raising an user error.
-
-    Returns
-    -------
-    sig : 1d array
-        Bursty oscillation.
-    df : pd.DataFrame
-        Cycle-by-cycle properties of the simulated oscillation.
-        Only returned if `return_cycle_df` is True.
-    """
-
-    # Define default parameters for cycle features
-    mean_period_samples = int(fs / freq)
-    cycle_features_use = {'amp_mean': 1, 'amp_burst_std': 0, 'amp_std': 0,
-                          'period_mean': mean_period_samples, 'period_burst_std': 0, 'period_std': 0,
-                          'rdsym_mean': rdsym, 'rdsym_burst_std': 0, 'rdsym_std': 0}
-
-    # Overwrite default cycle features with those specified
-    if cycle_features is not None:
-        for k in cycle_features:
-            cycle_features_use[k] = cycle_features[k]
-
-    # Determine number of cycles to generate
-    n_samples = int(n_seconds * fs)
-    n_cycles_overestimate = int(np.ceil(n_samples / mean_period_samples * 2))
-
-    # Determine which periods will be oscillating and the cycle properties for each cycle
-    is_oscillating = _make_is_osc(n_cycles_overestimate, enter_burst, leave_burst)
[…]
