### 2019-04-16 9d32347
Add back not-yet-updated periodic sims
