+++ neurodsp/sim/periodic.py
@@ -44,6 +44,56 @@ def sim_oscillation(n_seconds, fs, freq, cycle='sine', **cycle_params):
 
 
 @normalize
+def sim_jittered_oscillation(n_seconds, fs, freq, jitter=0, cycle='sine', **cycle_params):
+    """Simulate an oscillation with jitter.
+
+    # THIS 'CONVOLUITON APPROACH' GENERALIZE WITH KERNEL APPROACH.
+
+    Parameters
+    ----------
+    n_seconds : float
+        Simulation time, in seconds.
+    fs : float
+        Sampling rate of simulated signal, in Hz.
+    freq : float
+        Frequency of simulated oscillation, in Hz.
+    jitter : float
+        Maximum jitter of oscillation period, in seconds.
+    cycle : {'sine', 'asine', 'sawtooth', 'gaussian', 'exp', '2exp'}
+        What type of oscillation cycle to simulate.
+        See `sim_osc_cycle` for details on cycle types and parameters.
+    **cycle_params
+        Parameters for the simulated oscillation cycle.
+
+    Returns
+    -------
+    sig: 1d array
+        Simulated oscillation with jitter.
+    """
+
+    n_cycles = int(np.ceil(n_seconds * freq))
+    n_seconds_cycle = int(np.ceil(fs / freq)) / fs
+
+    # Create a single cycle of the oscillation
+    osc_cycle = sim_osc_cycle(n_seconds_cycle, fs, cycle, **cycle_params)
+
+    # Binary "spike-train" of when each cycle should occur & oscillation "event" indices
+    spks = np.zeros(int(n_seconds * fs + len(osc_cycle)) - 1)
+    osc_period = int(fs / freq)
+    spk_indices = np.arange(osc_period, len(spks), osc_period)
+
+    # Add jitter to "spike" indices
+    if jitter != 0:
+        spk_indices = spk_indices + \
+            randint(low=-int(fs * jitter), high=int(fs * jitter), size=len(spk_indices))
+
+    spks[spk_indices] = 1
+    sig = np.convolve(spks, osc_cycle, 'valid')
+
+    return sig
+
+
+@normalize
 def sim_bursty_oscillation(n_seconds, fs, freq, enter_burst=.2, leave_burst=.2,
                            cycle='sine', **cycle_params):
     """Simulate a bursty oscillation.
@@ -99,6 +149,100 @@ def sim_bursty_oscillation(n_seconds, fs, freq, enter_burst=.2, leave_burst=.2,
 
     return sig
 
+
+@normalize
+def sim_bursty_oscillation_features(n_seconds, fs, freq, rdsym=.5, enter_burst=.2, leave_burst=.2,
+                                    cycle_features=None, return_cycle_df=False, n_tries=5):
+    """Simulate a bursty oscillation, with defined cycle features.
+
+    # NEEDS REFACTORING TO USE KERNEL APPROACH
+    # WHEN DONE, THIS SHOULD BE ABLE TO MERGE TO SIM_BURSTY_OSCILLATION
+
+    Parameters
+    ----------
+    n_seconds : float
+        Simulation time, in seconds.
+    fs : float
+        Sampling rate of simulated signal, in Hz
+    freq : float
+        Oscillation frequency, in Hz.
+    rdsym : float
+        Rise-decay symmetry of the oscillation, as fraction of the period in the rise time:
+
+        - = 0.5: symmetric (sine wave)
+        - < 0.5: shorter rise, longer decay
+        - > 0.5: longer rise, shorter decay
+    enter_burst : float
+        Probability of a cycle being oscillating given the last cycle is not oscillating.
+    leave_burst : float
+        Probability of a cycle not being oscillating given the last cycle is oscillating.
+    cycle_features : dict
[…]
