### 2019-04-16 #1272 @TomDonoghue
Hey @arokem - sorry for the delay, we were just fixing up some last things in the `sim` module that got refactored in the updates - but yes, I did just merge the updates in!

What is now in the `paper` folder on master is the updated version of the paper (with the updates for the code updates), and all those code updates have now also been merged to master.

This means we should be ready to check the new rendered version of the article! After that, depositing an archive of the article / code makes sense - though I'm not sure how to do that / where to put it, if you can perhaps direct us on some further instructions on that!

cc: @voytek & @rdgao
