### 2019-04-16 #1272 @arokem
Is the first sentence missing a word?

Should that be two sentences? Something like: "Populations of neurons exhibit time-varying fluctuations in their aggregate electrical activity. These fluctuations are measured using ectrophysiological methods, such as magneto or electroencephalography (M/EEG), intracranial EEG (iEEG) or electrocorticography (ECoG), and local field potential (LFP) recordings [@buzsaki_origin_2012]."
