### 2019-04-16 663cb23
Update paper.md
correcting weird sentence error that propagated somehow
