### 2019-04-16 #1272 @TomDonoghue
Hey @arokem - thanks for the info, couple quick questions:

1) In the article proof, a couple of the in text citations are a little weird: specifically it cites `B. Voytek` in the first paragraph, and `Scott R Cole` in the second bullet point of the modules, inconsistently using first names / initials. It's not clear from looking at the paper.bib file what's different - do you happen to know what's making it do that?

2) The current version of the software is 2.0.0, though this is not yet tagged on Github and is a pre-release on pip (just so it gets a little more road tested after the revamp for 2.0). Can we set that as the software version number? That's the version that I will put up on Zenodo, if that all sounds good to you.
