### 2019-04-16 #1272 @arokem
Your last bib item uses a different format for first/last names than all the others.

Also, I think you have "Scott R Cole" here:

https://github.com/neurodsp-tools/neurodsp/blob/master/paper/paper.bib#L41

And "Scott Robert Cole" here:

https://github.com/neurodsp-tools/neurodsp/blob/master/paper/paper.bib#L31

Which is also causing some issues.
