### 2019-04-17 367353e
Standardize author lists in references
