+++ paper/paper.bib
@@ -17,7 +17,7 @@
     doi = {10.1038/nrn3241},
     number = {6},
     journal = {Nature Reviews Neuroscience},
-    author = {Buzsaki, Gyorgy and Anastassiou, Costas A. and Koch, Christof},
+    author = {Buzsáki, György and Anastassiou, Costas A. and Koch, Christof},
     year = {2012},
     month = may,
     pages = {407--420}
@@ -28,7 +28,7 @@
     url = {http://biorxiv.org/lookup/doi/10.1101/452987},
     doi = {10.1101/452987},
     journal = {{bioRxiv}},
-    author = {Cole, Scott Robert and Voytek, Bradley},
+    author = {Cole, Scott R. and Voytek, Bradley},
     year = {2018},
     month = oct
 }
@@ -38,7 +38,7 @@
     url = {http://biorxiv.org/lookup/doi/10.1101/302000},
     doi = {10.1101/302000},
     journal = {{bioRxiv}},
-    author = {Cole, Scott R and Voytek, Bradley},
+    author = {Cole, Scott R. and Voytek, Bradley},
     year = {2018},
     month = apr
 }
@@ -176,7 +176,7 @@
     pages = {13257--13265},
     number = {38},
     journal = {Journal of Neuroscience},
-    author = {Voytek, B. and Kramer, M. A. and Case, J. and Lepage, K. Q. and Tempesta, Z. R. and Knight, R. T. and Gazzaley, A.},
+    author = {Voytek, Bradley and Kramer, Mark A. and Case, John and Lepage, Kyle Q. and Tempesta, Zechari R. and Knight, Robert T. and Gazzaley, Adam},
     year = {2015},
     month = sep
 }
