### 2019-04-17 7003953
remove sklearn dependency, which is no longer used / required
