### 2019-04-17 #1272 @TomDonoghue
Okay thanks for the info / guidance!

- I stepped through and updated the references, trying to make sure that they all used the same format, so hopefully that fixed things up. Can you rebuild the paper?

- I enabled Zenodo for the NeuroDSP repository, and tagged a pre-release version of 2.0.0, and it got assigned a DOI through Zenodo:
DOI: 10.5281/zenodo.2643784
Zenodo badge: https://zenodo.org/badge/doi/10.5281/zenodo.2643784.svg
