### 2019-04-17 faf793b
Fix for references with multiple citations.
