### 2019-04-17 5c6d337
Add links to ByCycle & FOOOF
