### 2019-04-17 #1272 @TomDonoghue
Okay, two small fixes / updates:
- Fixed the references for when there are multiple refs for the same claim, so it doesn't add extra parentheses to the year.
- Added links to the ByCycle & FOOOF toolboxes.

Otherwise - I did another check through, and everything looks good to me now!
Can confirm to go through to the acceptance process! Thanks @arokem
