### 2019-04-17 #1272 @arokem
👋 @openjournals/joss-eics : I believe this paper is ready to be accepted. Could you please take a look?
